==> 2_venndiagram_from_fgsea_MaT4_and_McT4_epi_cell_GOBP_REACTOME.R <==
#2023.04.21
setwd('F:/Ko/67ESCC_PCN_integration_07-19-22/immune_cell_T4_M4/fgsea')

library(ggplot2)
library(RTCGA)
library(dplyr)
library(RTCGAToolbox)
library(TCGAbiolinks)
library(data.table)
#install.packages("RDocumentation")
library(RDocumentation)


##data from excel##
# GOBP
######use format -> cells -> custom -> \'@\'\, ########
MaT4 = c("GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",	"GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	
         "GOBP_ATP_BIOSYNTHETIC_PROCESS",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_PROTEIN_REFOLDING",	
         "GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_AEROBIC_RESPIRATION",	"GOBP_PROTEIN_LOCALIZATION_TO_ENDOPLASMIC_RETICULUM",	
         "GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_COLLAGEN_FIBRIL_ORGANIZATION",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_MATURATION",	"GOBP_ATP_METABOLIC_PROCESS",	
         "GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_BY_HOST_OF_VIRAL_PROCESS",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_CELLULAR_RESPIRATION",	
         "GOBP_ELECTRON_TRANSPORT_CHAIN",	"GOBP_COLLAGEN_METABOLIC_PROCESS",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	
         "GOBP_REGULATION_OF_B_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	"GOBP_RESPONSE_TO_ZINC_ION_STARVATION",	
         "GOBP_ESTABLISHMENT_OF_PROTEIN_LOCALIZATION_TO_ENDOPLASMIC_RETICULUM",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_PROTEIN_EXIT_FROM_ENDOPLASMIC_RETICULUM",	
         "GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_ENDOPLASMIC_RETICULUM_TO_GOLGI_VESICLE_MEDIATED_TRANSPORT",	
         "GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",	"GOBP_VESICLE_BUDDING_FROM_MEMBRANE",	"GOBP_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	
         "GOBP_CYTOPLASMIC_TRANSLATION",	"GOBP_PROTEIN_FOLDING",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	
         "GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_EXTERNAL_ENCAPSULATING_STRUCTURE_ORGANIZATION",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",	"GOBP_OSTEOBLAST_DIFFERENTIATION",	
         "GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	
         "GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_GOLGI_VESICLE_TRANSPORT",	"GOBP_CARBOHYDRATE_DERIVATIVE_BIOSYNTHETIC_PROCESS",	
         "GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_RESPONSE_TO_ENDOPLASMIC_RETICULUM_STRESS",	"GOBP_MITOCHONDRION_ORGANIZATION",	
         "GOBP_VIRAL_LIFE_CYCLE",	"GOBP_RESPONSE_TO_WOUNDING",	"GOBP_WOUND_HEALING",	"GOBP_REGULATION_OF_PROTEOLYSIS",	"GOBP_POSITIVE_REGULATION_OF_PROTEOLYSIS",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	
         "GOBP_PROTEIN_TARGETING",	"GOBP_MUSCLE_STRUCTURE_DEVELOPMENT"
)
McT4 = c("GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_COLLAGEN_METABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	
         "GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	
         "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_I",	"GOBP_COLLAGEN_FIBRIL_ORGANIZATION",	"GOBP_REGULATION_OF_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	
         "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_EXTERNAL_ENCAPSULATING_STRUCTURE_ORGANIZATION",	
         "GOBP_REGULATION_OF_T_CELL_MEDIATED_CYTOTOXICITY",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_COLLAGEN_CATABOLIC_PROCESS",	
         "GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_POSITIVE_REGULATION_OF_T_CELL_MEDIATED_CYTOTOXICITY",	"GOBP_RESPONSE_TO_INTERFERON_ALPHA",	
         "GOBP_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",	
         "GOBP_T_CELL_MEDIATED_CYTOTOXICITY",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_PLATELET_AGGREGATION",	"GOBP_POSITIVE_REGULATION_OF_CELL_KILLING",	"GOBP_NEUTROPHIL_EXTRAVASATION",	
         "GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_RESPONSE_TO_TYPE_I_INTERFERON",	
         "GOBP_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_BY_SYMBIONT_OF_ENTRY_INTO_HOST",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_DETECTION_OF_OTHER_ORGANISM",	
         "GOBP_PYRIMIDINE_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_HOMOTYPIC_CELL_CELL_ADHESION",	"GOBP_POSITIVE_REGULATION_OF_CELLULAR_EXTRAVASATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	
         "GOBP_UTP_METABOLIC_PROCESS",	"GOBP_COLLAGEN_BIOSYNTHETIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	
         "GOBP_POSITIVE_REGULATION_OF_PROTEIN_MATURATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_I_VIA_ER_PATHWAY",	
         "GOBP_REGULATION_OF_CELL_SUBSTRATE_ADHESION",	"GOBP_REGULATION_OF_MONOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_NEUTROPHIL_EXTRAVASATION",	"GOBP_REGULATION_OF_NEUTROPHIL_EXTRAVASATION",	
         "GOBP_POSITIVE_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_CELL_KILLING",	"GOBP_HEMOSTASIS",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	
         "GOBP_NUCLEOSIDE_DIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_GTP_BIOSYNTHETIC_PROCESS",	"GOBP_DETECTION_OF_EXTERNAL_BIOTIC_STIMULUS",	"GOBP_PROTEIN_ACTIVATION_CASCADE",	
         "GOBP_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	
         "GOBP_POSITIVE_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	"GOBP_PYRIMIDINE_RIBONUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_KILLING",	
         "GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_RECEPTOR_BINDING",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_DETOXIFICATION_OF_COPPER_ION",	"GOBP_PLASMINOGEN_ACTIVATION",	
         "GOBP_PLATELET_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_PLASMINOGEN_ACTIVATION",	"GOBP_ISG15_PROTEIN_CONJUGATION",	"GOBP_NUCLEOSIDE_DIPHOSPHATE_METABOLIC_PROCESS",	
         "GOBP_PURINE_NUCLEOSIDE_DIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_PLASMA_MEMBRANE_REPAIR",	"GOBP_ENTRY_INTO_HOST",	"GOBP_INTEGRIN_MEDIATED_SIGNALING_PATHWAY",	
         "GOBP_PYRIMIDINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_ENDOTHELIAL_CELL_PROLIFERATION",	"GOBP_REGULATION_OF_PROTEIN_MATURATION",	
         "GOBP_PYRIMIDINE_NUCLEOTIDE_METABOLIC_PROCESS",	"GOBP_CELL_SUBSTRATE_ADHESION",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_NEGATIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	
         "GOBP_PYRIMIDINE_RIBONUCLEOTIDE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_CELL_SUBSTRATE_JUNCTION_ORGANIZATION",	"GOBP_WOUND_HEALING",	"GOBP_ACTOMYOSIN_STRUCTURE_ORGANIZATION",	
         "GOBP_ACTIN_FILAMENT_BUNDLE_ORGANIZATION",	"GOBP_NUCLEOTIDE_PHOSPHORYLATION",	"GOBP_REGULATION_OF_T_CELL_MEDIATED_IMMUNITY",	"GOBP_CELL_MATRIX_ADHESION",	
         "GOBP_POSITIVE_REGULATION_OF_CELL_SUBSTRATE_ADHESION",	"GOBP_CELL_SUBSTRATE_JUNCTION_ORGANIZATION",	"GOBP_REGULATION_OF_FIBROBLAST_PROLIFERATION",	
         "GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_REGULATION_OF_PHAGOCYTOSIS",	"GOBP_VIRAL_GENOME_REPLICATION",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_ADP_METABOLIC_PROCESS",	
         "GOBP_NUCLEOSIDE_MONOPHOSPHATE_METABOLIC_PROCESS",	"GOBP_T_CELL_MEDIATED_IMMUNITY",	"GOBP_STRESS_FIBER_ASSEMBLY",	"GOBP_MAINTENANCE_OF_PROTEIN_LOCATION",	
         "GOBP_RIBONUCLEOSIDE_DIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	
         "GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_CELL_KILLING",	"GOBP_FIBROBLAST_PROLIFERATION",	"GOBP_FOCAL_ADHESION_ASSEMBLY",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	
         "GOBP_REGULATION_OF_BODY_FLUID_LEVELS",	"GOBP_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	
         "GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",	"GOBP_MUSCLE_CONTRACTION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_CELL_MATRIX_ADHESION",	
         "GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_REGULATION_OF_CELL_JUNCTION_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	
         "GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_BONE_DEVELOPMENT",	"GOBP_REGULATION_OF_CELL_MORPHOGENESIS",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	
         "GOBP_ADAPTIVE_IMMUNE_RESPONSE_BASED_ON_SOMATIC_RECOMBINATION_OF_IMMUNE_RECEPTORS_BUILT_FROM_IMMUNOGLOBULIN_SUPERFAMILY_DOMAINS",	"GOBP_CELL_JUNCTION_ASSEMBLY",	"GOBP_RESPONSE_TO_VIRUS",	
         "GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	
         "GOBP_CYTOKINE_MEDIATED_SIGNALING_PATHWAY"
)

data_1 = MaT4
data_2 = McT4

library(VennDiagram)
library(RColorBrewer)
library(ggvenn)

colors()

set1 = as.list(data_1)
set2 = as.list(data_2)

venn.diagram(x=list(set1, set2), 
             category.names = c("MaT4","McT4"), filename = "venndiagram_GOBP_MaT4_McT4_epi.png", output=T, 
             imagetype = "png", height = 2000, width = 2000, resolution = 300, compression = "lzw",
             fill=c("skyblue", "yellowgreen"), 
             cex = 2,  lwd=2, lty='blank', fontface="bold", fontfamily="sans", cat.cex=1, cat.fontface="bold", cat.default.pos="outer", cat.pos=c(0, 0),cat.fontfamily="sans")


a <- list("MaT4vs.all_pos_GOBP" = c("GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",	"GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	
          "GOBP_ATP_BIOSYNTHETIC_PROCESS",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_PROTEIN_REFOLDING",	
          "GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_AEROBIC_RESPIRATION",	"GOBP_PROTEIN_LOCALIZATION_TO_ENDOPLASMIC_RETICULUM",	
          "GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_COLLAGEN_FIBRIL_ORGANIZATION",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_MATURATION",	"GOBP_ATP_METABOLIC_PROCESS",	
          "GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_BY_HOST_OF_VIRAL_PROCESS",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_CELLULAR_RESPIRATION",	
          "GOBP_ELECTRON_TRANSPORT_CHAIN",	"GOBP_COLLAGEN_METABOLIC_PROCESS",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	
          "GOBP_REGULATION_OF_B_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	"GOBP_RESPONSE_TO_ZINC_ION_STARVATION",	
          "GOBP_ESTABLISHMENT_OF_PROTEIN_LOCALIZATION_TO_ENDOPLASMIC_RETICULUM",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_PROTEIN_EXIT_FROM_ENDOPLASMIC_RETICULUM",	
          "GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_ENDOPLASMIC_RETICULUM_TO_GOLGI_VESICLE_MEDIATED_TRANSPORT",	
          "GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",	"GOBP_VESICLE_BUDDING_FROM_MEMBRANE",	"GOBP_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	
          "GOBP_CYTOPLASMIC_TRANSLATION",	"GOBP_PROTEIN_FOLDING",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	
          "GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_EXTERNAL_ENCAPSULATING_STRUCTURE_ORGANIZATION",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",	"GOBP_OSTEOBLAST_DIFFERENTIATION",	
          "GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	
          "GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_GOLGI_VESICLE_TRANSPORT",	"GOBP_CARBOHYDRATE_DERIVATIVE_BIOSYNTHETIC_PROCESS",	
          "GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_RESPONSE_TO_ENDOPLASMIC_RETICULUM_STRESS",	"GOBP_MITOCHONDRION_ORGANIZATION",	
          "GOBP_VIRAL_LIFE_CYCLE",	"GOBP_RESPONSE_TO_WOUNDING",	"GOBP_WOUND_HEALING",	"GOBP_REGULATION_OF_PROTEOLYSIS",	"GOBP_POSITIVE_REGULATION_OF_PROTEOLYSIS",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	
          "GOBP_PROTEIN_TARGETING",	"GOBP_MUSCLE_STRUCTURE_DEVELOPMENT"),
          
          "McT4vs.all_pos_GOBP" = c("GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_COLLAGEN_METABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	
                                    "GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	
                                    "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_I",	"GOBP_COLLAGEN_FIBRIL_ORGANIZATION",	"GOBP_REGULATION_OF_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	
                                    "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_EXTERNAL_ENCAPSULATING_STRUCTURE_ORGANIZATION",	
                                    "GOBP_REGULATION_OF_T_CELL_MEDIATED_CYTOTOXICITY",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_COLLAGEN_CATABOLIC_PROCESS",	
                                    "GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_POSITIVE_REGULATION_OF_T_CELL_MEDIATED_CYTOTOXICITY",	"GOBP_RESPONSE_TO_INTERFERON_ALPHA",	
                                    "GOBP_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",	
                                    "GOBP_T_CELL_MEDIATED_CYTOTOXICITY",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_PLATELET_AGGREGATION",	"GOBP_POSITIVE_REGULATION_OF_CELL_KILLING",	"GOBP_NEUTROPHIL_EXTRAVASATION",	
                                    "GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_RESPONSE_TO_TYPE_I_INTERFERON",	
                                    "GOBP_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_BY_SYMBIONT_OF_ENTRY_INTO_HOST",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_DETECTION_OF_OTHER_ORGANISM",	
                                    "GOBP_PYRIMIDINE_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_HOMOTYPIC_CELL_CELL_ADHESION",	"GOBP_POSITIVE_REGULATION_OF_CELLULAR_EXTRAVASATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	
                                    "GOBP_UTP_METABOLIC_PROCESS",	"GOBP_COLLAGEN_BIOSYNTHETIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	
                                    "GOBP_POSITIVE_REGULATION_OF_PROTEIN_MATURATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_I_VIA_ER_PATHWAY",	
                                    "GOBP_REGULATION_OF_CELL_SUBSTRATE_ADHESION",	"GOBP_REGULATION_OF_MONOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_NEUTROPHIL_EXTRAVASATION",	"GOBP_REGULATION_OF_NEUTROPHIL_EXTRAVASATION",	
                                    "GOBP_POSITIVE_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_CELL_KILLING",	"GOBP_HEMOSTASIS",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	
                                    "GOBP_NUCLEOSIDE_DIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_GTP_BIOSYNTHETIC_PROCESS",	"GOBP_DETECTION_OF_EXTERNAL_BIOTIC_STIMULUS",	"GOBP_PROTEIN_ACTIVATION_CASCADE",	
                                    "GOBP_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	
                                    "GOBP_POSITIVE_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	"GOBP_PYRIMIDINE_RIBONUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_KILLING",	
                                    "GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_RECEPTOR_BINDING",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_DETOXIFICATION_OF_COPPER_ION",	"GOBP_PLASMINOGEN_ACTIVATION",	
                                    "GOBP_PLATELET_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_PLASMINOGEN_ACTIVATION",	"GOBP_ISG15_PROTEIN_CONJUGATION",	"GOBP_NUCLEOSIDE_DIPHOSPHATE_METABOLIC_PROCESS",	
                                    "GOBP_PURINE_NUCLEOSIDE_DIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_PLASMA_MEMBRANE_REPAIR",	"GOBP_ENTRY_INTO_HOST",	"GOBP_INTEGRIN_MEDIATED_SIGNALING_PATHWAY",	
                                    "GOBP_PYRIMIDINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_ENDOTHELIAL_CELL_PROLIFERATION",	"GOBP_REGULATION_OF_PROTEIN_MATURATION",	
                                    "GOBP_PYRIMIDINE_NUCLEOTIDE_METABOLIC_PROCESS",	"GOBP_CELL_SUBSTRATE_ADHESION",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_NEGATIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	
                                    "GOBP_PYRIMIDINE_RIBONUCLEOTIDE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_CELL_SUBSTRATE_JUNCTION_ORGANIZATION",	"GOBP_WOUND_HEALING",	"GOBP_ACTOMYOSIN_STRUCTURE_ORGANIZATION",	
                                    "GOBP_ACTIN_FILAMENT_BUNDLE_ORGANIZATION",	"GOBP_NUCLEOTIDE_PHOSPHORYLATION",	"GOBP_REGULATION_OF_T_CELL_MEDIATED_IMMUNITY",	"GOBP_CELL_MATRIX_ADHESION",	
                                    "GOBP_POSITIVE_REGULATION_OF_CELL_SUBSTRATE_ADHESION",	"GOBP_CELL_SUBSTRATE_JUNCTION_ORGANIZATION",	"GOBP_REGULATION_OF_FIBROBLAST_PROLIFERATION",	
                                    "GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_REGULATION_OF_PHAGOCYTOSIS",	"GOBP_VIRAL_GENOME_REPLICATION",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_ADP_METABOLIC_PROCESS",	
                                    "GOBP_NUCLEOSIDE_MONOPHOSPHATE_METABOLIC_PROCESS",	"GOBP_T_CELL_MEDIATED_IMMUNITY",	"GOBP_STRESS_FIBER_ASSEMBLY",	"GOBP_MAINTENANCE_OF_PROTEIN_LOCATION",	
                                    "GOBP_RIBONUCLEOSIDE_DIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	
                                    "GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_CELL_KILLING",	"GOBP_FIBROBLAST_PROLIFERATION",	"GOBP_FOCAL_ADHESION_ASSEMBLY",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	
                                    "GOBP_REGULATION_OF_BODY_FLUID_LEVELS",	"GOBP_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	
                                    "GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",	"GOBP_MUSCLE_CONTRACTION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_CELL_MATRIX_ADHESION",	
                                    "GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_REGULATION_OF_CELL_JUNCTION_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	
                                    "GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_BONE_DEVELOPMENT",	"GOBP_REGULATION_OF_CELL_MORPHOGENESIS",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	
                                    "GOBP_ADAPTIVE_IMMUNE_RESPONSE_BASED_ON_SOMATIC_RECOMBINATION_OF_IMMUNE_RECEPTORS_BUILT_FROM_IMMUNOGLOBULIN_SUPERFAMILY_DOMAINS",	"GOBP_CELL_JUNCTION_ASSEMBLY",	"GOBP_RESPONSE_TO_VIRUS",	
                                    "GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	
                                    "GOBP_CYTOKINE_MEDIATED_SIGNALING_PATHWAY")
)
ggvenn(a, c("MaT4vs.all_pos_GOBP","McT4vs.all_pos_GOBP" )) 


MaT4_McT4_pos_GOBP_intersect= as.data.frame(Reduce(intersect, list(set1,set2)))
write.csv(MaT4_McT4_pos_GOBP_intersect, "_MaT4_McT4_epi_pos_GOBP_intersect.csv")



##data from excel##
# REACTOME
######use format -> cells -> custom -> \'@\'\, ########
MaT4 = c("REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	
         "REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_COLLAGEN_BIOSYNTHESIS_AND_MODIFYING_ENZYMES",	
         "REACTOME_BINDING_AND_UPTAKE_OF_LIGANDS_BY_SCAVENGER_RECEPTORS",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	
         "REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_ROCKS",	"REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",
         "REACTOME_COLLAGEN_FORMATION",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION",	
         "REACTOME_REGULATION_OF_INSULIN_LIKE_GROWTH_FACTOR_IGF_TRANSPORT_AND_UPTAKE_BY_INSULIN_LIKE_GROWTH_FACTOR_BINDING_PROTEINS_IGFBPS",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",	"REACTOME_HSF1_ACTIVATION",	
         "REACTOME_CHAPERONE_MEDIATED_AUTOPHAGY",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_RHO_GTPASES_ACTIVATE_CIT",	
         "REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	
         "REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_RESPONSE_OF_EIF2AK4_GCN2_TO_AMINO_ACID_DEFICIENCY",	
         "REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",	
         "REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_AMYLOID_FIBER_FORMATION",	"REACTOME_ECM_PROTEOGLYCANS",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_PCP_CE_PATHWAY",	
         "REACTOME_AGGREPHAGY",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_LATE_ENDOSOMAL_MICROAUTOPHAGY",	"REACTOME_CELLULAR_RESPONSE_TO_STARVATION",
         "REACTOME_SELENOAMINO_ACID_METABOLISM",	"REACTOME_COLLAGEN_DEGRADATION",	"REACTOME_ELASTIC_FIBRE_FORMATION",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",
         "REACTOME_SCAVENGING_BY_CLASS_A_RECEPTORS",	"REACTOME_REGULATION_OF_EXPRESSION_OF_SLITS_AND_ROBOS",	"REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",	
         "REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_EXTRACELLULAR_MATRIX_ORGANIZATION",	
         "REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_ASSEMBLY_OF_THE_HIV_VIRION",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_ACTIVATION_OF_THE_AP_1_FAMILY_OF_TRANSCRIPTION_FACTORS",	
         "REACTOME_REGULATION_OF_BACH1_ACTIVITY",	"REACTOME_NEF_MEDIATES_DOWN_MODULATION_OF_CELL_SURFACE_RECEPTORS_BY_RECRUITING_THEM_TO_CLATHRIN_ADAPTERS",	
         "REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_COPI_MEDIATED_ANTEROGRADE_TRANSPORT",	"REACTOME_ACTIVATION_OF_MATRIX_METALLOPROTEINASES",	
         "REACTOME_DOWNREGULATION_OF_ERBB2_ERBB3_SIGNALING",	"REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_POST_CHAPERONIN_TUBULIN_FOLDING_PATHWAY",	"REACTOME_SIGNALING_BY_ROBO_RECEPTORS",
         "REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",
         "REACTOME_ASSEMBLY_OF_COLLAGEN_FIBRILS_AND_OTHER_MULTIMERIC_STRUCTURES",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_RHO_GTPASES_ACTIVATE_PAKS",
         "REACTOME_PLASMA_LIPOPROTEIN_CLEARANCE",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_COPI_DEPENDENT_GOLGI_TO_ER_RETROGRADE_TRAFFIC",	"REACTOME_ENDOSOMAL_VACUOLAR_PATHWAY",
         "REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_SYNDECAN_INTERACTIONS",	"REACTOME_ALPHA_PROTEIN_KINASE_1_SIGNALING_PATHWAY",
         "REACTOME_NONSENSE_MEDIATED_DECAY_NMD",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_DEGRADATION_OF_THE_EXTRACELLULAR_MATRIX",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	
         "REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_VLDLR_INTERNALISATION_AND_DEGRADATION",	"REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_CELL_EXTRACELLULAR_MATRIX_INTERACTIONS",	"REACTOME_EPH_EPHRIN_SIGNALING",
         "REACTOME_ER_TO_GOLGI_ANTEROGRADE_TRANSPORT",	"REACTOME_MOLECULES_ASSOCIATED_WITH_ELASTIC_FIBRES",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",
         "REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_HEDGEHOG_ON_STATE",	"REACTOME_EPHA_MEDIATED_GROWTH_CONE_COLLAPSE",
         "REACTOME_RHO_GTPASES_ACTIVATE_PKNS",	"REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_PLASMA_LIPOPROTEIN_ASSEMBLY_REMODELING_AND_CLEARANCE",	"REACTOME_MUSCLE_CONTRACTION",	
         "REACTOME_CALNEXIN_CALRETICULIN_CYCLE",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_CHONDROITIN_SULFATE_DERMATAN_SULFATE_METABOLISM",	"REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",
         "REACTOME_MITOPHAGY",	"REACTOME_UCH_PROTEINASES",	"REACTOME_IRON_UPTAKE_AND_TRANSPORT",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_O_GLYCOSYLATION_OF_TSR_DOMAIN_CONTAINING_PROTEINS",
         "REACTOME_COLLAGEN_CHAIN_TRIMERIZATION",	"REACTOME_SODIUM_CALCIUM_EXCHANGERS",	"REACTOME_TRANSPORT_TO_THE_GOLGI_AND_SUBSEQUENT_MODIFICATION",	"REACTOME_SIGNALING_BY_HEDGEHOG",	
         "REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_MODULATION_BY_MTB_OF_HOST_IMMUNE_SYSTEM",	"REACTOME_GPER1_SIGNALING",	"REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	
         "REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_NEF_MEDIATED_CD4_DOWN_REGULATION",	"REACTOME_TRANSLATION",	"REACTOME_AUTOPHAGY",	"REACTOME_N_GLYCAN_TRIMMING_IN_THE_ER_AND_CALNEXIN_CALRETICULIN_CYCLE",
         "REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_IRE1ALPHA_ACTIVATES_CHAPERONES",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",
         "REACTOME_E3_UBIQUITIN_LIGASES_UBIQUITINATE_TARGET_PROTEINS",	"REACTOME_CROSSLINKING_OF_COLLAGEN_FIBRILS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	
         "REACTOME_ASPARAGINE_N_LINKED_GLYCOSYLATION",	"REACTOME_TP53_REGULATES_METABOLIC_GENES",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	
         "REACTOME_COPII_MEDIATED_VESICLE_TRANSPORT",	"REACTOME_INTEGRIN_CELL_SURFACE_INTERACTIONS",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",
         "REACTOME_NERVOUS_SYSTEM_DEVELOPMENT",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_DNA_REPLICATION_PRE_INITIATION",
         "REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_CYTOPROTECTION_BY_HMOX1",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",	"REACTOME_SYNTHESIS_OF_DNA",
         "REACTOME_KEAP1_NFE2L2_PATHWAY",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",	"REACTOME_SARS_COV_2_HOST_INTERACTIONS",	"REACTOME_METABOLISM_OF_AMINO_ACIDS_AND_DERIVATIVES",
         "REACTOME_INTRA_GOLGI_AND_RETROGRADE_GOLGI_TO_ER_TRAFFIC",	"REACTOME_DNA_REPLICATION",	"REACTOME_RRNA_PROCESSING",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	
         "REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_PTEN_REGULATION",	"REACTOME_PROTEIN_LOCALIZATION",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",
         "REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",	"REACTOME_APOPTOSIS",	"REACTOME_NEDDYLATION",	"REACTOME_HIV_INFECTION",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_SARS_COV_2_INFECTION",	
         "REACTOME_MEMBRANE_TRAFFICKING",	"REACTOME_RHO_GTPASE_EFFECTORS",	"REACTOME_HEMOSTASIS",	"REACTOME_SARS_COV_INFECTIONS",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES"
)
McT4 = c("REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_COLLAGEN_FORMATION",
         "REACTOME_COLLAGEN_BIOSYNTHESIS_AND_MODIFYING_ENZYMES",	"REACTOME_EXTRACELLULAR_MATRIX_ORGANIZATION",	"REACTOME_ECM_PROTEOGLYCANS",	"REACTOME_CELL_EXTRACELLULAR_MATRIX_INTERACTIONS",	
         "REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_ASSEMBLY_OF_COLLAGEN_FIBRILS_AND_OTHER_MULTIMERIC_STRUCTURES",	
         "REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_DEGRADATION_OF_THE_EXTRACELLULAR_MATRIX",	"REACTOME_INTEGRIN_CELL_SURFACE_INTERACTIONS",
         "REACTOME_REGULATION_OF_INSULIN_LIKE_GROWTH_FACTOR_IGF_TRANSPORT_AND_UPTAKE_BY_INSULIN_LIKE_GROWTH_FACTOR_BINDING_PROTEINS_IGFBPS",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",	"REACTOME_INTERFERON_SIGNALING",	
         "REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",	"REACTOME_COLLAGEN_DEGRADATION",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_SYNDECAN_INTERACTIONS",	
         "REACTOME_ELASTIC_FIBRE_FORMATION",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_ENDOSOMAL_VACUOLAR_PATHWAY",	"REACTOME_COLLAGEN_CHAIN_TRIMERIZATION",	"REACTOME_NON_INTEGRIN_MEMBRANE_ECM_INTERACTIONS",	
         "REACTOME_CROSSLINKING_OF_COLLAGEN_FIBRILS",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_COPI_DEPENDENT_GOLGI_TO_ER_RETROGRADE_TRAFFIC",	
         "REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_REGULATION_OF_CYTOSKELETAL_REMODELING_AND_CELL_SPREADING_BY_IPP_COMPLEX_COMPONENTS",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	
         "REACTOME_GPER1_SIGNALING",	"REACTOME_NCAM1_INTERACTIONS",	"REACTOME_EPHA_MEDIATED_GROWTH_CONE_COLLAPSE",	"REACTOME_MUSCLE_CONTRACTION",	"REACTOME_ACTIVATION_OF_MATRIX_METALLOPROTEINASES",
         "REACTOME_RECYCLING_PATHWAY_OF_L1",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_PD_1_SIGNALING",	"REACTOME_SEMAPHORIN_INTERACTIONS",	
         "REACTOME_DAP12_INTERACTIONS",	"REACTOME_RHO_GTPASES_ACTIVATE_CIT",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_GAP_JUNCTION_ASSEMBLY",	"REACTOME_TRANSPORT_OF_CONNEXONS_TO_THE_PLASMA_MEMBRANE",	
         "REACTOME_MET_ACTIVATES_PTK2_SIGNALING",	"REACTOME_THE_ROLE_OF_NEF_IN_HIV_1_REPLICATION_AND_DISEASE_PATHOGENESIS",	"REACTOME_AZATHIOPRINE_ADME",	"REACTOME_OAS_ANTIVIRAL_RESPONSE",	
         "REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	"REACTOME_ER_TO_GOLGI_ANTEROGRADE_TRANSPORT",	"REACTOME_METALLOTHIONEINS_BIND_METALS",	"REACTOME_EPH_EPHRIN_MEDIATED_REPULSION_OF_CELLS",	
         "REACTOME_RHO_GTPASES_ACTIVATE_IQGAPS",	"REACTOME_PARASITE_INFECTION",	"REACTOME_CELL_JUNCTION_ORGANIZATION",	"REACTOME_COPI_MEDIATED_ANTEROGRADE_TRANSPORT",	"REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",
         "REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_MET_PROMOTES_CELL_MOTILITY",	"REACTOME_TRANSPORT_TO_THE_GOLGI_AND_SUBSEQUENT_MODIFICATION",	"REACTOME_SIGNALING_BY_MODERATE_KINASE_ACTIVITY_BRAF_MUTANTS",
         "REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_L1CAM_INTERACTIONS",	"REACTOME_METABOLISM_OF_NUCLEOTIDES",	"REACTOME_RHO_GTPASE_EFFECTORS",	
         "REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_CELL_SURFACE_INTERACTIONS_AT_THE_VASCULAR_WALL",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	
         "REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_G_ALPHA_S_SIGNALLING_EVENTS",	"REACTOME_SIGNALING_BY_BRAF_AND_RAF1_FUSIONS",	"REACTOME_COPII_MEDIATED_VESICLE_TRANSPORT",
         "REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING"
)

data_1 = MaT4
data_2 = McT4

library(VennDiagram)
library(RColorBrewer)

colors()

set1 = as.list(data_1)
set2 = as.list(data_2)

venn.diagram(x=list(set1, set2), 
             category.names = c("MaT4","McT4"), filename = "venndiagram_REACTOME_pos_MaT4_McT4_epi.png", output=T, 
             imagetype = "png", height = 2000, width = 2000, resolution = 300, compression = "lzw",
             fill=c("skyblue", "yellowgreen"), 
             cex = 2,  lwd=2, lty='blank', fontface="bold", fontfamily="sans", cat.cex=1, cat.fontface="bold", cat.default.pos="outer", cat.pos=c(0, 20),cat.fontfamily="sans")

a <- list("MaT4vs.all_pos_REACTOME" = c("REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	
                                        "REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_COLLAGEN_BIOSYNTHESIS_AND_MODIFYING_ENZYMES",	
                                        "REACTOME_BINDING_AND_UPTAKE_OF_LIGANDS_BY_SCAVENGER_RECEPTORS",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	
                                        "REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_ROCKS",	"REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",
                                        "REACTOME_COLLAGEN_FORMATION",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION",	
                                        "REACTOME_REGULATION_OF_INSULIN_LIKE_GROWTH_FACTOR_IGF_TRANSPORT_AND_UPTAKE_BY_INSULIN_LIKE_GROWTH_FACTOR_BINDING_PROTEINS_IGFBPS",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",	"REACTOME_HSF1_ACTIVATION",	
                                        "REACTOME_CHAPERONE_MEDIATED_AUTOPHAGY",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_RHO_GTPASES_ACTIVATE_CIT",	
                                        "REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	
                                        "REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_RESPONSE_OF_EIF2AK4_GCN2_TO_AMINO_ACID_DEFICIENCY",	
                                        "REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",	
                                        "REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_AMYLOID_FIBER_FORMATION",	"REACTOME_ECM_PROTEOGLYCANS",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_PCP_CE_PATHWAY",	
                                        "REACTOME_AGGREPHAGY",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_LATE_ENDOSOMAL_MICROAUTOPHAGY",	"REACTOME_CELLULAR_RESPONSE_TO_STARVATION",
                                        "REACTOME_SELENOAMINO_ACID_METABOLISM",	"REACTOME_COLLAGEN_DEGRADATION",	"REACTOME_ELASTIC_FIBRE_FORMATION",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",
                                        "REACTOME_SCAVENGING_BY_CLASS_A_RECEPTORS",	"REACTOME_REGULATION_OF_EXPRESSION_OF_SLITS_AND_ROBOS",	"REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",	
                                        "REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_EXTRACELLULAR_MATRIX_ORGANIZATION",	
                                        "REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_ASSEMBLY_OF_THE_HIV_VIRION",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_ACTIVATION_OF_THE_AP_1_FAMILY_OF_TRANSCRIPTION_FACTORS",	
                                        "REACTOME_REGULATION_OF_BACH1_ACTIVITY",	"REACTOME_NEF_MEDIATES_DOWN_MODULATION_OF_CELL_SURFACE_RECEPTORS_BY_RECRUITING_THEM_TO_CLATHRIN_ADAPTERS",	
                                        "REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_COPI_MEDIATED_ANTEROGRADE_TRANSPORT",	"REACTOME_ACTIVATION_OF_MATRIX_METALLOPROTEINASES",	
                                        "REACTOME_DOWNREGULATION_OF_ERBB2_ERBB3_SIGNALING",	"REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_POST_CHAPERONIN_TUBULIN_FOLDING_PATHWAY",	"REACTOME_SIGNALING_BY_ROBO_RECEPTORS",
                                        "REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",
                                        "REACTOME_ASSEMBLY_OF_COLLAGEN_FIBRILS_AND_OTHER_MULTIMERIC_STRUCTURES",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_RHO_GTPASES_ACTIVATE_PAKS",
                                        "REACTOME_PLASMA_LIPOPROTEIN_CLEARANCE",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_COPI_DEPENDENT_GOLGI_TO_ER_RETROGRADE_TRAFFIC",	"REACTOME_ENDOSOMAL_VACUOLAR_PATHWAY",
                                        "REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_SYNDECAN_INTERACTIONS",	"REACTOME_ALPHA_PROTEIN_KINASE_1_SIGNALING_PATHWAY",
                                        "REACTOME_NONSENSE_MEDIATED_DECAY_NMD",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_DEGRADATION_OF_THE_EXTRACELLULAR_MATRIX",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	
                                        "REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_VLDLR_INTERNALISATION_AND_DEGRADATION",	"REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_CELL_EXTRACELLULAR_MATRIX_INTERACTIONS",	"REACTOME_EPH_EPHRIN_SIGNALING",
                                        "REACTOME_ER_TO_GOLGI_ANTEROGRADE_TRANSPORT",	"REACTOME_MOLECULES_ASSOCIATED_WITH_ELASTIC_FIBRES",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",
                                        "REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_HEDGEHOG_ON_STATE",	"REACTOME_EPHA_MEDIATED_GROWTH_CONE_COLLAPSE",
                                        "REACTOME_RHO_GTPASES_ACTIVATE_PKNS",	"REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_PLASMA_LIPOPROTEIN_ASSEMBLY_REMODELING_AND_CLEARANCE",	"REACTOME_MUSCLE_CONTRACTION",	
                                        "REACTOME_CALNEXIN_CALRETICULIN_CYCLE",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_CHONDROITIN_SULFATE_DERMATAN_SULFATE_METABOLISM",	"REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",
                                        "REACTOME_MITOPHAGY",	"REACTOME_UCH_PROTEINASES",	"REACTOME_IRON_UPTAKE_AND_TRANSPORT",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_O_GLYCOSYLATION_OF_TSR_DOMAIN_CONTAINING_PROTEINS",
                                        "REACTOME_COLLAGEN_CHAIN_TRIMERIZATION",	"REACTOME_SODIUM_CALCIUM_EXCHANGERS",	"REACTOME_TRANSPORT_TO_THE_GOLGI_AND_SUBSEQUENT_MODIFICATION",	"REACTOME_SIGNALING_BY_HEDGEHOG",	
                                        "REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_MODULATION_BY_MTB_OF_HOST_IMMUNE_SYSTEM",	"REACTOME_GPER1_SIGNALING",	"REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	
                                        "REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_NEF_MEDIATED_CD4_DOWN_REGULATION",	"REACTOME_TRANSLATION",	"REACTOME_AUTOPHAGY",	"REACTOME_N_GLYCAN_TRIMMING_IN_THE_ER_AND_CALNEXIN_CALRETICULIN_CYCLE",
                                        "REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_IRE1ALPHA_ACTIVATES_CHAPERONES",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",
                                        "REACTOME_E3_UBIQUITIN_LIGASES_UBIQUITINATE_TARGET_PROTEINS",	"REACTOME_CROSSLINKING_OF_COLLAGEN_FIBRILS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	
                                        "REACTOME_ASPARAGINE_N_LINKED_GLYCOSYLATION",	"REACTOME_TP53_REGULATES_METABOLIC_GENES",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	
                                        "REACTOME_COPII_MEDIATED_VESICLE_TRANSPORT",	"REACTOME_INTEGRIN_CELL_SURFACE_INTERACTIONS",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",
                                        "REACTOME_NERVOUS_SYSTEM_DEVELOPMENT",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_DNA_REPLICATION_PRE_INITIATION",
                                        "REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_CYTOPROTECTION_BY_HMOX1",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",	"REACTOME_SYNTHESIS_OF_DNA",
                                        "REACTOME_KEAP1_NFE2L2_PATHWAY",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",	"REACTOME_SARS_COV_2_HOST_INTERACTIONS",	"REACTOME_METABOLISM_OF_AMINO_ACIDS_AND_DERIVATIVES",
                                        "REACTOME_INTRA_GOLGI_AND_RETROGRADE_GOLGI_TO_ER_TRAFFIC",	"REACTOME_DNA_REPLICATION",	"REACTOME_RRNA_PROCESSING",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	
                                        "REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_PTEN_REGULATION",	"REACTOME_PROTEIN_LOCALIZATION",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",
                                        "REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",	"REACTOME_APOPTOSIS",	"REACTOME_NEDDYLATION",	"REACTOME_HIV_INFECTION",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_SARS_COV_2_INFECTION",	
                                        "REACTOME_MEMBRANE_TRAFFICKING",	"REACTOME_RHO_GTPASE_EFFECTORS",	"REACTOME_HEMOSTASIS",	"REACTOME_SARS_COV_INFECTIONS",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES"),
          
          "McT4vs.all_pos_REACTOME" = c("REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_COLLAGEN_FORMATION",
                                        "REACTOME_COLLAGEN_BIOSYNTHESIS_AND_MODIFYING_ENZYMES",	"REACTOME_EXTRACELLULAR_MATRIX_ORGANIZATION",	"REACTOME_ECM_PROTEOGLYCANS",	"REACTOME_CELL_EXTRACELLULAR_MATRIX_INTERACTIONS",	
                                        "REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_ASSEMBLY_OF_COLLAGEN_FIBRILS_AND_OTHER_MULTIMERIC_STRUCTURES",	
                                        "REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_DEGRADATION_OF_THE_EXTRACELLULAR_MATRIX",	"REACTOME_INTEGRIN_CELL_SURFACE_INTERACTIONS",
                                        "REACTOME_REGULATION_OF_INSULIN_LIKE_GROWTH_FACTOR_IGF_TRANSPORT_AND_UPTAKE_BY_INSULIN_LIKE_GROWTH_FACTOR_BINDING_PROTEINS_IGFBPS",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",	"REACTOME_INTERFERON_SIGNALING",	
                                        "REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",	"REACTOME_COLLAGEN_DEGRADATION",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_SYNDECAN_INTERACTIONS",	
                                        "REACTOME_ELASTIC_FIBRE_FORMATION",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_ENDOSOMAL_VACUOLAR_PATHWAY",	"REACTOME_COLLAGEN_CHAIN_TRIMERIZATION",	"REACTOME_NON_INTEGRIN_MEMBRANE_ECM_INTERACTIONS",	
                                        "REACTOME_CROSSLINKING_OF_COLLAGEN_FIBRILS",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_COPI_DEPENDENT_GOLGI_TO_ER_RETROGRADE_TRAFFIC",	
                                        "REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_REGULATION_OF_CYTOSKELETAL_REMODELING_AND_CELL_SPREADING_BY_IPP_COMPLEX_COMPONENTS",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	
                                        "REACTOME_GPER1_SIGNALING",	"REACTOME_NCAM1_INTERACTIONS",	"REACTOME_EPHA_MEDIATED_GROWTH_CONE_COLLAPSE",	"REACTOME_MUSCLE_CONTRACTION",	"REACTOME_ACTIVATION_OF_MATRIX_METALLOPROTEINASES",
                                        "REACTOME_RECYCLING_PATHWAY_OF_L1",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_PD_1_SIGNALING",	"REACTOME_SEMAPHORIN_INTERACTIONS",	
                                        "REACTOME_DAP12_INTERACTIONS",	"REACTOME_RHO_GTPASES_ACTIVATE_CIT",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_GAP_JUNCTION_ASSEMBLY",	"REACTOME_TRANSPORT_OF_CONNEXONS_TO_THE_PLASMA_MEMBRANE",	
                                        "REACTOME_MET_ACTIVATES_PTK2_SIGNALING",	"REACTOME_THE_ROLE_OF_NEF_IN_HIV_1_REPLICATION_AND_DISEASE_PATHOGENESIS",	"REACTOME_AZATHIOPRINE_ADME",	"REACTOME_OAS_ANTIVIRAL_RESPONSE",	
                                        "REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	"REACTOME_ER_TO_GOLGI_ANTEROGRADE_TRANSPORT",	"REACTOME_METALLOTHIONEINS_BIND_METALS",	"REACTOME_EPH_EPHRIN_MEDIATED_REPULSION_OF_CELLS",	
                                        "REACTOME_RHO_GTPASES_ACTIVATE_IQGAPS",	"REACTOME_PARASITE_INFECTION",	"REACTOME_CELL_JUNCTION_ORGANIZATION",	"REACTOME_COPI_MEDIATED_ANTEROGRADE_TRANSPORT",	"REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",
                                        "REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_MET_PROMOTES_CELL_MOTILITY",	"REACTOME_TRANSPORT_TO_THE_GOLGI_AND_SUBSEQUENT_MODIFICATION",	"REACTOME_SIGNALING_BY_MODERATE_KINASE_ACTIVITY_BRAF_MUTANTS",
                                        "REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_L1CAM_INTERACTIONS",	"REACTOME_METABOLISM_OF_NUCLEOTIDES",	"REACTOME_RHO_GTPASE_EFFECTORS",	
                                        "REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_CELL_SURFACE_INTERACTIONS_AT_THE_VASCULAR_WALL",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	
                                        "REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_G_ALPHA_S_SIGNALLING_EVENTS",	"REACTOME_SIGNALING_BY_BRAF_AND_RAF1_FUSIONS",	"REACTOME_COPII_MEDIATED_VESICLE_TRANSPORT",
                                        "REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING"))

ggvenn(a, c("MaT4vs.all_pos_REACTOME","McT4vs.all_pos_REACTOME" )) 


MaT4_McT4_pos_REACTOME_intersect= as.data.frame(Reduce(intersect, list(set1,set2)))
write.csv(MaT4_McT4_pos_REACTOME_intersect, "_MaT4_McT4_epi_pos_REACTOME_intersect.csv")
