==> 2_venndiagram_from_fgsea_MaT4_and_McT4_Myeloid_cell_GOBP_REACTOME.R <==
#2023.04.21
setwd('F:/Ko/67ESCC_PCN_integration_07-19-22/immune_cell_T4_M4/fgsea')

library(ggplot2)
library(RTCGA)
library(dplyr)
library(RTCGAToolbox)
library(TCGAbiolinks)
library(data.table)
#install.packages("RDocumentation")
library(RDocumentation)


##data from excel##
# GOBP
######use format -> cells -> custom -> \'@\'\, ########
MaT4 = c("GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_ATP_BIOSYNTHETIC_PROCESS",	"GOBP_PROTEIN_REFOLDING",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",	"GOBP_NEUTROPHIL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_MYELOID_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",	"GOBP_MICROTUBULE_BASED_PROTEIN_TRANSPORT",	"GOBP_REGULATION_OF_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_REGULATION_OF_NEUTROPHIL_ACTIVATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_I",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_SUPEROXIDE_ANION_GENERATION",	"GOBP_NEUTROPHIL_DEGRANULATION",	"GOBP_REGULATION_OF_SUPEROXIDE_ANION_GENERATION",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_DEGRANULATION",	"GOBP_REGULATION_OF_LEUKOCYTE_DEGRANULATION",	"GOBP_SKELETAL_MUSCLE_CONTRACTION",	"GOBP_T_CELL_ACTIVATION_VIA_T_CELL_RECEPTOR_CONTACT_WITH_ANTIGEN_BOUND_TO_MHC_MOLECULE_ON_ANTIGEN_PRESENTING_CELL",	"GOBP_ICOSANOID_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CHEMOKINE_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	"GOBP_DE_NOVO_PROTEIN_FOLDING",	"GOBP_ATP_METABOLIC_PROCESS",	"GOBP_CHAPERONE_COFACTOR_DEPENDENT_PROTEIN_REFOLDING",	"GOBP_NEGATIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_CERAMIDE_CATABOLIC_PROCESS",	"GOBP_SUPEROXIDE_METABOLIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_GRANULOCYTE_ACTIVATION",	"GOBP_MACROPHAGE_ACTIVATION",	"GOBP_HEAT_ACCLIMATION",	"GOBP_ELECTRON_TRANSPORT_CHAIN",	"GOBP_AEROBIC_RESPIRATION",	"GOBP_POSITIVE_REGULATION_OF_NATURAL_KILLER_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_CHEMOKINE_PRODUCTION",	"GOBP_INCLUSION_BODY_ASSEMBLY",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_POSITIVE_REGULATION_OF_PHAGOCYTOSIS",	"GOBP_REGULATION_OF_PHAGOCYTOSIS",	"GOBP_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_POSITIVE_REGULATION_OF_LIPOPOLYSACCHARIDE_MEDIATED_SIGNALING_PATHWAY",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_BETA",	"GOBP_RESPONSE_TO_INTERFERON_ALPHA",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_LEUKOCYTE_ACTIVATION_INVOLVED_IN_INFLAMMATORY_RESPONSE",	"GOBP_MEMBRANE_LIPID_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_MAST_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_PLASMA_LIPOPROTEIN_PARTICLE_CLEARANCE",	"GOBP_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_MAST_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_INCLUSION_BODY_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_CELL_KILLING",	"GOBP_GLYCOLIPID_CATABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_PHAGOSOME_ACIDIFICATION",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_CYTOCHROME_C_TO_OXYGEN",	"GOBP_NEUROTRANSMITTER_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_LIPOPROTEIN_PARTICLE_CLEARANCE",	"GOBP_POSITIVE_REGULATION_OF_RECEPTOR_INTERNALIZATION",	"GOBP_GLUCOCORTICOID_METABOLIC_PROCESS",	"GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_CYTOKINE_PRODUCTION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_INTERLEUKIN_10_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_1_PRODUCTION",	"GOBP_REGULATION_OF_SUPEROXIDE_METABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_I_VIA_ER_PATHWAY",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_INTERLEUKIN_6_PRODUCTION",	"GOBP_FC_GAMMA_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION_VIA_T_CELL_RECEPTOR_CONTACT_WITH_ANTIGEN_BOUND_TO_MHC_MOLECULE_ON_ANTIGEN_PRESENTING_CELL",	"GOBP_NEGATIVE_REGULATION_OF_ACTIN_NUCLEATION",	"GOBP_NEGATIVE_REGULATION_OF_ARP2_3_COMPLEX_MEDIATED_ACTIN_NUCLEATION",	"GOBP_LEUKOCYTE_MIGRATION_INVOLVED_IN_INFLAMMATORY_RESPONSE",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",	"GOBP_MACROPHAGE_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_PROTEIN_LOCALIZATION_TO_PRESYNAPSE",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_INVOLVED_IN_INFLAMMATORY_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_8_PRODUCTION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_RESPONSE_TO_SALT_STRESS",	"GOBP_CELLULAR_RESPIRATION",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_ACTIN_CYTOSKELETON",	"GOBP_REGULATION_OF_BLOOD_VESSEL_REMODELING",	"GOBP_NEGATIVE_REGULATION_OF_MAST_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_INTERLEUKIN_6_PRODUCTION",	"GOBP_REGULATION_OF_NEUTROPHIL_DEGRANULATION",	"GOBP_POSITIVE_REGULATION_OF_MONOCYTE_CHEMOTACTIC_PROTEIN_1_PRODUCTION",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_STRUCTURE",	"GOBP_NEGATIVE_REGULATION_OF_MYOBLAST_DIFFERENTIATION",	"GOBP_PROTEIN_FOLDING",	"GOBP_CHOLESTEROL_EFFLUX",	"GOBP_POSITIVE_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_GLIAL_CELL_ACTIVATION",	"GOBP_MAST_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_MAST_CELL_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_PHOSPHATIDYLINOSITOL_3_KINASE_SIGNALING",	"GOBP_CYTOKINE_PRODUCTION_INVOLVED_IN_INFLAMMATORY_RESPONSE",	"GOBP_NEUROMUSCULAR_PROCESS_CONTROLLING_POSTURE",	"GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_REGULATION_OF_INCLUSION_BODY_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_AXO_DENDRITIC_PROTEIN_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_VITAMIN_D_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_CELL_DIFFERENTIATION",	"GOBP_NEGATIVE_REGULATION_OF_LIPOPROTEIN_PARTICLE_CLEARANCE",	"GOBP_INTERLEUKIN_1_PRODUCTION",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_TAU_PROTEIN_KINASE_ACTIVITY",	"GOBP_NEGATIVE_REGULATION_OF_EXOCYTOSIS",	"GOBP_RESPIRATORY_BURST",	"GOBP_DETECTION_OF_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_4_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_MONOCYTE_CHEMOTACTIC_PROTEIN_1_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_OSTEOCLAST_DIFFERENTIATION",	"GOBP_INTRACELLULAR_PH_REDUCTION",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_CELLULAR_DEFENSE_RESPONSE",	"GOBP_INTERLEUKIN_10_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_DEGRANULATION",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_LEUKOCYTE_DIFFERENTIATION",	"GOBP_VACUOLAR_ACIDIFICATION",	"GOBP_NEGATIVE_REGULATION_OF_REGULATED_SECRETORY_PATHWAY",	"GOBP_COLLAGEN_FIBRIL_ORGANIZATION",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_REGULATION_OF_CALCIUM_MEDIATED_SIGNALING",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_MYELOID_LEUKOCYTE_CYTOKINE_PRODUCTION",	"GOBP_REGULATION_OF_STEROL_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_REGULATED_SECRETORY_PATHWAY",	"GOBP_SUBSTANTIA_NIGRA_DEVELOPMENT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_OSSIFICATION",	"GOBP_REGULATION_OF_MAST_CELL_ACTIVATION",	"GOBP_REGULATION_OF_PLASMA_LIPOPROTEIN_PARTICLE_LEVELS",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_POSITIVE_REGULATION_OF_BONE_MINERALIZATION",	"GOBP_FOREBRAIN_GENERATION_OF_NEURONS",	"GOBP_NEGATIVE_REGULATION_OF_SIGNALING_RECEPTOR_ACTIVITY",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_6_PRODUCTION",	"GOBP_CELL_KILLING",	"GOBP_TOLL_LIKE_RECEPTOR_4_SIGNALING_PATHWAY",	"GOBP_FC_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_RESPONSE_TO_PROGESTERONE",	"GOBP_COLLAGEN_METABOLIC_PROCESS",	"GOBP_NEUROINFLAMMATORY_RESPONSE",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_REGULATED_SECRETORY_PATHWAY",	"GOBP_RESPONSE_TO_AMYLOID_BETA",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	"GOBP_REGULATION_OF_CELL_KILLING",	"GOBP_POSITIVE_REGULATION_OF_PATTERN_RECOGNITION_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_ANTIGEN_RECEPTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_GLYCOSYL_COMPOUND_METABOLIC_PROCESS",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_MEDIATED_SIGNALING",	"GOBP_RESPONSE_TO_TYPE_I_INTERFERON",	"GOBP_LYTIC_VACUOLE_ORGANIZATION",	"GOBP_MITOCHONDRIAL_TRANSLATION",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_POSITIVE_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEURAL_NUCLEUS_DEVELOPMENT",	"GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_CELLULAR_RESPONSE_TO_AMYLOID_BETA",	"GOBP_REGULATION_OF_NITRIC_OXIDE_METABOLIC_PROCESS",	"GOBP_APOPTOTIC_CELL_CLEARANCE",	"GOBP_CELLULAR_RESPONSE_TO_CALCIUM_ION",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_ICOSANOID_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_NIK_NF_KAPPAB_SIGNALING",	"GOBP_REGULATION_OF_B_CELL_PROLIFERATION",	"GOBP_PROTEIN_TARGETING_TO_MITOCHONDRION",	"GOBP_VACUOLAR_LOCALIZATION",	"GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_INTERLEUKIN_8_PRODUCTION",	"GOBP_CYTOKINE_PRODUCTION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_RUFFLE_ASSEMBLY",	"GOBP_PROTEIN_LOCALIZATION_TO_LYSOSOME",	"GOBP_REACTIVE_NITROGEN_SPECIES_METABOLIC_PROCESS",	"GOBP_CELLULAR_RESPONSE_TO_HEAT",	"GOBP_SYNCYTIUM_FORMATION",	"GOBP_NEGATIVE_REGULATION_OF_SECRETION",	"GOBP_RESPONSE_TO_HEAT",	"GOBP_REGULATION_OF_PH",	"GOBP_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_HETEROTYPIC_CELL_CELL_ADHESION",	"GOBP_ACTIVATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_NEGATIVE_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_INTERFERON_GAMMA_PRODUCTION",	"GOBP_ENTRY_INTO_HOST",	"GOBP_PHAGOCYTOSIS",	"GOBP_CARBOHYDRATE_DERIVATIVE_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_MYELOID_LEUKOCYTE_DIFFERENTIATION",	"GOBP_INTERLEUKIN_12_PRODUCTION",	"GOBP_DEFENSE_RESPONSE_TO_GRAM_POSITIVE_BACTERIUM",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",	"GOBP_REGULATION_OF_MYELOID_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_INFLAMMATORY_RESPONSE",	"GOBP_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_CELLULAR_MONOVALENT_INORGANIC_CATION_HOMEOSTASIS",	"GOBP_MITOCHONDRIAL_GENE_EXPRESSION",	"GOBP_OSTEOCLAST_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_NEGATIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_ENDOTHELIAL_CELL_PROLIFERATION",	"GOBP_INTEGRIN_MEDIATED_SIGNALING_PATHWAY",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_REGULATION_OF_PHOSPHATIDYLINOSITOL_3_KINASE_SIGNALING",	"GOBP_PROTEIN_LOCALIZATION_TO_MITOCHONDRION",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_REGULATION_OF_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_CALCIUM_MEDIATED_SIGNALING",	"GOBP_PHOSPHATIDYLINOSITOL_3_KINASE_SIGNALING",	"GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_GRANULOCYTE_CHEMOTAXIS",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_REGULATION_OF_PROTEIN_DEPHOSPHORYLATION",	"GOBP_REGULATION_OF_PROTEIN_LOCALIZATION_TO_PLASMA_MEMBRANE",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_INOSITOL_LIPID_MEDIATED_SIGNALING",	"GOBP_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_RECEPTOR_MEDIATED_ENDOCYTOSIS",	"GOBP_REGULATION_OF_NIK_NF_KAPPAB_SIGNALING",	"GOBP_REGULATION_OF_ION_TRANSPORT",	"GOBP_INORGANIC_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_MYELOID_LEUKOCYTE_DIFFERENTIATION",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_REACTIVE_OXYGEN_SPECIES_METABOLIC_PROCESS",	"GOBP_CALCIUM_ION_HOMEOSTASIS",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_MONOVALENT_INORGANIC_CATION_HOMEOSTASIS",	"GOBP_RESPONSE_TO_TEMPERATURE_STIMULUS",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_CATION_TRANSMEMBRANE_TRANSPORT",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_POSITIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_REGULATION_OF_LEUKOCYTE_CHEMOTAXIS",	"GOBP_REGULATION_OF_CALCIUM_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_MYELOID_LEUKOCYTE_MIGRATION",	"GOBP_ENDOTHELIAL_CELL_MIGRATION",	"GOBP_CELLULAR_ION_HOMEOSTASIS",	"GOBP_REGULATION_OF_ENDOTHELIAL_CELL_MIGRATION",	"GOBP_REGULATION_OF_VESICLE_MEDIATED_TRANSPORT",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_PATTERN_RECOGNITION_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_LEUKOCYTE_PROLIFERATION",	"GOBP_REGULATION_OF_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_CATION_TRANSMEMBRANE_TRANSPORT",	"GOBP_VIRAL_PROCESS",	"GOBP_B_CELL_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_REGULATION_OF_PROTEIN_LOCALIZATION_TO_CELL_PERIPHERY",	"GOBP_LIPID_CATABOLIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_CELLULAR_LIPID_CATABOLIC_PROCESS",	"GOBP_MEMBRANE_LIPID_METABOLIC_PROCESS",	"GOBP_INFLAMMATORY_RESPONSE",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_RESPONSE_TO_METAL_ION",	"GOBP_CARBOHYDRATE_DERIVATIVE_BIOSYNTHETIC_PROCESS",	"GOBP_DIVALENT_INORGANIC_CATION_HOMEOSTASIS",	"GOBP_MITOCHONDRION_ORGANIZATION",	"GOBP_LEUKOCYTE_CHEMOTAXIS",	"GOBP_SECOND_MESSENGER_MEDIATED_SIGNALING",	"GOBP_ION_HOMEOSTASIS",	"GOBP_TISSUE_MIGRATION",	"GOBP_DEFENSE_RESPONSE_TO_BACTERIUM",	"GOBP_REGULATION_OF_BLOOD_CIRCULATION",	"GOBP_REGULATION_OF_TRANSPORTER_ACTIVITY",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_REGULATION_OF_METAL_ION_TRANSPORT",	"GOBP_SPHINGOLIPID_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_EXOCYTOSIS",	"GOBP_CELLULAR_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_SECRETION",	"GOBP_RESPONSE_TO_MOLECULE_OF_BACTERIAL_ORIGIN",	"GOBP_VACUOLE_ORGANIZATION",	"GOBP_STEROID_METABOLIC_PROCESS",	"GOBP_RESPONSE_TO_TUMOR_NECROSIS_FACTOR",	"GOBP_ORGANOPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_TRANSPORT",	"GOBP_HEMOSTASIS",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_EXOCYTOSIS",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_CELLULAR_METAL_ION_HOMEOSTASIS",	"GOBP_METAL_ION_HOMEOSTASIS",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_RESPONSE_TO_BACTERIUM",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_RESPONSE_TO_STEROID_HORMONE",	"GOBP_ALCOHOL_METABOLIC_PROCESS",	"GOBP_ANATOMICAL_STRUCTURE_HOMEOSTASIS",	"GOBP_MAINTENANCE_OF_LOCATION",	"GOBP_PROTEIN_TARGETING",	"GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_BODY_FLUID_LEVELS",	"GOBP_RESPONSE_TO_WOUNDING",	"GOBP_PROTEIN_POLYMERIZATION",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_CALCIUM_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_MYELOID_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_PROTEOLYSIS",	"GOBP_REGULATION_OF_LEUKOCYTE_DIFFERENTIATION",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_ORGANIC_HYDROXY_COMPOUND_METABOLIC_PROCESS",	"GOBP_CALCIUM_ION_TRANSPORT",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_INFLAMMATORY_RESPONSE",	"GOBP_CYTOKINE_MEDIATED_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE_BASED_ON_SOMATIC_RECOMBINATION_OF_IMMUNE_RECEPTORS_BUILT_FROM_IMMUNOGLOBULIN_SUPERFAMILY_DOMAINS",	"GOBP_SMALL_MOLECULE_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_HYDROLASE_ACTIVITY",	"GOBP_TAXIS",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_LOCOMOTION",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_ANATOMICAL_STRUCTURE_SIZE",	"GOBP_ACTIVATION_OF_IMMUNE_RESPONSE",	"GOBP_ENDOCYTOSIS",	"GOBP_SMALL_MOLECULE_CATABOLIC_PROCESS",	"GOBP_CELLULAR_HOMEOSTASIS",	"GOBP_LIPID_LOCALIZATION",	"GOBP_LEUKOCYTE_DIFFERENTIATION",	"GOBP_VESICLE_ORGANIZATION",	"GOBP_CELLULAR_RESPONSE_TO_ORGANIC_CYCLIC_COMPOUND",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_CELLULAR_RESPONSE_TO_LIPID",	"GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_METAL_ION_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CELL_DEATH",	"GOBP_PROCESS_UTILIZING_AUTOPHAGIC_MECHANISM",	"GOBP_T_CELL_ACTIVATION",
)
McT4 = c()

data_1 = MaT4
data_2 = McT4

library(VennDiagram)
library(RColorBrewer)
library(ggvenn)

colors()

set1 = as.list(data_1)
set2 = as.list(data_2)

venn.diagram(x=list(set1, set2), 
             category.names = c("MaT4","McT4"), filename = "venndiagram_GOBP_MaT4_McT4_epi.png", output=T, 
             imagetype = "png", height = 2000, width = 2000, resolution = 300, compression = "lzw",
             fill=c("skyblue", "yellowgreen"), 
             cex = 2,  lwd=2, lty='blank', fontface="bold", fontfamily="sans", cat.cex=1, cat.fontface="bold", cat.default.pos="outer", cat.pos=c(0, 0),cat.fontfamily="sans")


a <- list("MaT4vs.all_pos_GOBP" = c("GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_ATP_BIOSYNTHETIC_PROCESS",	"GOBP_PROTEIN_REFOLDING",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",	"GOBP_NEUTROPHIL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_MYELOID_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",	"GOBP_MICROTUBULE_BASED_PROTEIN_TRANSPORT",	"GOBP_REGULATION_OF_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_REGULATION_OF_NEUTROPHIL_ACTIVATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_I",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_SUPEROXIDE_ANION_GENERATION",	"GOBP_NEUTROPHIL_DEGRANULATION",	"GOBP_REGULATION_OF_SUPEROXIDE_ANION_GENERATION",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_DEGRANULATION",	"GOBP_REGULATION_OF_LEUKOCYTE_DEGRANULATION",	"GOBP_SKELETAL_MUSCLE_CONTRACTION",	"GOBP_T_CELL_ACTIVATION_VIA_T_CELL_RECEPTOR_CONTACT_WITH_ANTIGEN_BOUND_TO_MHC_MOLECULE_ON_ANTIGEN_PRESENTING_CELL",	"GOBP_ICOSANOID_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CHEMOKINE_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	"GOBP_DE_NOVO_PROTEIN_FOLDING",	"GOBP_ATP_METABOLIC_PROCESS",	"GOBP_CHAPERONE_COFACTOR_DEPENDENT_PROTEIN_REFOLDING",	"GOBP_NEGATIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_CERAMIDE_CATABOLIC_PROCESS",	"GOBP_SUPEROXIDE_METABOLIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_GRANULOCYTE_ACTIVATION",	"GOBP_MACROPHAGE_ACTIVATION",	"GOBP_HEAT_ACCLIMATION",	"GOBP_ELECTRON_TRANSPORT_CHAIN",	"GOBP_AEROBIC_RESPIRATION",	"GOBP_POSITIVE_REGULATION_OF_NATURAL_KILLER_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_CHEMOKINE_PRODUCTION",	"GOBP_INCLUSION_BODY_ASSEMBLY",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_POSITIVE_REGULATION_OF_PHAGOCYTOSIS",	"GOBP_REGULATION_OF_PHAGOCYTOSIS",	"GOBP_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_POSITIVE_REGULATION_OF_LIPOPOLYSACCHARIDE_MEDIATED_SIGNALING_PATHWAY",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_BETA",	"GOBP_RESPONSE_TO_INTERFERON_ALPHA",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_LEUKOCYTE_ACTIVATION_INVOLVED_IN_INFLAMMATORY_RESPONSE",	"GOBP_MEMBRANE_LIPID_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_MAST_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_PLASMA_LIPOPROTEIN_PARTICLE_CLEARANCE",	"GOBP_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_MAST_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_INCLUSION_BODY_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_CELL_KILLING",	"GOBP_GLYCOLIPID_CATABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_PHAGOSOME_ACIDIFICATION",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_CYTOCHROME_C_TO_OXYGEN",	"GOBP_NEUROTRANSMITTER_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_LIPOPROTEIN_PARTICLE_CLEARANCE",	"GOBP_POSITIVE_REGULATION_OF_RECEPTOR_INTERNALIZATION",	"GOBP_GLUCOCORTICOID_METABOLIC_PROCESS",	"GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_CYTOKINE_PRODUCTION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_INTERLEUKIN_10_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_1_PRODUCTION",	"GOBP_REGULATION_OF_SUPEROXIDE_METABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_I_VIA_ER_PATHWAY",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_INTERLEUKIN_6_PRODUCTION",	"GOBP_FC_GAMMA_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION_VIA_T_CELL_RECEPTOR_CONTACT_WITH_ANTIGEN_BOUND_TO_MHC_MOLECULE_ON_ANTIGEN_PRESENTING_CELL",	"GOBP_NEGATIVE_REGULATION_OF_ACTIN_NUCLEATION",	"GOBP_NEGATIVE_REGULATION_OF_ARP2_3_COMPLEX_MEDIATED_ACTIN_NUCLEATION",	"GOBP_LEUKOCYTE_MIGRATION_INVOLVED_IN_INFLAMMATORY_RESPONSE",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",	"GOBP_MACROPHAGE_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_PROTEIN_LOCALIZATION_TO_PRESYNAPSE",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_INVOLVED_IN_INFLAMMATORY_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_8_PRODUCTION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_RESPONSE_TO_SALT_STRESS",	"GOBP_CELLULAR_RESPIRATION",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_ACTIN_CYTOSKELETON",	"GOBP_REGULATION_OF_BLOOD_VESSEL_REMODELING",	"GOBP_NEGATIVE_REGULATION_OF_MAST_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_INTERLEUKIN_6_PRODUCTION",	"GOBP_REGULATION_OF_NEUTROPHIL_DEGRANULATION",	"GOBP_POSITIVE_REGULATION_OF_MONOCYTE_CHEMOTACTIC_PROTEIN_1_PRODUCTION",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_STRUCTURE",	"GOBP_NEGATIVE_REGULATION_OF_MYOBLAST_DIFFERENTIATION",	"GOBP_PROTEIN_FOLDING",	"GOBP_CHOLESTEROL_EFFLUX",	"GOBP_POSITIVE_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_GLIAL_CELL_ACTIVATION",	"GOBP_MAST_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_MAST_CELL_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_PHOSPHATIDYLINOSITOL_3_KINASE_SIGNALING",	"GOBP_CYTOKINE_PRODUCTION_INVOLVED_IN_INFLAMMATORY_RESPONSE",	"GOBP_NEUROMUSCULAR_PROCESS_CONTROLLING_POSTURE",	"GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_REGULATION_OF_INCLUSION_BODY_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_AXO_DENDRITIC_PROTEIN_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_VITAMIN_D_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_CELL_DIFFERENTIATION",	"GOBP_NEGATIVE_REGULATION_OF_LIPOPROTEIN_PARTICLE_CLEARANCE",	"GOBP_INTERLEUKIN_1_PRODUCTION",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_TAU_PROTEIN_KINASE_ACTIVITY",	"GOBP_NEGATIVE_REGULATION_OF_EXOCYTOSIS",	"GOBP_RESPIRATORY_BURST",	"GOBP_DETECTION_OF_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_4_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_MONOCYTE_CHEMOTACTIC_PROTEIN_1_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_OSTEOCLAST_DIFFERENTIATION",	"GOBP_INTRACELLULAR_PH_REDUCTION",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_CELLULAR_DEFENSE_RESPONSE",	"GOBP_INTERLEUKIN_10_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_DEGRANULATION",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_LEUKOCYTE_DIFFERENTIATION",	"GOBP_VACUOLAR_ACIDIFICATION",	"GOBP_NEGATIVE_REGULATION_OF_REGULATED_SECRETORY_PATHWAY",	"GOBP_COLLAGEN_FIBRIL_ORGANIZATION",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_REGULATION_OF_CALCIUM_MEDIATED_SIGNALING",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_MYELOID_LEUKOCYTE_CYTOKINE_PRODUCTION",	"GOBP_REGULATION_OF_STEROL_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_REGULATED_SECRETORY_PATHWAY",	"GOBP_SUBSTANTIA_NIGRA_DEVELOPMENT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_OSSIFICATION",	"GOBP_REGULATION_OF_MAST_CELL_ACTIVATION",	"GOBP_REGULATION_OF_PLASMA_LIPOPROTEIN_PARTICLE_LEVELS",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_POSITIVE_REGULATION_OF_BONE_MINERALIZATION",	"GOBP_FOREBRAIN_GENERATION_OF_NEURONS",	"GOBP_NEGATIVE_REGULATION_OF_SIGNALING_RECEPTOR_ACTIVITY",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_6_PRODUCTION",	"GOBP_CELL_KILLING",	"GOBP_TOLL_LIKE_RECEPTOR_4_SIGNALING_PATHWAY",	"GOBP_FC_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_RESPONSE_TO_PROGESTERONE",	"GOBP_COLLAGEN_METABOLIC_PROCESS",	"GOBP_NEUROINFLAMMATORY_RESPONSE",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_REGULATED_SECRETORY_PATHWAY",	"GOBP_RESPONSE_TO_AMYLOID_BETA",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	"GOBP_REGULATION_OF_CELL_KILLING",	"GOBP_POSITIVE_REGULATION_OF_PATTERN_RECOGNITION_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_ANTIGEN_RECEPTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_GLYCOSYL_COMPOUND_METABOLIC_PROCESS",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_MEDIATED_SIGNALING",	"GOBP_RESPONSE_TO_TYPE_I_INTERFERON",	"GOBP_LYTIC_VACUOLE_ORGANIZATION",	"GOBP_MITOCHONDRIAL_TRANSLATION",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_POSITIVE_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEURAL_NUCLEUS_DEVELOPMENT",	"GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_CELLULAR_RESPONSE_TO_AMYLOID_BETA",	"GOBP_REGULATION_OF_NITRIC_OXIDE_METABOLIC_PROCESS",	"GOBP_APOPTOTIC_CELL_CLEARANCE",	"GOBP_CELLULAR_RESPONSE_TO_CALCIUM_ION",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_ICOSANOID_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_NIK_NF_KAPPAB_SIGNALING",	"GOBP_REGULATION_OF_B_CELL_PROLIFERATION",	"GOBP_PROTEIN_TARGETING_TO_MITOCHONDRION",	"GOBP_VACUOLAR_LOCALIZATION",	"GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_INTERLEUKIN_8_PRODUCTION",	"GOBP_CYTOKINE_PRODUCTION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_RUFFLE_ASSEMBLY",	"GOBP_PROTEIN_LOCALIZATION_TO_LYSOSOME",	"GOBP_REACTIVE_NITROGEN_SPECIES_METABOLIC_PROCESS",	"GOBP_CELLULAR_RESPONSE_TO_HEAT",	"GOBP_SYNCYTIUM_FORMATION",	"GOBP_NEGATIVE_REGULATION_OF_SECRETION",	"GOBP_RESPONSE_TO_HEAT",	"GOBP_REGULATION_OF_PH",	"GOBP_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_HETEROTYPIC_CELL_CELL_ADHESION",	"GOBP_ACTIVATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_NEGATIVE_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_INTERFERON_GAMMA_PRODUCTION",	"GOBP_ENTRY_INTO_HOST",	"GOBP_PHAGOCYTOSIS",	"GOBP_CARBOHYDRATE_DERIVATIVE_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_MYELOID_LEUKOCYTE_DIFFERENTIATION",	"GOBP_INTERLEUKIN_12_PRODUCTION",	"GOBP_DEFENSE_RESPONSE_TO_GRAM_POSITIVE_BACTERIUM",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",	"GOBP_REGULATION_OF_MYELOID_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_INFLAMMATORY_RESPONSE",	"GOBP_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_CELLULAR_MONOVALENT_INORGANIC_CATION_HOMEOSTASIS",	"GOBP_MITOCHONDRIAL_GENE_EXPRESSION",	"GOBP_OSTEOCLAST_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_NEGATIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_ENDOTHELIAL_CELL_PROLIFERATION",	"GOBP_INTEGRIN_MEDIATED_SIGNALING_PATHWAY",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_REGULATION_OF_PHOSPHATIDYLINOSITOL_3_KINASE_SIGNALING",	"GOBP_PROTEIN_LOCALIZATION_TO_MITOCHONDRION",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_REGULATION_OF_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_CALCIUM_MEDIATED_SIGNALING",	"GOBP_PHOSPHATIDYLINOSITOL_3_KINASE_SIGNALING",	"GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_GRANULOCYTE_CHEMOTAXIS",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_REGULATION_OF_PROTEIN_DEPHOSPHORYLATION",	"GOBP_REGULATION_OF_PROTEIN_LOCALIZATION_TO_PLASMA_MEMBRANE",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_INOSITOL_LIPID_MEDIATED_SIGNALING",	"GOBP_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_RECEPTOR_MEDIATED_ENDOCYTOSIS",	"GOBP_REGULATION_OF_NIK_NF_KAPPAB_SIGNALING",	"GOBP_REGULATION_OF_ION_TRANSPORT",	"GOBP_INORGANIC_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_MYELOID_LEUKOCYTE_DIFFERENTIATION",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_REACTIVE_OXYGEN_SPECIES_METABOLIC_PROCESS",	"GOBP_CALCIUM_ION_HOMEOSTASIS",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_MONOVALENT_INORGANIC_CATION_HOMEOSTASIS",	"GOBP_RESPONSE_TO_TEMPERATURE_STIMULUS",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_CATION_TRANSMEMBRANE_TRANSPORT",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_POSITIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_REGULATION_OF_LEUKOCYTE_CHEMOTAXIS",	"GOBP_REGULATION_OF_CALCIUM_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_MYELOID_LEUKOCYTE_MIGRATION",	"GOBP_ENDOTHELIAL_CELL_MIGRATION",	"GOBP_CELLULAR_ION_HOMEOSTASIS",	"GOBP_REGULATION_OF_ENDOTHELIAL_CELL_MIGRATION",	"GOBP_REGULATION_OF_VESICLE_MEDIATED_TRANSPORT",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_PATTERN_RECOGNITION_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_LEUKOCYTE_PROLIFERATION",	"GOBP_REGULATION_OF_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_CATION_TRANSMEMBRANE_TRANSPORT",	"GOBP_VIRAL_PROCESS",	"GOBP_B_CELL_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_REGULATION_OF_PROTEIN_LOCALIZATION_TO_CELL_PERIPHERY",	"GOBP_LIPID_CATABOLIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_CELLULAR_LIPID_CATABOLIC_PROCESS",	"GOBP_MEMBRANE_LIPID_METABOLIC_PROCESS",	"GOBP_INFLAMMATORY_RESPONSE",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_RESPONSE_TO_METAL_ION",	"GOBP_CARBOHYDRATE_DERIVATIVE_BIOSYNTHETIC_PROCESS",	"GOBP_DIVALENT_INORGANIC_CATION_HOMEOSTASIS",	"GOBP_MITOCHONDRION_ORGANIZATION",	"GOBP_LEUKOCYTE_CHEMOTAXIS",	"GOBP_SECOND_MESSENGER_MEDIATED_SIGNALING",	"GOBP_ION_HOMEOSTASIS",	"GOBP_TISSUE_MIGRATION",	"GOBP_DEFENSE_RESPONSE_TO_BACTERIUM",	"GOBP_REGULATION_OF_BLOOD_CIRCULATION",	"GOBP_REGULATION_OF_TRANSPORTER_ACTIVITY",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_REGULATION_OF_METAL_ION_TRANSPORT",	"GOBP_SPHINGOLIPID_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_EXOCYTOSIS",	"GOBP_CELLULAR_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_SECRETION",	"GOBP_RESPONSE_TO_MOLECULE_OF_BACTERIAL_ORIGIN",	"GOBP_VACUOLE_ORGANIZATION",	"GOBP_STEROID_METABOLIC_PROCESS",	"GOBP_RESPONSE_TO_TUMOR_NECROSIS_FACTOR",	"GOBP_ORGANOPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_TRANSPORT",	"GOBP_HEMOSTASIS",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_EXOCYTOSIS",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_CELLULAR_METAL_ION_HOMEOSTASIS",	"GOBP_METAL_ION_HOMEOSTASIS",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_RESPONSE_TO_BACTERIUM",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_RESPONSE_TO_STEROID_HORMONE",	"GOBP_ALCOHOL_METABOLIC_PROCESS",	"GOBP_ANATOMICAL_STRUCTURE_HOMEOSTASIS",	"GOBP_MAINTENANCE_OF_LOCATION",	"GOBP_PROTEIN_TARGETING",	"GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_BODY_FLUID_LEVELS",	"GOBP_RESPONSE_TO_WOUNDING",	"GOBP_PROTEIN_POLYMERIZATION",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_CALCIUM_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_MYELOID_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_PROTEOLYSIS",	"GOBP_REGULATION_OF_LEUKOCYTE_DIFFERENTIATION",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_ORGANIC_HYDROXY_COMPOUND_METABOLIC_PROCESS",	"GOBP_CALCIUM_ION_TRANSPORT",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_INFLAMMATORY_RESPONSE",	"GOBP_CYTOKINE_MEDIATED_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE_BASED_ON_SOMATIC_RECOMBINATION_OF_IMMUNE_RECEPTORS_BUILT_FROM_IMMUNOGLOBULIN_SUPERFAMILY_DOMAINS",	"GOBP_SMALL_MOLECULE_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_HYDROLASE_ACTIVITY",	"GOBP_TAXIS",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_LOCOMOTION",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_ANATOMICAL_STRUCTURE_SIZE",	"GOBP_ACTIVATION_OF_IMMUNE_RESPONSE",	"GOBP_ENDOCYTOSIS",	"GOBP_SMALL_MOLECULE_CATABOLIC_PROCESS",	"GOBP_CELLULAR_HOMEOSTASIS",	"GOBP_LIPID_LOCALIZATION",	"GOBP_LEUKOCYTE_DIFFERENTIATION",	"GOBP_VESICLE_ORGANIZATION",	"GOBP_CELLULAR_RESPONSE_TO_ORGANIC_CYCLIC_COMPOUND",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_CELLULAR_RESPONSE_TO_LIPID",	"GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_METAL_ION_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CELL_DEATH",	"GOBP_PROCESS_UTILIZING_AUTOPHAGIC_MECHANISM",	"GOBP_T_CELL_ACTIVATION",
),
          
          "McT4vs.all_pos_GOBP" = c()
)
ggvenn(a, c("MaT4vs.all_pos_GOBP","McT4vs.all_pos_GOBP" )) 


MaT4_McT4_pos_GOBP_intersect= as.data.frame(Reduce(intersect, list(set1,set2)))
write.csv(MaT4_McT4_pos_GOBP_intersect, "_MaT4_McT4_myeloid_pos_GOBP_intersect.csv")



##data from excel##
# REACTOME
######use format -> cells -> custom -> \'@\'\, ########
MaT4 = c("REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",
         "REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",
         "REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	
         "REACTOME_LYSOSOME_VESICLE_BIOGENESIS",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_ANTIMICROBIAL_PEPTIDES",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	
         "REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_SIGNAL_REGULATORY_PROTEIN_FAMILY_INTERACTIONS",
         "REACTOME_NEGATIVE_REGULATION_OF_NMDA_RECEPTOR_MEDIATED_NEURONAL_TRANSMISSION",	"REACTOME_COLLAGEN_BIOSYNTHESIS_AND_MODIFYING_ENZYMES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_NADPH_OXIDASES",	
         "REACTOME_ADRENALINE_NORADRENALINE_INHIBITS_INSULIN_SECRETION",	"REACTOME_COLLAGEN_FORMATION",	"REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	"REACTOME_INTERFERON_SIGNALING",
         "REACTOME_GLYCOSPHINGOLIPID_METABOLISM",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_CHL1_INTERACTIONS",	"REACTOME_ROS_AND_RNS_PRODUCTION_IN_PHAGOCYTES",	"REACTOME_ENDOSOMAL_VACUOLAR_PATHWAY",	
         "REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_G_ALPHA_Z_SIGNALLING_EVENTS",	"REACTOME_DAP12_SIGNALING",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_GABA_B_RECEPTOR_ACTIVATION",	
         "REACTOME_DAP12_INTERACTIONS",	"REACTOME_SODIUM_CALCIUM_EXCHANGERS",	"REACTOME_CYTOPROTECTION_BY_HMOX1",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_PCP_CE_PATHWAY",	
         "REACTOME_TRANS_GOLGI_NETWORK_VESICLE_BUDDING",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_PLATELET_HOMEOSTASIS",	
         "REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	
         "REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_MITOCHONDRIAL_PROTEIN_IMPORT",	
         "REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	
         "REACTOME_RAC2_GTPASE_CYCLE",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_EXTRACELLULAR_MATRIX_ORGANIZATION",	"REACTOME_LEISHMANIA_INFECTION",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	
         "REACTOME_HEMOSTASIS",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES"
)
McT4 = c("REACTOME_G2_M_DNA_REPLICATION_CHECKPOINT",	"REACTOME_UNWINDING_OF_DNA")

data_1 = MaT4
data_2 = McT4

library(VennDiagram)
library(RColorBrewer)

colors()

set1 = as.list(data_1)
set2 = as.list(data_2)

venn.diagram(x=list(set1, set2), 
             category.names = c("MaT4","McT4"), filename = "venndiagram_REACTOME_pos_MaT4_McT4_myeloid.png", output=T, 
             imagetype = "png", height = 2000, width = 2000, resolution = 300, compression = "lzw",
             fill=c("skyblue", "yellowgreen"), 
             cex = 2,  lwd=2, lty='blank', fontface="bold", fontfamily="sans", cat.cex=1, cat.fontface="bold", cat.default.pos="outer", cat.pos=c(0, 20),cat.fontfamily="sans")

a <- list("MaT4vs.all_pos_REACTOME" = c("REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",
                                         "REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",
                                         "REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	
                                         "REACTOME_LYSOSOME_VESICLE_BIOGENESIS",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_ANTIMICROBIAL_PEPTIDES",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	
                                         "REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_SIGNAL_REGULATORY_PROTEIN_FAMILY_INTERACTIONS",
                                         "REACTOME_NEGATIVE_REGULATION_OF_NMDA_RECEPTOR_MEDIATED_NEURONAL_TRANSMISSION",	"REACTOME_COLLAGEN_BIOSYNTHESIS_AND_MODIFYING_ENZYMES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_NADPH_OXIDASES",	
                                         "REACTOME_ADRENALINE_NORADRENALINE_INHIBITS_INSULIN_SECRETION",	"REACTOME_COLLAGEN_FORMATION",	"REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	"REACTOME_INTERFERON_SIGNALING",
                                         "REACTOME_GLYCOSPHINGOLIPID_METABOLISM",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_CHL1_INTERACTIONS",	"REACTOME_ROS_AND_RNS_PRODUCTION_IN_PHAGOCYTES",	"REACTOME_ENDOSOMAL_VACUOLAR_PATHWAY",	
                                         "REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_G_ALPHA_Z_SIGNALLING_EVENTS",	"REACTOME_DAP12_SIGNALING",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_GABA_B_RECEPTOR_ACTIVATION",	
                                         "REACTOME_DAP12_INTERACTIONS",	"REACTOME_SODIUM_CALCIUM_EXCHANGERS",	"REACTOME_CYTOPROTECTION_BY_HMOX1",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_PCP_CE_PATHWAY",	
                                         "REACTOME_TRANS_GOLGI_NETWORK_VESICLE_BUDDING",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_PLATELET_HOMEOSTASIS",	
                                         "REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	
                                         "REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_MITOCHONDRIAL_PROTEIN_IMPORT",	
                                         "REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	
                                         "REACTOME_RAC2_GTPASE_CYCLE",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_EXTRACELLULAR_MATRIX_ORGANIZATION",	"REACTOME_LEISHMANIA_INFECTION",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	
                                         "REACTOME_HEMOSTASIS",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES"),
          
          "McT4vs.all_pos_REACTOME" = c("REACTOME_G2_M_DNA_REPLICATION_CHECKPOINT",	"REACTOME_UNWINDING_OF_DNA"))

ggvenn(a, c("MaT4vs.all_pos_REACTOME","McT4vs.all_pos_REACTOME" )) 


MaT4_McT4_pos_REACTOME_intersect= as.data.frame(Reduce(intersect, list(set1,set2)))
write.csv(MaT4_McT4_pos_REACTOME_intersect, "_MaT4_McT4_myeloid_pos_REACTOME_intersect.csv")
