==> 2_venndiagram_from_fgsea_MaT4_and_McT4_T_cell_GOBP_REACTOME.R <==
#2023.04.21
setwd('F:/Ko/67ESCC_PCN_integration_07-19-22/immune_cell_T4_M4/fgsea')

library(ggplot2)
library(RTCGA)
library(dplyr)
library(RTCGAToolbox)
library(TCGAbiolinks)
library(data.table)
#install.packages("RDocumentation")
library(RDocumentation)


##data from excel##
# GOBP
######use format -> cells -> custom -> \'@\'\, ########
MaT4 = c("GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",	"GOBP_PROTEIN_REFOLDING",	"GOBP_DE_NOVO_PROTEIN_FOLDING",	"GOBP_RESPONSE_TO_HEAT",	"GOBP_CHAPERONE_COFACTOR_DEPENDENT_PROTEIN_REFOLDING",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_ACTIVATION",	"GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_8_PRODUCTION",	"GOBP_PROTEIN_FOLDING",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ATP_BIOSYNTHETIC_PROCESS",	"GOBP_CELLULAR_RESPONSE_TO_HEAT",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_COMPLEX_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_ATP_METABOLIC_PROCESS",	"GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_INCLUSION_BODY_ASSEMBLY",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_ICOSANOID_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_VASOCONSTRICTION",	"GOBP_SKELETAL_MUSCLE_CONTRACTION",	"GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",	"GOBP_AEROBIC_RESPIRATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_RESPONSE_TO_TUMOR_CELL",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_RESPONSE_TO_TEMPERATURE_STIMULUS",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_STRUCTURE",	"GOBP_MICROTUBULE_BASED_PROTEIN_TRANSPORT",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_CHEMOTAXIS",
         "GOBP_POSITIVE_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_HOMOTYPIC_CELL_CELL_ADHESION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	"GOBP_IMMUNE_RESPONSE_TO_TUMOR_CELL",	"GOBP_PLATELET_AGGREGATION",	"GOBP_POSITIVE_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_BY_HOST_OF_VIRAL_TRANSCRIPTION",	"GOBP_POSITIVE_REGULATION_OF_TELOMERASE_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_MIGRATION",	"GOBP_REGULATION_OF_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_PYRIMIDINE_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_MONOCYTE_CHEMOTAXIS",	"GOBP_REGULATION_OF_ENDOPLASMIC_RETICULUM_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_LAMELLIPODIUM_ASSEMBLY",	"GOBP_MACROPHAGE_CHEMOTAXIS",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_ACTIN_CYTOSKELETON",	"GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_REGULATION_OF_INCLUSION_BODY_ASSEMBLY",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_SMOOTH_MUSCLE_CONTRACTION",	"GOBP_MUSCLE_CELL_CELLULAR_HOMEOSTASIS",	"GOBP_ADHESION_OF_SYMBIONT_TO_HOST",	"GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_CELLULAR_RESPIRATION",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_CELLULAR_RESPONSE_TO_VASCULAR_ENDOTHELIAL_GROWTH_FACTOR_STIMULUS",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_PODOSOME_ASSEMBLY",	"GOBP_RESPONSE_TO_SALT_STRESS",	"GOBP_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_ELECTRON_TRANSPORT_CHAIN",	"GOBP_PYROPTOSIS",	"GOBP_PYRIMIDINE_CONTAINING_COMPOUND_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_POSITIVE_REGULATION_OF_NUCLEOTIDE_BINDING_OLIGOMERIZATION_DOMAIN_CONTAINING_2_SIGNALING_PATHWAY",
         "GOBP_POSITIVE_REGULATION_OF_NUCLEOTIDE_BINDING_OLIGOMERIZATION_DOMAIN_CONTAINING_SIGNALING_PATHWAY",	"GOBP_MACROPHAGE_MIGRATION",	"GOBP_NATURAL_KILLER_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NATURAL_KILLER_CELL_DEGRANULATION",	"GOBP_REGULATION_OF_PROTEIN_POLYMERIZATION",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_MACROPHAGE_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_CELL_DIFFERENTIATION",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_RELAXATION_OF_VASCULAR_ASSOCIATED_SMOOTH_MUSCLE",	"GOBP_ENDOTHELIAL_CELL_CHEMOTAXIS",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_1_PRODUCTION",	"GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CHEMOKINE_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_ENDOPLASMIC_RETICULUM_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_REGULATION_OF_RESPONSE_TO_TUMOR_CELL",	"GOBP_POSITIVE_REGULATION_OF_MICROTUBULE_NUCLEATION",	"GOBP_PLATELET_ACTIVATION",
         "GOBP_REGULATION_OF_MITOCHONDRIAL_MEMBRANE_POTENTIAL",	"GOBP_MYELOID_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	"GOBP_SIGNAL_TRANSDUCTION_IN_ABSENCE_OF_LIGAND",	"GOBP_REGULATION_OF_LEUKOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_CHEMOTAXIS",	"GOBP_HYPOTHALAMUS_CELL_DIFFERENTIATION",	"GOBP_HYPOTHALAMUS_GONADOTROPHIN_RELEASING_HORMONE_NEURON_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_MICROGLIAL_CELL_ACTIVATION",	"GOBP_MITOCHONDRIAL_TRANSLATION",	"GOBP_REGULATION_OF_NITRIC_OXIDE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_TUBE_SIZE",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_FIBROBLAST_PROLIFERATION",	"GOBP_CENTRAL_NERVOUS_SYSTEM_NEURON_DEVELOPMENT",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_INTEGRIN_MEDIATED_SIGNALING_PATHWAY",	"GOBP_PROTEIN_LOCALIZATION_TO_MITOCHONDRION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_SYMBIONT",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",
         "GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_MORPHOGENESIS_OF_A_POLARIZED_EPITHELIUM",	"GOBP_CELL_KILLING",	"GOBP_INTERLEUKIN_8_PRODUCTION",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_CELLULAR_RESPONSE_TO_UNFOLDED_PROTEIN",	"GOBP_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY_IN_RESPONSE_TO_ENDOPLASMIC_RETICULUM_STRESS",	"GOBP_REGULATION_OF_MUSCLE_CONTRACTION",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",
         "GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_POLYMERIZATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_MITOCHONDRIAL_GENE_EXPRESSION",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_INTERLEUKIN_1_PRODUCTION",
         "GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_CELLULAR_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_HEMOSTASIS",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",	"GOBP_POSITIVE_REGULATION_OF_ION_TRANSPORT",	"GOBP_MITOCHONDRIAL_TRANSPORT",	"GOBP_RESPONSE_TO_TOXIC_SUBSTANCE",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_PATTERN_RECOGNITION_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_PROTEIN_POLYMERIZATION",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",
         "GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ANATOMICAL_STRUCTURE_SIZE",	"GOBP_NEGATIVE_REGULATION_OF_DNA_BINDING_TRANSCRIPTION_FACTOR_ACTIVITY",	"GOBP_TISSUE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_HYDROLASE_ACTIVITY",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_REGULATION_OF_MEMBRANE_POTENTIAL",	"GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",
         "GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_T_CELL_ACTIVATION",	"GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_MITOCHONDRION_ORGANIZATION",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_VIRAL_PROCESS",
         "GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_ORGANOPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_DNA_BINDING_TRANSCRIPTION_FACTOR_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_SMALL_MOLECULE_CATABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",
         "GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_I_KAPPAB_KINASE_NF_KAPPAB_SIGNALING",	"GOBP_EXOCYTOSIS",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_TAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_REGULATION_OF_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_HYDROLASE_ACTIVITY",
         "GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_TRANSPORT",	"GOBP_REGULATION_OF_VESICLE_MEDIATED_TRANSPORT",
         "GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_CARBOHYDRATE_DERIVATIVE_BIOSYNTHETIC_PROCESS",	"GOBP_ESTABLISHMENT_OF_PROTEIN_LOCALIZATION_TO_ORGANELLE",	"GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",	"GOBP_SMALL_GTPASE_MEDIATED_SIGNAL_TRANSDUCTION",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE")
McT4 = c("GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_POLARIZED_EPITHELIAL_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_MORPHOGENESIS_OF_A_POLARIZED_EPITHELIUM",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_VIRAL_GENOME_REPLICATION",	"GOBP_ESTABLISHMENT_OF_EPITHELIAL_CELL_APICAL_BASAL_POLARITY",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_ESTABLISHMENT_OR_MAINTENANCE_OF_MONOPOLAR_CELL_POLARITY",	"GOBP_RESPONSE_TO_TYPE_I_INTERFERON",	"GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_DETECTION_OF_BIOTIC_STIMULUS",	"GOBP_CELL_KILLING",	"GOBP_CYTOLYSIS",	"GOBP_DETECTION_OF_OTHER_ORGANISM",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_VIRAL_PROCESS",	"GOBP_NATURAL_KILLER_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_DETECTION_OF_EXTERNAL_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_CELL_KILLING",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",
         "GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_NEGATIVE_REGULATION_OF_T_CELL_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_TOLL_LIKE_RECEPTOR_4_SIGNALING_PATHWAY",
         "GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_MIGRATION",	"GOBP_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_GLAND_MORPHOGENESIS",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_TOLERANCE_INDUCTION",	"GOBP_INTERFERON_BETA_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_BETA_PRODUCTION",	"GOBP_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_INTERFERON_GAMMA_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_SEMAPHORIN_PLEXIN_SIGNALING_PATHWAY",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_NEGATIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_CAMP_MEDIATED_SIGNALING",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNE_RESPONSE_TO_TUMOR_CELL",
         "GOBP_EXOCYTOSIS",	"GOBP_ERK1_AND_ERK2_CASCADE",	"GOBP_ADENYLATE_CYCLASE_MODULATING_G_PROTEIN_COUPLED_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",	"GOBP_CELLULAR_RESPONSE_TO_VIRUS",	"GOBP_MAST_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_NEGATIVE_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_REGULATION_OF_CELL_KILLING",	"GOBP_CELLULAR_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_RESPONSE_TO_FIBROBLAST_GROWTH_FACTOR",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_EOSINOPHIL_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_ACTIN_FILAMENT_BUNDLE_ASSEMBLY",	"GOBP_CELL_CELL_RECOGNITION",
         "GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_TISSUE_MIGRATION",	"GOBP_REGULATION_OF_SYNAPSE_STRUCTURE_OR_ACTIVITY",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_TOLL_LIKE_RECEPTOR_7_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_RESPONSE_TO_CYTOKINE_STIMULUS",	"GOBP_STRESS_FIBER_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_CLEAVAGE_FURROW_FORMATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_CADHERIN",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_INTERLEUKIN_6_PRODUCTION",	"GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_ACTIN_FILAMENT_BUNDLE_ORGANIZATION",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION_VIA_T_CELL_RECEPTOR_CONTACT_WITH_ANTIGEN_BOUND_TO_MHC_MOLECULE_ON_ANTIGEN_PRESENTING_CELL",
         "GOBP_REGULATION_OF_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_T_CELL_ACTIVATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_RESPONSE_TO_MOLECULE_OF_BACTERIAL_ORIGIN",	"GOBP_LYMPHOCYTE_MIGRATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_T_CELL_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_CELL_SUBSTRATE_ADHESION",	"GOBP_ACTOMYOSIN_STRUCTURE_ORGANIZATION",	"GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_CYTOKINE_MEDIATED_SIGNALING_PATHWAY",
         "GOBP_PROTEIN_DEPHOSPHORYLATION",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",
         "GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_REGULATION_OF_CELL_MORPHOGENESIS",	"GOBP_DEPHOSPHORYLATION",	"GOBP_REGULATION_OF_LEUKOCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_HEMOPOIESIS",
         "GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_REGULATION_OF_ION_TRANSPORT",	"GOBP_TAXIS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_POPULATION_PROLIFERATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE_BASED_ON_SOMATIC_RECOMBINATION_OF_IMMUNE_RECEPTORS_BUILT_FROM_IMMUNOGLOBULIN_SUPERFAMILY_DOMAINS",	"GOBP_G_PROTEIN_COUPLED_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_MORPHOGENESIS_OF_AN_EPITHELIUM",	"GOBP_ACTIVATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_CELLULAR_COMPONENT_BIOGENESIS",	"GOBP_REGULATION_OF_SECRETION",	"GOBP_TRANSMEMBRANE_RECEPTOR_PROTEIN_TYROSINE_KINASE_SIGNALING_PATHWAY",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_TUBE_MORPHOGENESIS"
)

data_1 = MaT4
data_2 = McT4

library(VennDiagram)
library(RColorBrewer)
library(ggvenn)

colors()

set1 = as.list(data_1)
set2 = as.list(data_2)

venn.diagram(x=list(set1, set2), 
             category.names = c("MaT4","McT4"), filename = "venndiagram_GOBP_MaT4_McT4.png", output=T, 
             imagetype = "png", height = 2000, width = 2000, resolution = 300, compression = "lzw",
             fill=c("skyblue", "yellowgreen"), 
             cex = 2,  lwd=2, lty='blank', fontface="bold", fontfamily="sans", cat.cex=1, cat.fontface="bold", cat.default.pos="outer", cat.pos=c(0, 0),cat.fontfamily="sans")


a <- list("MaT4vs.all_pos_GOBP" = c("GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",	"GOBP_PROTEIN_REFOLDING",	"GOBP_DE_NOVO_PROTEIN_FOLDING",	"GOBP_RESPONSE_TO_HEAT",	"GOBP_CHAPERONE_COFACTOR_DEPENDENT_PROTEIN_REFOLDING",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_ACTIVATION",	"GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_8_PRODUCTION",	"GOBP_PROTEIN_FOLDING",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ATP_BIOSYNTHETIC_PROCESS",	"GOBP_CELLULAR_RESPONSE_TO_HEAT",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_COMPLEX_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_ATP_METABOLIC_PROCESS",	"GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_INCLUSION_BODY_ASSEMBLY",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_ICOSANOID_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_VASOCONSTRICTION",	"GOBP_SKELETAL_MUSCLE_CONTRACTION",	"GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",	"GOBP_AEROBIC_RESPIRATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_RESPONSE_TO_TUMOR_CELL",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_RESPONSE_TO_TEMPERATURE_STIMULUS",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_STRUCTURE",	"GOBP_MICROTUBULE_BASED_PROTEIN_TRANSPORT",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_CHEMOTAXIS",
                                    "GOBP_POSITIVE_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_HOMOTYPIC_CELL_CELL_ADHESION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	"GOBP_IMMUNE_RESPONSE_TO_TUMOR_CELL",	"GOBP_PLATELET_AGGREGATION",	"GOBP_POSITIVE_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_BY_HOST_OF_VIRAL_TRANSCRIPTION",	"GOBP_POSITIVE_REGULATION_OF_TELOMERASE_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_MIGRATION",	"GOBP_REGULATION_OF_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_PYRIMIDINE_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_MONOCYTE_CHEMOTAXIS",	"GOBP_REGULATION_OF_ENDOPLASMIC_RETICULUM_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_LAMELLIPODIUM_ASSEMBLY",	"GOBP_MACROPHAGE_CHEMOTAXIS",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_ACTIN_CYTOSKELETON",	"GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_REGULATION_OF_INCLUSION_BODY_ASSEMBLY",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_SMOOTH_MUSCLE_CONTRACTION",	"GOBP_MUSCLE_CELL_CELLULAR_HOMEOSTASIS",	"GOBP_ADHESION_OF_SYMBIONT_TO_HOST",	"GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_CELLULAR_RESPIRATION",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_CELLULAR_RESPONSE_TO_VASCULAR_ENDOTHELIAL_GROWTH_FACTOR_STIMULUS",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_PODOSOME_ASSEMBLY",	"GOBP_RESPONSE_TO_SALT_STRESS",	"GOBP_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_ELECTRON_TRANSPORT_CHAIN",	"GOBP_PYROPTOSIS",	"GOBP_PYRIMIDINE_CONTAINING_COMPOUND_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_POSITIVE_REGULATION_OF_NUCLEOTIDE_BINDING_OLIGOMERIZATION_DOMAIN_CONTAINING_2_SIGNALING_PATHWAY",
                                    "GOBP_POSITIVE_REGULATION_OF_NUCLEOTIDE_BINDING_OLIGOMERIZATION_DOMAIN_CONTAINING_SIGNALING_PATHWAY",	"GOBP_MACROPHAGE_MIGRATION",	"GOBP_NATURAL_KILLER_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NATURAL_KILLER_CELL_DEGRANULATION",	"GOBP_REGULATION_OF_PROTEIN_POLYMERIZATION",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_MACROPHAGE_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_CELL_DIFFERENTIATION",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_RELAXATION_OF_VASCULAR_ASSOCIATED_SMOOTH_MUSCLE",	"GOBP_ENDOTHELIAL_CELL_CHEMOTAXIS",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_1_PRODUCTION",	"GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CHEMOKINE_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_ENDOPLASMIC_RETICULUM_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_REGULATION_OF_RESPONSE_TO_TUMOR_CELL",	"GOBP_POSITIVE_REGULATION_OF_MICROTUBULE_NUCLEATION",	"GOBP_PLATELET_ACTIVATION",
                                    "GOBP_REGULATION_OF_MITOCHONDRIAL_MEMBRANE_POTENTIAL",	"GOBP_MYELOID_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	"GOBP_SIGNAL_TRANSDUCTION_IN_ABSENCE_OF_LIGAND",	"GOBP_REGULATION_OF_LEUKOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_CHEMOTAXIS",	"GOBP_HYPOTHALAMUS_CELL_DIFFERENTIATION",	"GOBP_HYPOTHALAMUS_GONADOTROPHIN_RELEASING_HORMONE_NEURON_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_MICROGLIAL_CELL_ACTIVATION",	"GOBP_MITOCHONDRIAL_TRANSLATION",	"GOBP_REGULATION_OF_NITRIC_OXIDE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_TUBE_SIZE",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_FIBROBLAST_PROLIFERATION",	"GOBP_CENTRAL_NERVOUS_SYSTEM_NEURON_DEVELOPMENT",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_INTEGRIN_MEDIATED_SIGNALING_PATHWAY",	"GOBP_PROTEIN_LOCALIZATION_TO_MITOCHONDRION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_SYMBIONT",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",
                                    "GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_MORPHOGENESIS_OF_A_POLARIZED_EPITHELIUM",	"GOBP_CELL_KILLING",	"GOBP_INTERLEUKIN_8_PRODUCTION",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_CELLULAR_RESPONSE_TO_UNFOLDED_PROTEIN",	"GOBP_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY_IN_RESPONSE_TO_ENDOPLASMIC_RETICULUM_STRESS",	"GOBP_REGULATION_OF_MUSCLE_CONTRACTION",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",
                                    "GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_POLYMERIZATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_MITOCHONDRIAL_GENE_EXPRESSION",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_INTERLEUKIN_1_PRODUCTION",
                                    "GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_CELLULAR_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_HEMOSTASIS",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",	"GOBP_POSITIVE_REGULATION_OF_ION_TRANSPORT",	"GOBP_MITOCHONDRIAL_TRANSPORT",	"GOBP_RESPONSE_TO_TOXIC_SUBSTANCE",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_PATTERN_RECOGNITION_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_PROTEIN_POLYMERIZATION",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",
                                    "GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ANATOMICAL_STRUCTURE_SIZE",	"GOBP_NEGATIVE_REGULATION_OF_DNA_BINDING_TRANSCRIPTION_FACTOR_ACTIVITY",	"GOBP_TISSUE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_HYDROLASE_ACTIVITY",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_REGULATION_OF_MEMBRANE_POTENTIAL",	"GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",
                                    "GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_T_CELL_ACTIVATION",	"GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_MITOCHONDRION_ORGANIZATION",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_VIRAL_PROCESS",
                                    "GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_ORGANOPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_DNA_BINDING_TRANSCRIPTION_FACTOR_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_SMALL_MOLECULE_CATABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",
                                    "GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_I_KAPPAB_KINASE_NF_KAPPAB_SIGNALING",	"GOBP_EXOCYTOSIS",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_TAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_REGULATION_OF_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_HYDROLASE_ACTIVITY",
                                    "GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_TRANSPORT",	"GOBP_REGULATION_OF_VESICLE_MEDIATED_TRANSPORT",
                                    "GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_CARBOHYDRATE_DERIVATIVE_BIOSYNTHETIC_PROCESS",	"GOBP_ESTABLISHMENT_OF_PROTEIN_LOCALIZATION_TO_ORGANELLE",	"GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",	"GOBP_SMALL_GTPASE_MEDIATED_SIGNAL_TRANSDUCTION",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE"),
          
          "McT4vs.all_pos_GOBP" = c("GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_POLARIZED_EPITHELIAL_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_MORPHOGENESIS_OF_A_POLARIZED_EPITHELIUM",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_VIRAL_GENOME_REPLICATION",	"GOBP_ESTABLISHMENT_OF_EPITHELIAL_CELL_APICAL_BASAL_POLARITY",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_ESTABLISHMENT_OR_MAINTENANCE_OF_MONOPOLAR_CELL_POLARITY",	"GOBP_RESPONSE_TO_TYPE_I_INTERFERON",	"GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_DETECTION_OF_BIOTIC_STIMULUS",	"GOBP_CELL_KILLING",	"GOBP_CYTOLYSIS",	"GOBP_DETECTION_OF_OTHER_ORGANISM",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_VIRAL_PROCESS",	"GOBP_NATURAL_KILLER_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_DETECTION_OF_EXTERNAL_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_CELL_KILLING",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",
                                    "GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_NEGATIVE_REGULATION_OF_T_CELL_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_TOLL_LIKE_RECEPTOR_4_SIGNALING_PATHWAY",
                                    "GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_MIGRATION",	"GOBP_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_GLAND_MORPHOGENESIS",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_TOLERANCE_INDUCTION",	"GOBP_INTERFERON_BETA_PRODUCTION",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_BETA_PRODUCTION",	"GOBP_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_INTERFERON_GAMMA_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_SEMAPHORIN_PLEXIN_SIGNALING_PATHWAY",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_NEGATIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_CAMP_MEDIATED_SIGNALING",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNE_RESPONSE_TO_TUMOR_CELL",
                                    "GOBP_EXOCYTOSIS",	"GOBP_ERK1_AND_ERK2_CASCADE",	"GOBP_ADENYLATE_CYCLASE_MODULATING_G_PROTEIN_COUPLED_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",	"GOBP_CELLULAR_RESPONSE_TO_VIRUS",	"GOBP_MAST_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_NEGATIVE_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_REGULATION_OF_CELL_KILLING",	"GOBP_CELLULAR_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_RESPONSE_TO_FIBROBLAST_GROWTH_FACTOR",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_EOSINOPHIL_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_ACTIN_FILAMENT_BUNDLE_ASSEMBLY",	"GOBP_CELL_CELL_RECOGNITION",
                                    "GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_TISSUE_MIGRATION",	"GOBP_REGULATION_OF_SYNAPSE_STRUCTURE_OR_ACTIVITY",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_TOLL_LIKE_RECEPTOR_7_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_RESPONSE_TO_CYTOKINE_STIMULUS",	"GOBP_STRESS_FIBER_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_CLEAVAGE_FURROW_FORMATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_CADHERIN",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_INTERLEUKIN_6_PRODUCTION",	"GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_ACTIN_FILAMENT_BUNDLE_ORGANIZATION",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION_VIA_T_CELL_RECEPTOR_CONTACT_WITH_ANTIGEN_BOUND_TO_MHC_MOLECULE_ON_ANTIGEN_PRESENTING_CELL",
                                    "GOBP_REGULATION_OF_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_T_CELL_ACTIVATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_RESPONSE_TO_MOLECULE_OF_BACTERIAL_ORIGIN",	"GOBP_LYMPHOCYTE_MIGRATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_T_CELL_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_CELL_SUBSTRATE_ADHESION",	"GOBP_ACTOMYOSIN_STRUCTURE_ORGANIZATION",	"GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_CYTOKINE_MEDIATED_SIGNALING_PATHWAY",
                                    "GOBP_PROTEIN_DEPHOSPHORYLATION",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",
                                    "GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_REGULATION_OF_CELL_MORPHOGENESIS",	"GOBP_DEPHOSPHORYLATION",	"GOBP_REGULATION_OF_LEUKOCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_HEMOPOIESIS",
                                    "GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_REGULATION_OF_ION_TRANSPORT",	"GOBP_TAXIS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_POPULATION_PROLIFERATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE_BASED_ON_SOMATIC_RECOMBINATION_OF_IMMUNE_RECEPTORS_BUILT_FROM_IMMUNOGLOBULIN_SUPERFAMILY_DOMAINS",	"GOBP_G_PROTEIN_COUPLED_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_MORPHOGENESIS_OF_AN_EPITHELIUM",	"GOBP_ACTIVATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_CELLULAR_COMPONENT_BIOGENESIS",	"GOBP_REGULATION_OF_SECRETION",	"GOBP_TRANSMEMBRANE_RECEPTOR_PROTEIN_TYROSINE_KINASE_SIGNALING_PATHWAY",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_TUBE_MORPHOGENESIS")
)
ggvenn(a, c("MaT4vs.all_pos_GOBP","McT4vs.all_pos_GOBP" )) 


MaT4_McT4_pos_GOBP_intersect= as.data.frame(Reduce(intersect, list(set1,set2)))
write.csv(MaT4_McT4_pos_GOBP_intersect, "_MaT4_McT4_Tcell_pos_GOBP_intersect.csv")



##data from excel##
# REACTOME
######use format -> cells -> custom -> \'@\'\, ########
MaT4 = c("REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_PD_1_SIGNALING",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_PYROPTOSIS",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_NUCLEOTIDE_CATABOLISM",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TNFS_BIND_THEIR_PHYSIOLOGICAL_RECEPTORS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_ADHERENS_JUNCTIONS_INTERACTIONS",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_TCR_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_UCH_PROTEINASES",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	"REACTOME_APOPTOSIS",	"REACTOME_PTEN_REGULATION",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_NEDDYLATION",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_CELL_CYCLE_MITOTIC"
)
McT4 = c("REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_PD_1_SIGNALING",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TCR_SIGNALING",	"REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",	"REACTOME_NOTCH2_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_INACTIVATION_OF_CDC42_AND_RAC1",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_APOPTOSIS",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_HEMOSTASIS",	"REACTOME_SIGNALING_BY_INTERLEUKINS"
)

data_1 = MaT4
data_2 = McT4

library(VennDiagram)
library(RColorBrewer)

colors()

set1 = as.list(data_1)
set2 = as.list(data_2)

venn.diagram(x=list(set1, set2), 
             category.names = c("MaT4","McT4"), filename = "venndiagram_REACTOME_pos_MaT4_McT4.png", output=T, 
             imagetype = "png", height = 2000, width = 2000, resolution = 300, compression = "lzw",
             fill=c("skyblue", "yellowgreen"), 
             cex = 2,  lwd=2, lty='blank', fontface="bold", fontfamily="sans", cat.cex=1, cat.fontface="bold", cat.default.pos="outer", cat.pos=c(0, 20),cat.fontfamily="sans")

a <- list("MaT4vs.all_pos_REACTOME" = c("REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_PD_1_SIGNALING",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_PYROPTOSIS",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_NUCLEOTIDE_CATABOLISM",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TNFS_BIND_THEIR_PHYSIOLOGICAL_RECEPTORS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_ADHERENS_JUNCTIONS_INTERACTIONS",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_TCR_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_UCH_PROTEINASES",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	"REACTOME_APOPTOSIS",	"REACTOME_PTEN_REGULATION",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_NEDDYLATION",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_CELL_CYCLE_MITOTIC"
),
          
          "McT4vs.all_pos_REACTOME" = c("REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_PD_1_SIGNALING",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TCR_SIGNALING",	"REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",	"REACTOME_NOTCH2_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_INACTIVATION_OF_CDC42_AND_RAC1",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_APOPTOSIS",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_HEMOSTASIS",	"REACTOME_SIGNALING_BY_INTERLEUKINS"
          ))
ggvenn(a, c("MaT4vs.all_pos_REACTOME","McT4vs.all_pos_REACTOME" )) 


MaT4_McT4_pos_REACTOME_intersect= as.data.frame(Reduce(intersect, list(set1,set2)))
write.csv(MaT4_McT4_pos_REACTOME_intersect, "_MaT4_McT4_Tcell_pos_REACTOME_intersect.csv")
