==> 2_venndiagram_from_fgsea_quad.venn_MaT4_and_MbT3.R <==
rm(list=ls())
#install.packages("RDocumentation")
#install.packages("VennDiagram")
#install.packages('ggvenn')

#library(VennDiagram)
library(RColorBrewer)
library(ggplot2)
library(RTCGA)
library(dplyr)
library(RTCGAToolbox)
library(TCGAbiolinks)
library(data.table)
library(RDocumentation)
library(ggvenn)

setwd("/Users/KKo1/OneDrive - mdanderson.org/13. KyungPil Ko/Experiment_results/scRNA-seq/2022_Front_Immn_PD-1_responder/PD-1_therapy_patients/fgsea")


#fgsea: REACTOME results comparison
# F_T_patient_cluster
# T cell
# (F1T4 vs. all) and (Responder vs. Nonreponder)


M1T4vs.all_pos_reactome = c("REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_PD_1_SIGNALING",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_PYROPTOSIS",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_NUCLEOTIDE_CATABOLISM",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TNFS_BIND_THEIR_PHYSIOLOGICAL_RECEPTORS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_ADHERENS_JUNCTIONS_INTERACTIONS",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_TCR_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_UCH_PROTEINASES",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	"REACTOME_APOPTOSIS",	"REACTOME_PTEN_REGULATION",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_NEDDYLATION",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_CELL_CYCLE_MITOTIC"
)
M1T4vs.all_neg_reactome = c("REACTOME_DCC_MEDIATED_ATTRACTIVE_SIGNALING",	"REACTOME_SCAVENGING_OF_HEME_FROM_PLASMA",	"REACTOME_COMPLEMENT_CASCADE",	"REACTOME_CREATION_OF_C4_AND_C2_ACTIVATORS",	"REACTOME_INITIAL_TRIGGERING_OF_COMPLEMENT",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_ACTIVATION_OF_THE_MRNA_UPON_BINDING_OF_THE_CAP_BINDING_COMPLEX_AND_EIFS_AND_SUBSEQUENT_BINDING_TO_43S",	"REACTOME_CELLULAR_RESPONSE_TO_STARVATION",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_RESPONSE_OF_EIF2AK4_GCN2_TO_AMINO_ACID_DEFICIENCY",	"REACTOME_NONSENSE_MEDIATED_DECAY_NMD",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_SELENOAMINO_ACID_METABOLISM",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION"
)

M2T3_pos_reactome =c("REACTOME_CREATION_OF_C4_AND_C2_ACTIVATORS",	"REACTOME_SCAVENGING_OF_HEME_FROM_PLASMA",	"REACTOME_INITIAL_TRIGGERING_OF_COMPLEMENT",	"REACTOME_COMPLEMENT_CASCADE",	"REACTOME_CROSSLINKING_OF_COLLAGEN_FIBRILS"
)
M2T3_neg_reactome = c("REACTOME_CELL_CYCLE_MITOTIC",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES",	"REACTOME_ASPARAGINE_N_LINKED_GLYCOSYLATION",
                      "REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_M_PHASE",	"REACTOME_NEDDYLATION",	"REACTOME_SIGNALING_BY_GPCR",
                      "REACTOME_ANTIGEN_PROCESSING_UBIQUITINATION_PROTEASOME_DEGRADATION",	"REACTOME_SIGNALING_BY_NUCLEAR_RECEPTORS",	"REACTOME_SUMOYLATION",
                      "REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	"REACTOME_NEURONAL_SYSTEM",	"REACTOME_POTENTIAL_THERAPEUTICS_FOR_SARS",
                      "REACTOME_PROCESSING_OF_CAPPED_INTRON_CONTAINING_PRE_MRNA",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",
                      "REACTOME_NEUROTRANSMITTER_RECEPTORS_AND_POSTSYNAPTIC_SIGNAL_TRANSMISSION",	"REACTOME_TRANSPORT_TO_THE_GOLGI_AND_SUBSEQUENT_MODIFICATION",
                      "REACTOME_ER_TO_GOLGI_ANTEROGRADE_TRANSPORT",	"REACTOME_SARS_COV_2_ACTIVATES_MODULATES_INNATE_AND_ADAPTIVE_IMMUNE_RESPONSES",
                      "REACTOME_TRANSMISSION_ACROSS_CHEMICAL_SYNAPSES",	"REACTOME_MRNA_SPLICING",	"REACTOME_AUTOPHAGY",	"REACTOME_INTRACELLULAR_SIGNALING_BY_SECOND_MESSENGERS",
                      "REACTOME_MITOTIC_METAPHASE_AND_ANAPHASE",	"REACTOME_CELL_SURFACE_INTERACTIONS_AT_THE_VASCULAR_WALL",	"REACTOME_HIV_INFECTION",
                      "REACTOME_PI3K_AKT_SIGNALING_IN_CANCER",	"REACTOME_COPI_DEPENDENT_GOLGI_TO_ER_RETROGRADE_TRAFFIC",	"REACTOME_ESR_MEDIATED_SIGNALING",	"REACTOME_PTEN_REGULATION",
                      "REACTOME_SIGNALING_BY_NOTCH",	"REACTOME_SEPARATION_OF_SISTER_CHROMATIDS",	"REACTOME_MITOCHONDRIAL_BIOGENESIS",	"REACTOME_MYD88_INDEPENDENT_TLR4_CASCADE",
                      "REACTOME_RHO_GTPASES_ACTIVATE_FORMINS",	"REACTOME_SIGNALING_BY_RECEPTOR_TYROSINE_KINASES",	"REACTOME_CDC42_GTPASE_CYCLE",
                      "REACTOME_CELL_DEATH_SIGNALLING_VIA_NRAGE_NRIF_AND_NADE",	"REACTOME_SIGNALING_BY_NTRKS",	"REACTOME_TRANS_GOLGI_NETWORK_VESICLE_BUDDING",
                      "REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",	"REACTOME_RHOB_GTPASE_CYCLE",	"REACTOME_MUSCLE_CONTRACTION",	"REACTOME_CIRCADIAN_CLOCK",
                      "REACTOME_RAB_GERANYLGERANYLATION",	"REACTOME_TOLL_LIKE_RECEPTOR_CASCADES",	"REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",
                      "REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_RAC2_GTPASE_CYCLE",
                      "REACTOME_REGULATION_OF_INSULIN_SECRETION",	"REACTOME_P75_NTR_RECEPTOR_MEDIATED_SIGNALLING",	"REACTOME_RHOA_GTPASE_CYCLE",	"REACTOME_HEMOSTASIS",
                      "REACTOME_INTRINSIC_PATHWAY_FOR_APOPTOSIS",	"REACTOME_GOLGI_ASSOCIATED_VESICLE_BIOGENESIS",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",
                      "REACTOME_ESTROGEN_DEPENDENT_GENE_EXPRESSION",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_DEATH_RECEPTOR_SIGNALLING",	"REACTOME_RHOG_GTPASE_CYCLE",
                      "REACTOME_SIGNALING_BY_MET",	"REACTOME_DEUBIQUITINATION",	"REACTOME_RHOF_GTPASE_CYCLE",	"REACTOME_RAC1_GTPASE_CYCLE",	"REACTOME_OPIOID_SIGNALLING",
                      "REACTOME_COPI_MEDIATED_ANTEROGRADE_TRANSPORT",	"REACTOME_RHOC_GTPASE_CYCLE",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",
                      "REACTOME_SENSORY_PERCEPTION",	"REACTOME_APOPTOTIC_CLEAVAGE_OF_CELLULAR_PROTEINS",	"REACTOME_SIGNALING_BY_VEGF",	"REACTOME_CLATHRIN_MEDIATED_ENDOCYTOSIS",
                      "REACTOME_UB_SPECIFIC_PROCESSING_PROTEASES",	"REACTOME_SIGNALING_BY_HEDGEHOG",	"REACTOME_COPI_INDEPENDENT_GOLGI_TO_ER_RETROGRADE_TRAFFIC",
                      "REACTOME_DISEASES_OF_SIGNAL_TRANSDUCTION_BY_GROWTH_FACTOR_RECEPTORS_AND_SECOND_MESSENGERS",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",
                      "REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",
                      "REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	"REACTOME_APOPTOSIS",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",
                      "REACTOME_RAC3_GTPASE_CYCLE",	"REACTOME_G_ALPHA_12_13_SIGNALLING_EVENTS",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",
                      "REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX1",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_BBSOME_MEDIATED_CARGO_TARGETING_TO_CILIUM",	"REACTOME_CRISTAE_FORMATION",
                      "REACTOME_MET_PROMOTES_CELL_MOTILITY",	"REACTOME_NETRIN_1_SIGNALING",	"REACTOME_G_BETA_GAMMA_SIGNALLING_THROUGH_PI3KGAMMA",	"REACTOME_KEAP1_NFE2L2_PATHWAY",
                      "REACTOME_POTASSIUM_CHANNELS",	"REACTOME_PLATELET_HOMEOSTASIS",	"REACTOME_SIGNALING_BY_SCF_KIT",	"REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",
                      "REACTOME_UPTAKE_AND_ACTIONS_OF_BACTERIAL_TOXINS",	"REACTOME_UCH_PROTEINASES",	"REACTOME_SIGNALING_BY_KIT_IN_DISEASE",	"REACTOME_PARASITE_INFECTION",
                      "REACTOME_REGULATION_OF_KIT_SIGNALING",	"REACTOME_ASSEMBLY_OF_THE_HIV_VIRION",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",
                      "REACTOME_COSTIMULATION_BY_THE_CD28_FAMILY",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_EPH_EPHRIN_MEDIATED_REPULSION_OF_CELLS",
                      "REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",
                      "REACTOME_DDX58_IFIH1_MEDIATED_INDUCTION_OF_INTERFERON_ALPHA_BETA",
                      "REACTOME_RHO_GTPASE_EFFECTORS",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_ONCOGENIC_MAPK_SIGNALING",
                      "REACTOME_L1CAM_INTERACTIONS",	"REACTOME_VASOPRESSIN_REGULATES_RENAL_WATER_HOMEOSTASIS_VIA_AQUAPORINS",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",
                      "REACTOME_LEISHMANIA_INFECTION",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",
                      "REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_P75NTR_SIGNALS_VIA_NF_KB",	"REACTOME_AQUAPORIN_MEDIATED_TRANSPORT",	"REACTOME_SIGNALING_BY_EGFR_IN_CANCER",
                      "REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_NEUTROPHIL_DEGRANULATION",
                      "REACTOME_CONSTITUTIVE_SIGNALING_BY_ABERRANT_PI3K_IN_CANCER",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",
                      "REACTOME_NEGATIVE_REGULATION_OF_FGFR1_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",
                      "REACTOME_TGF_BETA_RECEPTOR_SIGNALING_IN_EMT_EPITHELIAL_TO_MESENCHYMAL_TRANSITION",	"REACTOME_SIGNALING_BY_BRAF_AND_RAF1_FUSIONS",
                      "REACTOME_ACTIVATION_OF_BAD_AND_TRANSLOCATION_TO_MITOCHONDRIA",	"REACTOME_CELL_JUNCTION_ORGANIZATION",	"REACTOME_INTERLEUKIN_15_SIGNALING",
                      "REACTOME_HEDGEHOG_ON_STATE",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",
                      "REACTOME_NEGATIVE_REGULATION_OF_MAPK_PATHWAY",	"REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",
                      "REACTOME_GLUCAGON_LIKE_PEPTIDE_1_GLP1_REGULATES_INSULIN_SECRETION",	"REACTOME_FOXO_MEDIATED_TRANSCRIPTION_OF_CELL_CYCLE_GENES",
                      "REACTOME_INTERLEUKIN_3_INTERLEUKIN_5_AND_GM_CSF_SIGNALING",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",
                      "REACTOME_ASSEMBLY_AND_CELL_SURFACE_PRESENTATION_OF_NMDA_RECEPTORS",	"REACTOME_GLUCAGON_SIGNALING_IN_METABOLIC_REGULATION",	"REACTOME_ABC_TRANSPORTER_DISORDERS",
                      "REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_RIPK1_MEDIATED_REGULATED_NECROSIS",	"REACTOME_EPHRIN_SIGNALING",	"REACTOME_CLASS_B_2_SECRETIN_FAMILY_RECEPTORS",
                      "REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_INLB_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELL",
                      "REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",	"REACTOME_NEGATIVE_REGULATORS_OF_DDX58_IFIH1_SIGNALING",	"REACTOME_SPRY_REGULATION_OF_FGF_SIGNALING",
                      "REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	"REACTOME_SUMOYLATION_OF_INTRACELLULAR_RECEPTORS",
                      "REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_C_TYPE_LECTIN_RECEPTORS_CLRS",	"REACTOME_CHK1_CHK2_CDS1_MEDIATED_INACTIVATION_OF_CYCLIN_B_CDK1_COMPLEX",
                      "REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_PLATELET_AGGREGATION_PLUG_FORMATION",	"REACTOME_MAPK6_MAPK4_SIGNALING",
                      "REACTOME_FGFR1_MUTANT_RECEPTOR_ACTIVATION",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",
                      "REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",
                      "REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_REGULATED_NECROSIS",	"REACTOME_INTEGRIN_SIGNALING",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",
                      "REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",	"REACTOME_LISTERIA_MONOCYTOGENES_ENTRY_INTO_HOST_CELLS",
                      "REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RAP1_SIGNALLING",	"REACTOME_INTERLEUKIN_4_AND_INTERLEUKIN_13_SIGNALING",	"REACTOME_RECYCLING_PATHWAY_OF_L1",
                      "REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_ARYL_HYDROCARBON_RECEPTOR_SIGNALLING",
                      "REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_ACTIVATION_OF_AMPK_DOWNSTREAM_OF_NMDARS",	"REACTOME_CELL_CELL_COMMUNICATION",
                      "REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",
                      "REACTOME_NUCLEOTIDE_BINDING_DOMAIN_LEUCINE_RICH_REPEAT_CONTAINING_RECEPTOR_NLR_SIGNALING_PATHWAYS",	"REACTOME_STABILIZATION_OF_P53",
                      "REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	"REACTOME_DEGRADATION_OF_DVL",
                      "REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_REGULATION_OF_FZD_BY_UBIQUITINATION",
                      "REACTOME_PCP_CE_PATHWAY",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_UPTAKE_AND_FUNCTION_OF_DIPHTHERIA_TOXIN",	"REACTOME_RHOH_GTPASE_CYCLE",
                      "REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_SIGNALING_BY_CYTOSOLIC_FGFR1_FUSION_MUTANTS",
                      "REACTOME_SEMA3A_PLEXIN_REPULSION_SIGNALING_BY_INHIBITING_INTEGRIN_ADHESION",	"REACTOME_RESPONSE_OF_MTB_TO_PHAGOCYTOSIS",	"REACTOME_DAP12_SIGNALING",
                      "REACTOME_GRB2_SOS_PROVIDES_LINKAGE_TO_MAPK_SIGNALING_FOR_INTEGRINS",
                      "REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_FLT3_SIGNALING",	"REACTOME_ENOS_ACTIVATION",
                      "REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",
                      "REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	"REACTOME_ERBB2_REGULATES_CELL_MOTILITY",	"REACTOME_INTERLEUKIN_2_SIGNALING",
                      "REACTOME_INTERLEUKIN_1_PROCESSING",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",
                      "REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",	"REACTOME_DARPP_32_EVENTS",	"REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",
                      "REACTOME_CD28_DEPENDENT_VAV1_PATHWAY",	"REACTOME_TP53_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_G1_CELL_CYCLE_ARREST",
                      "REACTOME_POST_CHAPERONIN_TUBULIN_FOLDING_PATHWAY",	"REACTOME_CELL_CELL_JUNCTION_ORGANIZATION",	"REACTOME_GLUTAMATE_NEUROTRANSMITTER_RELEASE_CYCLE",
                      "REACTOME_NF_KB_IS_ACTIVATED_AND_SIGNALS_SURVIVAL",	"REACTOME_P130CAS_LINKAGE_TO_MAPK_SIGNALING_FOR_INTEGRINS",
                      "REACTOME_THE_ROLE_OF_NEF_IN_HIV_1_REPLICATION_AND_DISEASE_PATHOGENESIS",	"REACTOME_MAP2K_AND_MAPK_ACTIVATION",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",
                      "REACTOME_SEMA4D_MEDIATED_INHIBITION_OF_CELL_ATTACHMENT_AND_MIGRATION",	"REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_PD_1_SIGNALING",	"REACTOME_TCR_SIGNALING",
                      "REACTOME_INTERLEUKIN_2_FAMILY_SIGNALING",	"REACTOME_GABA_SYNTHESIS_RELEASE_REUPTAKE_AND_DEGRADATION",	"REACTOME_MET_ACTIVATES_RAP1_AND_RAC1",
                      "REACTOME_SIGNAL_REGULATORY_PROTEIN_FAMILY_INTERACTIONS",	"REACTOME_CELL_EXTRACELLULAR_MATRIX_INTERACTIONS",	"REACTOME_INFECTION_WITH_MYCOBACTERIUM_TUBERCULOSIS",
                      "REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_MODULATION_BY_MTB_OF_HOST_IMMUNE_SYSTEM",	"REACTOME_SIGNALING_BY_PTK6",
                      "REACTOME_GAP_JUNCTION_ASSEMBLY",	"REACTOME_TRANSPORT_OF_CONNEXONS_TO_THE_PLASMA_MEMBRANE",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",
                      "REACTOME_INTERACTION_BETWEEN_L1_AND_ANKYRINS",
                      "REACTOME_GPVI_MEDIATED_ACTIVATION_CASCADE",	"REACTOME_NEGATIVE_REGULATION_OF_FLT3",	"REACTOME_PURINERGIC_SIGNALING_IN_LEISHMANIASIS_INFECTION",
                      "REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_SIGNALING_BY_MODERATE_KINASE_ACTIVITY_BRAF_MUTANTS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",
                      "REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_NEUROTRANSMITTER_RELEASE_CYCLE",	"REACTOME_INTERLEUKIN_9_SIGNALING",
                      "REACTOME_SENSORY_PROCESSING_OF_SOUND_BY_OUTER_HAIR_CELLS_OF_THE_COCHLEA",	"REACTOME_SIGNALING_BY_CSF3_G_CSF",	"REACTOME_RHO_GTPASES_ACTIVATE_PAKS",
                      "REACTOME_DAP12_INTERACTIONS",	"REACTOME_INACTIVATION_OF_CSF3_G_CSF_SIGNALING",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",
                      "REACTOME_THE_NLRP3_INFLAMMASOME",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_FLT3_SIGNALING_THROUGH_SRC_FAMILY_KINASES",
                      "REACTOME_COOPERATION_OF_PREFOLDIN_AND_TRIC_CCT_IN_ACTIN_AND_TUBULIN_FOLDING",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",
                      "REACTOME_ADHERENS_JUNCTIONS_INTERACTIONS",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",
                      "REACTOME_INFLAMMASOMES",	"REACTOME_SENSORY_PROCESSING_OF_SOUND",
                      "REACTOME_EPHA_MEDIATED_GROWTH_CONE_COLLAPSE",
                      "REACTOME_FOLDING_OF_ACTIN_BY_CCT_TRIC",
                      "REACTOME_EPH_EPHRIN_SIGNALING",
                      "REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",
                      "REACTOME_SEMAPHORIN_INTERACTIONS",
                      "REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",
                      "REACTOME_ENDOSOMAL_VACUOLAR_PATHWAY",
                      "REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",
                      "REACTOME_INTERLEUKIN_21_SIGNALING",
                      "REACTOME_RHO_GTPASES_ACTIVATE_IQGAPS",
                      "REACTOME_RHO_GTPASES_ACTIVATE_ROCKS",	"REACTOME_AGGREPHAGY",	"REACTOME_RHO_GTPASES_ACTIVATE_PKNS",	"REACTOME_NEF_AND_SIGNAL_TRANSDUCTION",
                      "REACTOME_FORMATION_OF_TUBULIN_FOLDING_INTERMEDIATES_BY_CCT_TRIC",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_SEMA3A_PAK_DEPENDENT_AXON_REPULSION",
                      "REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_CHAPERONE_MEDIATED_AUTOPHAGY",	"REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",
                      "REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_INTERLEUKIN_12_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_CIT",
                      "REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",
                      "REACTOME_GENE_AND_PROTEIN_EXPRESSION_BY_JAK_STAT_SIGNALING_AFTER_INTERLEUKIN_12_STIMULATION",
                      "REACTOME_HSF1_ACTIVATION",
                      "REACTOME_ATTENUATION_PHASE"
                      )



data_1 = M1T4vs.all_pos_reactome 
data_2 = M1T4vs.all_neg_reactome
data_3 = M2T3_pos_reactome 
data_4 = M2T3_neg_reactome


set1 = as.list(data_1)
set2 = as.list(data_2)
set3 = as.list(data_3)
set4 = as.list(data_4)

a <- list("M1T4vs.all_pos_reactome" = c("REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_PD_1_SIGNALING",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_PYROPTOSIS",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_NUCLEOTIDE_CATABOLISM",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TNFS_BIND_THEIR_PHYSIOLOGICAL_RECEPTORS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_ADHERENS_JUNCTIONS_INTERACTIONS",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_TCR_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_UCH_PROTEINASES",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	"REACTOME_APOPTOSIS",	"REACTOME_PTEN_REGULATION",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_NEDDYLATION",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_CELL_CYCLE_MITOTIC"),
          "M1T4vs.all_neg_reactome" = c("REACTOME_DCC_MEDIATED_ATTRACTIVE_SIGNALING",	"REACTOME_SCAVENGING_OF_HEME_FROM_PLASMA",	"REACTOME_COMPLEMENT_CASCADE",	"REACTOME_CREATION_OF_C4_AND_C2_ACTIVATORS",	"REACTOME_INITIAL_TRIGGERING_OF_COMPLEMENT",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_ACTIVATION_OF_THE_MRNA_UPON_BINDING_OF_THE_CAP_BINDING_COMPLEX_AND_EIFS_AND_SUBSEQUENT_BINDING_TO_43S",	"REACTOME_CELLULAR_RESPONSE_TO_STARVATION",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_RESPONSE_OF_EIF2AK4_GCN2_TO_AMINO_ACID_DEFICIENCY",	"REACTOME_NONSENSE_MEDIATED_DECAY_NMD",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_SELENOAMINO_ACID_METABOLISM",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION"),
          "M2T3_pos_reactome" = c("REACTOME_DCC_MEDIATED_ATTRACTIVE_SIGNALING",	"REACTOME_SCAVENGING_OF_HEME_FROM_PLASMA",	"REACTOME_COMPLEMENT_CASCADE",	"REACTOME_CREATION_OF_C4_AND_C2_ACTIVATORS",	"REACTOME_INITIAL_TRIGGERING_OF_COMPLEMENT",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_ACTIVATION_OF_THE_MRNA_UPON_BINDING_OF_THE_CAP_BINDING_COMPLEX_AND_EIFS_AND_SUBSEQUENT_BINDING_TO_43S",	"REACTOME_CELLULAR_RESPONSE_TO_STARVATION",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_RESPONSE_OF_EIF2AK4_GCN2_TO_AMINO_ACID_DEFICIENCY",	"REACTOME_NONSENSE_MEDIATED_DECAY_NMD",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_SELENOAMINO_ACID_METABOLISM",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION"),
          "M2T3_neg_reactome" = c("REACTOME_CELL_CYCLE_MITOTIC",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES",	"REACTOME_ASPARAGINE_N_LINKED_GLYCOSYLATION",
                                  "REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_M_PHASE",	"REACTOME_NEDDYLATION",	"REACTOME_SIGNALING_BY_GPCR",
                                  "REACTOME_ANTIGEN_PROCESSING_UBIQUITINATION_PROTEASOME_DEGRADATION",	"REACTOME_SIGNALING_BY_NUCLEAR_RECEPTORS",	"REACTOME_SUMOYLATION",
                                  "REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	"REACTOME_NEURONAL_SYSTEM",	"REACTOME_POTENTIAL_THERAPEUTICS_FOR_SARS",
                                  "REACTOME_PROCESSING_OF_CAPPED_INTRON_CONTAINING_PRE_MRNA",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",
                                  "REACTOME_NEUROTRANSMITTER_RECEPTORS_AND_POSTSYNAPTIC_SIGNAL_TRANSMISSION",	"REACTOME_TRANSPORT_TO_THE_GOLGI_AND_SUBSEQUENT_MODIFICATION",
                                  "REACTOME_ER_TO_GOLGI_ANTEROGRADE_TRANSPORT",	"REACTOME_SARS_COV_2_ACTIVATES_MODULATES_INNATE_AND_ADAPTIVE_IMMUNE_RESPONSES",
                                  "REACTOME_TRANSMISSION_ACROSS_CHEMICAL_SYNAPSES",	"REACTOME_MRNA_SPLICING",	"REACTOME_AUTOPHAGY",	"REACTOME_INTRACELLULAR_SIGNALING_BY_SECOND_MESSENGERS",
                                  "REACTOME_MITOTIC_METAPHASE_AND_ANAPHASE",	"REACTOME_CELL_SURFACE_INTERACTIONS_AT_THE_VASCULAR_WALL",	"REACTOME_HIV_INFECTION",
                                  "REACTOME_PI3K_AKT_SIGNALING_IN_CANCER",	"REACTOME_COPI_DEPENDENT_GOLGI_TO_ER_RETROGRADE_TRAFFIC",	"REACTOME_ESR_MEDIATED_SIGNALING",	"REACTOME_PTEN_REGULATION",
                                  "REACTOME_SIGNALING_BY_NOTCH",	"REACTOME_SEPARATION_OF_SISTER_CHROMATIDS",	"REACTOME_MITOCHONDRIAL_BIOGENESIS",	"REACTOME_MYD88_INDEPENDENT_TLR4_CASCADE",
                                  "REACTOME_RHO_GTPASES_ACTIVATE_FORMINS",	"REACTOME_SIGNALING_BY_RECEPTOR_TYROSINE_KINASES",	"REACTOME_CDC42_GTPASE_CYCLE",
                                  "REACTOME_CELL_DEATH_SIGNALLING_VIA_NRAGE_NRIF_AND_NADE",	"REACTOME_SIGNALING_BY_NTRKS",	"REACTOME_TRANS_GOLGI_NETWORK_VESICLE_BUDDING",
                                  "REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",	"REACTOME_RHOB_GTPASE_CYCLE",	"REACTOME_MUSCLE_CONTRACTION",	"REACTOME_CIRCADIAN_CLOCK",
                                  "REACTOME_RAB_GERANYLGERANYLATION",	"REACTOME_TOLL_LIKE_RECEPTOR_CASCADES",	"REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",
                                  "REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_RAC2_GTPASE_CYCLE",
                                  "REACTOME_REGULATION_OF_INSULIN_SECRETION",	"REACTOME_P75_NTR_RECEPTOR_MEDIATED_SIGNALLING",	"REACTOME_RHOA_GTPASE_CYCLE",	"REACTOME_HEMOSTASIS",
                                  "REACTOME_INTRINSIC_PATHWAY_FOR_APOPTOSIS",	"REACTOME_GOLGI_ASSOCIATED_VESICLE_BIOGENESIS",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",
                                  "REACTOME_ESTROGEN_DEPENDENT_GENE_EXPRESSION",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_DEATH_RECEPTOR_SIGNALLING",	"REACTOME_RHOG_GTPASE_CYCLE",
                                  "REACTOME_SIGNALING_BY_MET",	"REACTOME_DEUBIQUITINATION",	"REACTOME_RHOF_GTPASE_CYCLE",	"REACTOME_RAC1_GTPASE_CYCLE",	"REACTOME_OPIOID_SIGNALLING",
                                  "REACTOME_COPI_MEDIATED_ANTEROGRADE_TRANSPORT",	"REACTOME_RHOC_GTPASE_CYCLE",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",
                                  "REACTOME_SENSORY_PERCEPTION",	"REACTOME_APOPTOTIC_CLEAVAGE_OF_CELLULAR_PROTEINS",	"REACTOME_SIGNALING_BY_VEGF",	"REACTOME_CLATHRIN_MEDIATED_ENDOCYTOSIS",
                                  "REACTOME_UB_SPECIFIC_PROCESSING_PROTEASES",	"REACTOME_SIGNALING_BY_HEDGEHOG",	"REACTOME_COPI_INDEPENDENT_GOLGI_TO_ER_RETROGRADE_TRAFFIC",
                                  "REACTOME_DISEASES_OF_SIGNAL_TRANSDUCTION_BY_GROWTH_FACTOR_RECEPTORS_AND_SECOND_MESSENGERS",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",
                                  "REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",
                                  "REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	"REACTOME_APOPTOSIS",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",
                                  "REACTOME_RAC3_GTPASE_CYCLE",	"REACTOME_G_ALPHA_12_13_SIGNALLING_EVENTS",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",
                                  "REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX1",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_BBSOME_MEDIATED_CARGO_TARGETING_TO_CILIUM",	"REACTOME_CRISTAE_FORMATION",
                                  "REACTOME_MET_PROMOTES_CELL_MOTILITY",	"REACTOME_NETRIN_1_SIGNALING",	"REACTOME_G_BETA_GAMMA_SIGNALLING_THROUGH_PI3KGAMMA",	"REACTOME_KEAP1_NFE2L2_PATHWAY",
                                  "REACTOME_POTASSIUM_CHANNELS",	"REACTOME_PLATELET_HOMEOSTASIS",	"REACTOME_SIGNALING_BY_SCF_KIT",	"REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",
                                  "REACTOME_UPTAKE_AND_ACTIONS_OF_BACTERIAL_TOXINS",	"REACTOME_UCH_PROTEINASES",	"REACTOME_SIGNALING_BY_KIT_IN_DISEASE",	"REACTOME_PARASITE_INFECTION",
                                  "REACTOME_REGULATION_OF_KIT_SIGNALING",	"REACTOME_ASSEMBLY_OF_THE_HIV_VIRION",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",
                                  "REACTOME_COSTIMULATION_BY_THE_CD28_FAMILY",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_EPH_EPHRIN_MEDIATED_REPULSION_OF_CELLS",
                                  "REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",
                                  "REACTOME_DDX58_IFIH1_MEDIATED_INDUCTION_OF_INTERFERON_ALPHA_BETA",
                                  "REACTOME_RHO_GTPASE_EFFECTORS",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_ONCOGENIC_MAPK_SIGNALING",
                                  "REACTOME_L1CAM_INTERACTIONS",	"REACTOME_VASOPRESSIN_REGULATES_RENAL_WATER_HOMEOSTASIS_VIA_AQUAPORINS",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",
                                  "REACTOME_LEISHMANIA_INFECTION",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",
                                  "REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_P75NTR_SIGNALS_VIA_NF_KB",	"REACTOME_AQUAPORIN_MEDIATED_TRANSPORT",	"REACTOME_SIGNALING_BY_EGFR_IN_CANCER",
                                  "REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_NEUTROPHIL_DEGRANULATION",
                                  "REACTOME_CONSTITUTIVE_SIGNALING_BY_ABERRANT_PI3K_IN_CANCER",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",
                                  "REACTOME_NEGATIVE_REGULATION_OF_FGFR1_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",
                                  "REACTOME_TGF_BETA_RECEPTOR_SIGNALING_IN_EMT_EPITHELIAL_TO_MESENCHYMAL_TRANSITION",	"REACTOME_SIGNALING_BY_BRAF_AND_RAF1_FUSIONS",
                                  "REACTOME_ACTIVATION_OF_BAD_AND_TRANSLOCATION_TO_MITOCHONDRIA",	"REACTOME_CELL_JUNCTION_ORGANIZATION",	"REACTOME_INTERLEUKIN_15_SIGNALING",
                                  "REACTOME_HEDGEHOG_ON_STATE",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",
                                  "REACTOME_NEGATIVE_REGULATION_OF_MAPK_PATHWAY",	"REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",
                                  "REACTOME_GLUCAGON_LIKE_PEPTIDE_1_GLP1_REGULATES_INSULIN_SECRETION",	"REACTOME_FOXO_MEDIATED_TRANSCRIPTION_OF_CELL_CYCLE_GENES",
                                  "REACTOME_INTERLEUKIN_3_INTERLEUKIN_5_AND_GM_CSF_SIGNALING",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",
                                  "REACTOME_ASSEMBLY_AND_CELL_SURFACE_PRESENTATION_OF_NMDA_RECEPTORS",	"REACTOME_GLUCAGON_SIGNALING_IN_METABOLIC_REGULATION",	"REACTOME_ABC_TRANSPORTER_DISORDERS",
                                  "REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_RIPK1_MEDIATED_REGULATED_NECROSIS",	"REACTOME_EPHRIN_SIGNALING",	"REACTOME_CLASS_B_2_SECRETIN_FAMILY_RECEPTORS",
                                  "REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_INLB_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELL",
                                  "REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",	"REACTOME_NEGATIVE_REGULATORS_OF_DDX58_IFIH1_SIGNALING",	"REACTOME_SPRY_REGULATION_OF_FGF_SIGNALING",
                                  "REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	"REACTOME_SUMOYLATION_OF_INTRACELLULAR_RECEPTORS",
                                  "REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_C_TYPE_LECTIN_RECEPTORS_CLRS",	"REACTOME_CHK1_CHK2_CDS1_MEDIATED_INACTIVATION_OF_CYCLIN_B_CDK1_COMPLEX",
                                  "REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_PLATELET_AGGREGATION_PLUG_FORMATION",	"REACTOME_MAPK6_MAPK4_SIGNALING",
                                  "REACTOME_FGFR1_MUTANT_RECEPTOR_ACTIVATION",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",
                                  "REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",
                                  "REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_REGULATED_NECROSIS",	"REACTOME_INTEGRIN_SIGNALING",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",
                                  "REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",	"REACTOME_LISTERIA_MONOCYTOGENES_ENTRY_INTO_HOST_CELLS",
                                  "REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RAP1_SIGNALLING",	"REACTOME_INTERLEUKIN_4_AND_INTERLEUKIN_13_SIGNALING",	"REACTOME_RECYCLING_PATHWAY_OF_L1",
                                  "REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_ARYL_HYDROCARBON_RECEPTOR_SIGNALLING",
                                  "REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_ACTIVATION_OF_AMPK_DOWNSTREAM_OF_NMDARS",	"REACTOME_CELL_CELL_COMMUNICATION",
                                  "REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",
                                  "REACTOME_NUCLEOTIDE_BINDING_DOMAIN_LEUCINE_RICH_REPEAT_CONTAINING_RECEPTOR_NLR_SIGNALING_PATHWAYS",	"REACTOME_STABILIZATION_OF_P53",
                                  "REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	"REACTOME_DEGRADATION_OF_DVL",
                                  "REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_REGULATION_OF_FZD_BY_UBIQUITINATION",
                                  "REACTOME_PCP_CE_PATHWAY",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_UPTAKE_AND_FUNCTION_OF_DIPHTHERIA_TOXIN",	"REACTOME_RHOH_GTPASE_CYCLE",
                                  "REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_SIGNALING_BY_CYTOSOLIC_FGFR1_FUSION_MUTANTS",
                                  "REACTOME_SEMA3A_PLEXIN_REPULSION_SIGNALING_BY_INHIBITING_INTEGRIN_ADHESION",	"REACTOME_RESPONSE_OF_MTB_TO_PHAGOCYTOSIS",	"REACTOME_DAP12_SIGNALING",
                                  "REACTOME_GRB2_SOS_PROVIDES_LINKAGE_TO_MAPK_SIGNALING_FOR_INTEGRINS",
                                  "REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_FLT3_SIGNALING",	"REACTOME_ENOS_ACTIVATION",
                                  "REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",
                                  "REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	"REACTOME_ERBB2_REGULATES_CELL_MOTILITY",	"REACTOME_INTERLEUKIN_2_SIGNALING",
                                  "REACTOME_INTERLEUKIN_1_PROCESSING",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",
                                  "REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",	"REACTOME_DARPP_32_EVENTS",	"REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",
                                  "REACTOME_CD28_DEPENDENT_VAV1_PATHWAY",	"REACTOME_TP53_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_G1_CELL_CYCLE_ARREST",
                                  "REACTOME_POST_CHAPERONIN_TUBULIN_FOLDING_PATHWAY",	"REACTOME_CELL_CELL_JUNCTION_ORGANIZATION",	"REACTOME_GLUTAMATE_NEUROTRANSMITTER_RELEASE_CYCLE",
                                  "REACTOME_NF_KB_IS_ACTIVATED_AND_SIGNALS_SURVIVAL",	"REACTOME_P130CAS_LINKAGE_TO_MAPK_SIGNALING_FOR_INTEGRINS",
                                  "REACTOME_THE_ROLE_OF_NEF_IN_HIV_1_REPLICATION_AND_DISEASE_PATHOGENESIS",	"REACTOME_MAP2K_AND_MAPK_ACTIVATION",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",
                                  "REACTOME_SEMA4D_MEDIATED_INHIBITION_OF_CELL_ATTACHMENT_AND_MIGRATION",	"REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_PD_1_SIGNALING",	"REACTOME_TCR_SIGNALING",
                                  "REACTOME_INTERLEUKIN_2_FAMILY_SIGNALING",	"REACTOME_GABA_SYNTHESIS_RELEASE_REUPTAKE_AND_DEGRADATION",	"REACTOME_MET_ACTIVATES_RAP1_AND_RAC1",
                                  "REACTOME_SIGNAL_REGULATORY_PROTEIN_FAMILY_INTERACTIONS",	"REACTOME_CELL_EXTRACELLULAR_MATRIX_INTERACTIONS",	"REACTOME_INFECTION_WITH_MYCOBACTERIUM_TUBERCULOSIS",
                                  "REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_MODULATION_BY_MTB_OF_HOST_IMMUNE_SYSTEM",	"REACTOME_SIGNALING_BY_PTK6",
                                  "REACTOME_GAP_JUNCTION_ASSEMBLY",	"REACTOME_TRANSPORT_OF_CONNEXONS_TO_THE_PLASMA_MEMBRANE",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",
                                  "REACTOME_INTERACTION_BETWEEN_L1_AND_ANKYRINS",
                                  "REACTOME_GPVI_MEDIATED_ACTIVATION_CASCADE",	"REACTOME_NEGATIVE_REGULATION_OF_FLT3",	"REACTOME_PURINERGIC_SIGNALING_IN_LEISHMANIASIS_INFECTION",
                                  "REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_SIGNALING_BY_MODERATE_KINASE_ACTIVITY_BRAF_MUTANTS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",
                                  "REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_NEUROTRANSMITTER_RELEASE_CYCLE",	"REACTOME_INTERLEUKIN_9_SIGNALING",
                                  "REACTOME_SENSORY_PROCESSING_OF_SOUND_BY_OUTER_HAIR_CELLS_OF_THE_COCHLEA",	"REACTOME_SIGNALING_BY_CSF3_G_CSF",	"REACTOME_RHO_GTPASES_ACTIVATE_PAKS",
                                  "REACTOME_DAP12_INTERACTIONS",	"REACTOME_INACTIVATION_OF_CSF3_G_CSF_SIGNALING",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",
                                  "REACTOME_THE_NLRP3_INFLAMMASOME",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_FLT3_SIGNALING_THROUGH_SRC_FAMILY_KINASES",
                                  "REACTOME_COOPERATION_OF_PREFOLDIN_AND_TRIC_CCT_IN_ACTIN_AND_TUBULIN_FOLDING",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",
                                  "REACTOME_ADHERENS_JUNCTIONS_INTERACTIONS",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",
                                  "REACTOME_INFLAMMASOMES",	"REACTOME_SENSORY_PROCESSING_OF_SOUND",
                                  "REACTOME_EPHA_MEDIATED_GROWTH_CONE_COLLAPSE",
                                  "REACTOME_FOLDING_OF_ACTIN_BY_CCT_TRIC",
                                  "REACTOME_EPH_EPHRIN_SIGNALING",
                                  "REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",
                                  "REACTOME_SEMAPHORIN_INTERACTIONS",
                                  "REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",
                                  "REACTOME_ENDOSOMAL_VACUOLAR_PATHWAY",
                                  "REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",
                                  "REACTOME_INTERLEUKIN_21_SIGNALING",
                                  "REACTOME_RHO_GTPASES_ACTIVATE_IQGAPS",
                                  "REACTOME_RHO_GTPASES_ACTIVATE_ROCKS",	"REACTOME_AGGREPHAGY",	"REACTOME_RHO_GTPASES_ACTIVATE_PKNS",	"REACTOME_NEF_AND_SIGNAL_TRANSDUCTION",
                                  "REACTOME_FORMATION_OF_TUBULIN_FOLDING_INTERMEDIATES_BY_CCT_TRIC",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_SEMA3A_PAK_DEPENDENT_AXON_REPULSION",
                                  "REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_CHAPERONE_MEDIATED_AUTOPHAGY",	"REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",
                                  "REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_INTERLEUKIN_12_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_CIT",
                                  "REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",
                                  "REACTOME_GENE_AND_PROTEIN_EXPRESSION_BY_JAK_STAT_SIGNALING_AFTER_INTERLEUKIN_12_STIMULATION",
                                  "REACTOME_HSF1_ACTIVATION",
                                  "REACTOME_ATTENUATION_PHASE"))


#data_1 = M1T4vs.all_pos_reactome 
#data_2 = M1T4vs.all_neg_reactome
#data_3 = M2T3_pos_reactome 
#data_4 = M2T3_neg_reactome


ggvenn(a, c("M1T4vs.all_pos_reactome", "M2T3_neg_reactome"))            # draw two-set venn
ggvenn(a, c("M1T4vs.all_neg_reactome", "M2T3_pos_reactome"))
#ggvenn(a, c("Set 1", "Set 2", "Set 3"))   # draw three-set venn
ggvenn(a)   # without set names, the first 4 elements in list will be chose to draw four-set venn








M1T4pos_M2T3neg_intersect= as.data.frame(Reduce(intersect, list(set1,set4)))
M1T4pos_M2T3neg_intersect =t(M1T4pos_M2T3neg_intersect)
M1T4neg_M2T3pos_intersect= as.data.frame(Reduce(intersect, list(set2,set3)))
M1T4neg_M2T3pos_intersect =t(M1T4neg_M2T3pos_intersect)


n<-max(length(M1T4pos_M2T3neg_intersect),length(M1T4neg_M2T3pos_intersect))
length(M1T4pos_M2T3neg_intersect)<-n
length(M1T4neg_M2T3pos_intersect)<-n

M1T4_M2T3_intersect = as.data.frame(cbind(M1T4pos_M2T3neg_intersect,M1T4neg_M2T3pos_intersect))
M1T4_M2T3_intersect$M1T4pos_M2T3neg_intersect <- M1T4_M2T3_intersect$V1

write.csv(M1T4_M2T3_intersect, "_MaT4_MbT3_intersect_pos_neg_neg_pos.csv")
