==> 2_venndiagram_from_fgsea_triple.venn_Tcell_MaT4_McT4_and_PBMC-Responder.R <==
rm(list=ls())


library(RTCGA)
library(dplyr)
library(RTCGAToolbox)
library(TCGAbiolinks)
library(data.table)
library(RDocumentation)
library(ggvenn)

setwd("F:/Ko/PD-1_responders/fgsea")


########################################################################################################################################################################################################################################################################################
#######################################################################    fgsea: REACTOME results comparison     ############################################################################################################################################
#######################################################################    T cell                                 ############################################################################################################################################
########################################################################################################################################################################################################################################################################################

M1T4vs.all_pos_reactome = c("REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_PD_1_SIGNALING",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_PYROPTOSIS",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_NUCLEOTIDE_CATABOLISM",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TNFS_BIND_THEIR_PHYSIOLOGICAL_RECEPTORS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_ADHERENS_JUNCTIONS_INTERACTIONS",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_TCR_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_UCH_PROTEINASES",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	"REACTOME_APOPTOSIS",	"REACTOME_PTEN_REGULATION",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_NEDDYLATION",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_CELL_CYCLE_MITOTIC"
)
M3T4vs.all_po_reactome = c("REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_PD_1_SIGNALING",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TCR_SIGNALING",	"REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",	"REACTOME_NOTCH2_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_INACTIVATION_OF_CDC42_AND_RAC1",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_APOPTOSIS",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_HEMOSTASIS",	"REACTOME_SIGNALING_BY_INTERLEUKINS"
)

RespondervsNonreponder_pos_reactome =c("REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_CHAPERONE_MEDIATED_AUTOPHAGY",	"REACTOME_CONSTITUTIVE_SIGNALING_BY_LIGAND_RESPONSIVE_EGFR_CANCER_VARIANTS",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_INLB_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELL",	"REACTOME_RESPONSE_TO_METAL_IONS",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY",	"REACTOME_SIGNALING_BY_EGFR_IN_CANCER",	"REACTOME_DOWNREGULATION_OF_ERBB2_ERBB3_SIGNALING",	"REACTOME_LISTERIA_MONOCYTOGENES_ENTRY_INTO_HOST_CELLS",
                                       "REACTOME_EGFR_DOWNREGULATION",	"REACTOME_NEGATIVE_REGULATION_OF_MET_ACTIVITY",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION",	"REACTOME_TGF_BETA_RECEPTOR_SIGNALING_IN_EMT_EPITHELIAL_TO_MESENCHYMAL_TRANSITION",	"REACTOME_IRAK2_MEDIATED_ACTIVATION_OF_TAK1_COMPLEX",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_METALLOTHIONEINS_BIND_METALS",	"REACTOME_RHO_GTPASES_ACTIVATE_KTN1",	"REACTOME_PTK6_REGULATES_RTKS_AND_THEIR_EFFECTORS_AKT1_AND_DOK1",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_TICAM1_TRAF6_DEPENDENT_INDUCTION_OF_TAK1_COMPLEX",	
                                       "REACTOME_FLT3_SIGNALING_BY_CBL_MUTANTS",	"REACTOME_BINDING_AND_UPTAKE_OF_LIGANDS_BY_SCAVENGER_RECEPTORS",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_REGULATION_OF_BACH1_ACTIVITY",	"REACTOME_IRAK1_RECRUITS_IKK_COMPLEX",	"REACTOME_SIGNALING_BY_NOTCH1_HD_DOMAIN_MUTANTS_IN_CANCER",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_FOLDING_OF_ACTIN_BY_CCT_TRIC",	"REACTOME_INLA_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELLS",	"REACTOME_TICAM1_RIP1_MEDIATED_IKK_COMPLEX_RECRUITMENT",	"REACTOME_GENE_AND_PROTEIN_EXPRESSION_BY_JAK_STAT_SIGNALING_AFTER_INTERLEUKIN_12_STIMULATION",	
                                       "REACTOME_MITOPHAGY",	"REACTOME_PEXOPHAGY",	"REACTOME_SIGNALING_BY_NTRK3_TRKC",	"REACTOME_GLYCOGEN_STORAGE_DISEASES",	"REACTOME_COOPERATION_OF_PREFOLDIN_AND_TRIC_CCT_IN_ACTIN_AND_TUBULIN_FOLDING",	"REACTOME_RESPONSE_OF_MTB_TO_PHAGOCYTOSIS",	"REACTOME_PINK1_PRKN_MEDIATED_MITOPHAGY",	"REACTOME_KILLING_MECHANISMS",	"REACTOME_UPTAKE_AND_ACTIONS_OF_BACTERIAL_TOXINS",	"REACTOME_ACTIVATION_OF_THE_MRNA_UPON_BINDING_OF_THE_CAP_BINDING_COMPLEX_AND_EIFS_AND_SUBSEQUENT_BINDING_TO_43S",	"REACTOME_GLYCOGEN_SYNTHESIS",	"REACTOME_FORMATION_OF_TUBULIN_FOLDING_INTERMEDIATES_BY_CCT_TRIC",	"REACTOME_TRAF6_MEDIATED_INDUCTION_OF_TAK1_COMPLEX_WITHIN_TLR4_COMPLEX",	
                                       "REACTOME_IKK_COMPLEX_RECRUITMENT_MEDIATED_BY_RIP1",	"REACTOME_CONSTITUTIVE_SIGNALING_BY_EGFRVIII",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_PD_1_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_IQGAPS",	"REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_SIGNALING_BY_FGFR4_IN_DISEASE",	"REACTOME_PTK6_REGULATES_RHO_GTPASES_RAS_GTPASE_AND_MAP_KINASES",	"REACTOME_ASSEMBLY_OF_THE_HIV_VIRION",	"REACTOME_SEMA3A_PAK_DEPENDENT_AXON_REPULSION",	"REACTOME_AGGREPHAGY",	"REACTOME_NF_KB_IS_ACTIVATED_AND_SIGNALS_SURVIVAL",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	
                                       "REACTOME_SIGNALING_BY_ERBB2",	"REACTOME_SIGNALING_BY_FGFR1",	"REACTOME_SIGNALING_BY_FGFR3",	"REACTOME_SIGNALING_BY_FGFR4",	"REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",	"REACTOME_DETOXIFICATION_OF_REACTIVE_OXYGEN_SPECIES",	"REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_P75NTR_SIGNALS_VIA_NF_KB",	"REACTOME_PREVENTION_OF_PHAGOSOMAL_LYSOSOMAL_FUSION",	"REACTOME_DOWNREGULATION_OF_ERBB4_SIGNALING",	"REACTOME_SIGNALING_BY_ERBB2_ECD_MUTANTS",	"REACTOME_SIGNALING_BY_PTK6",	"REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_INTEGRATION_OF_PROVIRUS",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR2_SIGNALING",
                                       "REACTOME_NEGATIVE_REGULATION_OF_FGFR3_SIGNALING",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR4_SIGNALING",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR1_SIGNALING",	"REACTOME_SIGNALING_BY_ROBO_RECEPTORS",	"REACTOME_JNK_C_JUN_KINASES_PHOSPHORYLATION_AND_ACTIVATION_MEDIATED_BY_ACTIVATED_HUMAN_TAK1",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_REGULATION_OF_INNATE_IMMUNE_RESPONSES_TO_CYTOSOLIC_DNA",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_INFECTION_WITH_MYCOBACTERIUM_TUBERCULOSIS",	"REACTOME_NEGATIVE_REGULATION_OF_FLT3",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",	"REACTOME_FLT3_SIGNALING_IN_DISEASE",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_PCP_CE_PATHWAY",	
                                       "REACTOME_SPRY_REGULATION_OF_FGF_SIGNALING",	"REACTOME_NUCLEAR_EVENTS_STIMULATED_BY_ALK_SIGNALING_IN_CANCER",	"REACTOME_REGULATION_OF_EXPRESSION_OF_SLITS_AND_ROBOS",	"REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_SIGNALING_BY_MET",	"REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_GABA_SYNTHESIS_RELEASE_REUPTAKE_AND_DEGRADATION",	"REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",	"REACTOME_P75NTR_RECRUITS_SIGNALLING_COMPLEXES",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_MYOCLONIC_EPILEPSY_OF_LAFORA",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_REGULATION_OF_SIGNALING_BY_CBL",	
                                       "REACTOME_SIGNALING_BY_EGFR",	"REACTOME_STIMULI_SENSING_CHANNELS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_ALPHA_PROTEIN_KINASE_1_SIGNALING_PATHWAY",	"REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	"REACTOME_INTERLEUKIN_12_SIGNALING",	"REACTOME_UPTAKE_AND_FUNCTION_OF_DIPHTHERIA_TOXIN",	"REACTOME_RAS_PROCESSING",	"REACTOME_MRNA_SPLICING",	"REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_MAP3K8_TPL2_DEPENDENT_MAPK1_3_ACTIVATION",	"REACTOME_TRANSPORT_OF_NUCLEOSIDES_AND_FREE_PURINE_AND_PYRIMIDINE_BASES_ACROSS_THE_PLASMA_MEMBRANE",	
                                       "REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	"REACTOME_ACTIVATION_OF_RAC1_DOWNSTREAM_OF_NMDARS",	"REACTOME_REGULATION_OF_PTEN_LOCALIZATION",	"REACTOME_SUPPRESSION_OF_PHAGOSOMAL_MATURATION",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_SCAVENGING_BY_CLASS_A_RECEPTORS",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",	"REACTOME_GP1B_IX_V_ACTIVATION_SIGNALLING",	"REACTOME_REGULATION_OF_NF_KAPPA_B_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	
                                       "REACTOME_CALCINEURIN_ACTIVATES_NFAT",	"REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_2_LTR_CIRCLE_FORMATION",	"REACTOME_TCR_SIGNALING",	"REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_SIGNALING_BY_ERBB2_IN_CANCER",	"REACTOME_GPVI_MEDIATED_ACTIVATION_CASCADE",	"REACTOME_EARLY_PHASE_OF_HIV_LIFE_CYCLE",	"REACTOME_NEGATIVE_REGULATION_OF_MAPK_PATHWAY",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",
                                       "REACTOME_VLDLR_INTERNALISATION_AND_DEGRADATION",	"REACTOME_NOTCH3_ACTIVATION_AND_TRANSMISSION_OF_SIGNAL_TO_THE_NUCLEUS",	"REACTOME_KEAP1_NFE2L2_PATHWAY",	"REACTOME_SHC_MEDIATED_CASCADE_FGFR1",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_TP53_REGULATES_METABOLIC_GENES",	"REACTOME_FLT3_SIGNALING",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_GLYCOGEN_METABOLISM",
                                       "REACTOME_SYNTHESIS_OF_5_EICOSATETRAENOIC_ACIDS",	"REACTOME_MITOCHONDRIAL_BIOGENESIS",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_ATF6_ATF6_ALPHA_ACTIVATES_CHAPERONES",	"REACTOME_POST_CHAPERONIN_TUBULIN_FOLDING_PATHWAY",	"REACTOME_RHOBTB1_GTPASE_CYCLE",	"REACTOME_HEDGEHOG_ON_STATE",	"REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_TRANSLATION",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_TGF_BETA_RECEPTOR_SIGNALING_ACTIVATES_SMADS",	"REACTOME_ACTIVATED_NTRK2_SIGNALS_THROUGH_RAS",	"REACTOME_ACTIVATED_NTRK3_SIGNALS_THROUGH_RAS",
                                       "REACTOME_SHC_MEDIATED_CASCADE_FGFR3",	"REACTOME_SHC_MEDIATED_CASCADE_FGFR4",	"REACTOME_SIGNALING_BY_ERBB4",	"REACTOME_EGFR_TRANSACTIVATION_BY_GASTRIN",	"REACTOME_VEGFR2_MEDIATED_CELL_PROLIFERATION",	"REACTOME_SIGNALING_BY_NTRK2_TRKB",	"REACTOME_AMYLOID_FIBER_FORMATION",	"REACTOME_RHO_GTPASES_ACTIVATE_PAKS",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_LATE_ENDOSOMAL_MICROAUTOPHAGY",	"REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",	"REACTOME_COSTIMULATION_BY_THE_CD28_FAMILY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",
                                       "REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_SIGNALING_BY_ALK_IN_CANCER",	"REACTOME_PROCESSING_OF_CAPPED_INTRON_CONTAINING_PRE_MRNA",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_MRNA_SPLICING_MINOR_PATHWAY",	"REACTOME_IRON_UPTAKE_AND_TRANSPORT",	"REACTOME_DDX58_IFIH1_MEDIATED_INDUCTION_OF_INTERFERON_ALPHA_BETA",	"REACTOME_DEFECTIVE_INTRINSIC_PATHWAY_FOR_APOPTOSIS",	"REACTOME_FCERI_MEDIATED_MAPK_ACTIVATION",	"REACTOME_SIGNALING_BY_FGFR2",	"REACTOME_C_TYPE_LECTIN_RECEPTORS_CLRS",	"REACTOME_SIGNALING_BY_FGFR",	"REACTOME_SIGNALING_BY_HEDGEHOG",	"REACTOME_SIGNALING_BY_CSF3_G_CSF",
                                       "REACTOME_SIGNALING_BY_ERYTHROPOIETIN",	"REACTOME_NERVOUS_SYSTEM_DEVELOPMENT",	"REACTOME_OVARIAN_TUMOR_DOMAIN_PROTEASES",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_DNA_REPLICATION_PRE_INITIATION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_B_WICH_COMPLEX_POSITIVELY_REGULATES_RRNA_EXPRESSION",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_PTEN_REGULATION",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_HIV_INFECTION",	"REACTOME_SMAD2_SMAD3_SMAD4_HETEROTRIMER_REGULATES_TRANSCRIPTION",	"REACTOME_CELL_SURFACE_INTERACTIONS_AT_THE_VASCULAR_WALL",	"REACTOME_NOD1_2_SIGNALING_PATHWAY",	
                                       "REACTOME_NEGATIVE_REGULATORS_OF_DDX58_IFIH1_SIGNALING",	"REACTOME_TAK1_DEPENDENT_IKK_AND_NF_KAPPA_B_ACTIVATION",	"REACTOME_NOTCH1_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_DISEASES_OF_PROGRAMMED_CELL_DEATH",	"REACTOME_POSITIVE_EPIGENETIC_REGULATION_OF_RRNA_EXPRESSION",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_UCH_PROTEINASES",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_REGULATION_OF_TNFR1_SIGNALING",	"REACTOME_PLASMA_LIPOPROTEIN_ASSEMBLY_REMODELING_AND_CLEARANCE",	"REACTOME_FORMATION_OF_INCISION_COMPLEX_IN_GG_NER",	"REACTOME_SIGNALING_BY_TGF_BETA_RECEPTOR_COMPLEX",	"REACTOME_SIGNALING_BY_TGFB_FAMILY_MEMBERS",
                                       "REACTOME_SENESCENCE_ASSOCIATED_SECRETORY_PHENOTYPE_SASP",	"REACTOME_MITOCHONDRIAL_PROTEIN_IMPORT",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_DNA_REPLICATION",	"REACTOME_RNA_POLYMERASE_II_TRANSCRIPTION_TERMINATION",	"REACTOME_FORMATION_OF_TC_NER_PRE_INCISION_COMPLEX",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_EPIGENETIC_REGULATION_OF_GENE_EXPRESSION",	"REACTOME_SIGNALING_BY_NOTCH",	"REACTOME_APOPTOSIS",	"REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",	
                                       "REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_S_PHASE",	"REACTOME_ESTROGEN_DEPENDENT_GENE_EXPRESSION",	"REACTOME_EXTRA_NUCLEAR_ESTROGEN_SIGNALING",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_L1CAM_INTERACTIONS",	"REACTOME_UB_SPECIFIC_PROCESSING_PROTEASES",	"REACTOME_AUTOPHAGY",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_SIGNALING_BY_VEGF",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX1",	"REACTOME_PROTEIN_UBIQUITINATION",	"REACTOME_INTRACELLULAR_SIGNALING_BY_SECOND_MESSENGERS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_DISEASES_OF_SIGNAL_TRANSDUCTION_BY_GROWTH_FACTOR_RECEPTORS_AND_SECOND_MESSENGERS",
                                       "REACTOME_SEPARATION_OF_SISTER_CHROMATIDS",	"REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	"REACTOME_REGULATION_OF_PLK1_ACTIVITY_AT_G2_M_TRANSITION",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_CYTOPROTECTION_BY_HMOX1",	"REACTOME_PROTEIN_LOCALIZATION",	"REACTOME_RHOG_GTPASE_CYCLE",	"REACTOME_TRANSCRIPTION_OF_THE_HIV_GENOME",	"REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	"REACTOME_DUAL_INCISION_IN_TC_NER",	"REACTOME_CLATHRIN_MEDIATED_ENDOCYTOSIS",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_METAPHASE_AND_ANAPHASE",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_TP53",	"REACTOME_POTENTIAL_THERAPEUTICS_FOR_SARS",	"REACTOME_RHO_GTPASE_EFFECTORS",
                                       "REACTOME_ESR_MEDIATED_SIGNALING",	"REACTOME_HEMOSTASIS",	"REACTOME_NEUROTRANSMITTER_RECEPTORS_AND_POSTSYNAPTIC_SIGNAL_TRANSMISSION",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_DEVELOPMENTAL_BIOLOGY",	"REACTOME_ANTIGEN_PROCESSING_UBIQUITINATION_PROTEASOME_DEGRADATION",	"REACTOME_NEDDYLATION",	"REACTOME_SARS_COV_2_HOST_INTERACTIONS",	"REACTOME_M_PHASE",	"REACTOME_NUCLEOTIDE_EXCISION_REPAIR",	"REACTOME_TRANSMISSION_ACROSS_CHEMICAL_SYNAPSES",	"REACTOME_HIV_LIFE_CYCLE",	"REACTOME_METABOLISM_OF_AMINO_ACIDS_AND_DERIVATIVES",	"REACTOME_DEUBIQUITINATION",	"REACTOME_SIGNALING_BY_RECEPTOR_TYROSINE_KINASES",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",
                                       "REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_CELL_CYCLE_CHECKPOINTS",	"REACTOME_SARS_COV_INFECTIONS",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES",	"REACTOME_CELL_CYCLE_MITOTIC",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_SIGNALING_BY_RHO_GTPASES_MIRO_GTPASES_AND_RHOBTB3",	"REACTOME_NEUTROPHIL_DEGRANULATION")
#RespondervsNonreponder_neg_reactome = c("REACTOME_INTERLEUKIN_18_SIGNALING")



data_1 = M1T4vs.all_pos_reactome 
data_2 = M3T4vs.all_po_reactome
data_3 = RespondervsNonreponder_pos_reactome 
#data_4 = RespondervsNonreponder_neg_reactome


set1 = as.list(data_1)
set2 = as.list(data_2)
set3 = as.list(data_3)
#set4 = as.list(data_4)

venn.diagram(x=list(set1, set2, set3), 
             category.names = c("MaT4","McT4","R"), filename = "venndiagram_Tcell_REACTOME_MaT4_McT4_Responder.png", output=T, 
             imagetype = "png", height = 2000, width = 2000, resolution = 300, compression = "lzw",
             fill=c("skyblue", "yellowgreen","pink"), 
             cex = 2,  lwd=2, lty='blank', fontface="bold", fontfamily="sans", cat.cex=1, cat.fontface="bold", cat.default.pos="outer", cat.pos=c(0, 0, 200),cat.fontfamily="sans")




a <- list("M1T4vs.all_pos_reactome" = c("REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_PD_1_SIGNALING",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_PYROPTOSIS",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_NUCLEOTIDE_CATABOLISM",	"REACTOME_GAP_JUNCTION_DEGRADATION",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TNFS_BIND_THEIR_PHYSIOLOGICAL_RECEPTORS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_ADHERENS_JUNCTIONS_INTERACTIONS",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_TCR_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_UCH_PROTEINASES",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	"REACTOME_APOPTOSIS",	"REACTOME_PTEN_REGULATION",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_NEDDYLATION",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_CELL_CYCLE_MITOTIC"
),
"M3T4vs.all_pos_reactome" = c("REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_PD_1_SIGNALING",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_TCR_SIGNALING",	"REACTOME_GOLGI_TO_ER_RETROGRADE_TRANSPORT",	"REACTOME_NOTCH2_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_INACTIVATION_OF_CDC42_AND_RAC1",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_APOPTOSIS",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_HEMOSTASIS",	"REACTOME_SIGNALING_BY_INTERLEUKINS"
),
"RespondervsNonreponder_pos_reactome" = c("REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_CHAPERONE_MEDIATED_AUTOPHAGY",	"REACTOME_CONSTITUTIVE_SIGNALING_BY_LIGAND_RESPONSIVE_EGFR_CANCER_VARIANTS",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_INLB_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELL",	"REACTOME_RESPONSE_TO_METAL_IONS",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY",	"REACTOME_SIGNALING_BY_EGFR_IN_CANCER",	"REACTOME_DOWNREGULATION_OF_ERBB2_ERBB3_SIGNALING",	"REACTOME_LISTERIA_MONOCYTOGENES_ENTRY_INTO_HOST_CELLS",
                                          "REACTOME_EGFR_DOWNREGULATION",	"REACTOME_NEGATIVE_REGULATION_OF_MET_ACTIVITY",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION",	"REACTOME_TGF_BETA_RECEPTOR_SIGNALING_IN_EMT_EPITHELIAL_TO_MESENCHYMAL_TRANSITION",	"REACTOME_IRAK2_MEDIATED_ACTIVATION_OF_TAK1_COMPLEX",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_METALLOTHIONEINS_BIND_METALS",	"REACTOME_RHO_GTPASES_ACTIVATE_KTN1",	"REACTOME_PTK6_REGULATES_RTKS_AND_THEIR_EFFECTORS_AKT1_AND_DOK1",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_TICAM1_TRAF6_DEPENDENT_INDUCTION_OF_TAK1_COMPLEX",	
                                          "REACTOME_FLT3_SIGNALING_BY_CBL_MUTANTS",	"REACTOME_BINDING_AND_UPTAKE_OF_LIGANDS_BY_SCAVENGER_RECEPTORS",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_REGULATION_OF_BACH1_ACTIVITY",	"REACTOME_IRAK1_RECRUITS_IKK_COMPLEX",	"REACTOME_SIGNALING_BY_NOTCH1_HD_DOMAIN_MUTANTS_IN_CANCER",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_FOLDING_OF_ACTIN_BY_CCT_TRIC",	"REACTOME_INLA_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELLS",	"REACTOME_TICAM1_RIP1_MEDIATED_IKK_COMPLEX_RECRUITMENT",	"REACTOME_GENE_AND_PROTEIN_EXPRESSION_BY_JAK_STAT_SIGNALING_AFTER_INTERLEUKIN_12_STIMULATION",	
                                          "REACTOME_MITOPHAGY",	"REACTOME_PEXOPHAGY",	"REACTOME_SIGNALING_BY_NTRK3_TRKC",	"REACTOME_GLYCOGEN_STORAGE_DISEASES",	"REACTOME_COOPERATION_OF_PREFOLDIN_AND_TRIC_CCT_IN_ACTIN_AND_TUBULIN_FOLDING",	"REACTOME_RESPONSE_OF_MTB_TO_PHAGOCYTOSIS",	"REACTOME_PINK1_PRKN_MEDIATED_MITOPHAGY",	"REACTOME_KILLING_MECHANISMS",	"REACTOME_UPTAKE_AND_ACTIONS_OF_BACTERIAL_TOXINS",	"REACTOME_ACTIVATION_OF_THE_MRNA_UPON_BINDING_OF_THE_CAP_BINDING_COMPLEX_AND_EIFS_AND_SUBSEQUENT_BINDING_TO_43S",	"REACTOME_GLYCOGEN_SYNTHESIS",	"REACTOME_FORMATION_OF_TUBULIN_FOLDING_INTERMEDIATES_BY_CCT_TRIC",	"REACTOME_TRAF6_MEDIATED_INDUCTION_OF_TAK1_COMPLEX_WITHIN_TLR4_COMPLEX",	
                                          "REACTOME_IKK_COMPLEX_RECRUITMENT_MEDIATED_BY_RIP1",	"REACTOME_CONSTITUTIVE_SIGNALING_BY_EGFRVIII",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_PD_1_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_IQGAPS",	"REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_SIGNALING_BY_FGFR4_IN_DISEASE",	"REACTOME_PTK6_REGULATES_RHO_GTPASES_RAS_GTPASE_AND_MAP_KINASES",	"REACTOME_ASSEMBLY_OF_THE_HIV_VIRION",	"REACTOME_SEMA3A_PAK_DEPENDENT_AXON_REPULSION",	"REACTOME_AGGREPHAGY",	"REACTOME_NF_KB_IS_ACTIVATED_AND_SIGNALS_SURVIVAL",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	
                                          "REACTOME_SIGNALING_BY_ERBB2",	"REACTOME_SIGNALING_BY_FGFR1",	"REACTOME_SIGNALING_BY_FGFR3",	"REACTOME_SIGNALING_BY_FGFR4",	"REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",	"REACTOME_DETOXIFICATION_OF_REACTIVE_OXYGEN_SPECIES",	"REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_P75NTR_SIGNALS_VIA_NF_KB",	"REACTOME_PREVENTION_OF_PHAGOSOMAL_LYSOSOMAL_FUSION",	"REACTOME_DOWNREGULATION_OF_ERBB4_SIGNALING",	"REACTOME_SIGNALING_BY_ERBB2_ECD_MUTANTS",	"REACTOME_SIGNALING_BY_PTK6",	"REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_INTEGRATION_OF_PROVIRUS",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR2_SIGNALING",
                                          "REACTOME_NEGATIVE_REGULATION_OF_FGFR3_SIGNALING",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR4_SIGNALING",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR1_SIGNALING",	"REACTOME_SIGNALING_BY_ROBO_RECEPTORS",	"REACTOME_JNK_C_JUN_KINASES_PHOSPHORYLATION_AND_ACTIVATION_MEDIATED_BY_ACTIVATED_HUMAN_TAK1",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_REGULATION_OF_INNATE_IMMUNE_RESPONSES_TO_CYTOSOLIC_DNA",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_INFECTION_WITH_MYCOBACTERIUM_TUBERCULOSIS",	"REACTOME_NEGATIVE_REGULATION_OF_FLT3",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",	"REACTOME_FLT3_SIGNALING_IN_DISEASE",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_PCP_CE_PATHWAY",	
                                          "REACTOME_SPRY_REGULATION_OF_FGF_SIGNALING",	"REACTOME_NUCLEAR_EVENTS_STIMULATED_BY_ALK_SIGNALING_IN_CANCER",	"REACTOME_REGULATION_OF_EXPRESSION_OF_SLITS_AND_ROBOS",	"REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_SIGNALING_BY_MET",	"REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_GABA_SYNTHESIS_RELEASE_REUPTAKE_AND_DEGRADATION",	"REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",	"REACTOME_P75NTR_RECRUITS_SIGNALLING_COMPLEXES",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_MYOCLONIC_EPILEPSY_OF_LAFORA",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_REGULATION_OF_SIGNALING_BY_CBL",	
                                          "REACTOME_SIGNALING_BY_EGFR",	"REACTOME_STIMULI_SENSING_CHANNELS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_ALPHA_PROTEIN_KINASE_1_SIGNALING_PATHWAY",	"REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	"REACTOME_INTERLEUKIN_12_SIGNALING",	"REACTOME_UPTAKE_AND_FUNCTION_OF_DIPHTHERIA_TOXIN",	"REACTOME_RAS_PROCESSING",	"REACTOME_MRNA_SPLICING",	"REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_MAP3K8_TPL2_DEPENDENT_MAPK1_3_ACTIVATION",	"REACTOME_TRANSPORT_OF_NUCLEOSIDES_AND_FREE_PURINE_AND_PYRIMIDINE_BASES_ACROSS_THE_PLASMA_MEMBRANE",	
                                          "REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	"REACTOME_ACTIVATION_OF_RAC1_DOWNSTREAM_OF_NMDARS",	"REACTOME_REGULATION_OF_PTEN_LOCALIZATION",	"REACTOME_SUPPRESSION_OF_PHAGOSOMAL_MATURATION",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_SCAVENGING_BY_CLASS_A_RECEPTORS",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",	"REACTOME_GP1B_IX_V_ACTIVATION_SIGNALLING",	"REACTOME_REGULATION_OF_NF_KAPPA_B_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	
                                          "REACTOME_CALCINEURIN_ACTIVATES_NFAT",	"REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_2_LTR_CIRCLE_FORMATION",	"REACTOME_TCR_SIGNALING",	"REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_SIGNALING_BY_ERBB2_IN_CANCER",	"REACTOME_GPVI_MEDIATED_ACTIVATION_CASCADE",	"REACTOME_EARLY_PHASE_OF_HIV_LIFE_CYCLE",	"REACTOME_NEGATIVE_REGULATION_OF_MAPK_PATHWAY",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",
                                          "REACTOME_VLDLR_INTERNALISATION_AND_DEGRADATION",	"REACTOME_NOTCH3_ACTIVATION_AND_TRANSMISSION_OF_SIGNAL_TO_THE_NUCLEUS",	"REACTOME_KEAP1_NFE2L2_PATHWAY",	"REACTOME_SHC_MEDIATED_CASCADE_FGFR1",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_TP53_REGULATES_METABOLIC_GENES",	"REACTOME_FLT3_SIGNALING",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_GLYCOGEN_METABOLISM",
                                          "REACTOME_SYNTHESIS_OF_5_EICOSATETRAENOIC_ACIDS",	"REACTOME_MITOCHONDRIAL_BIOGENESIS",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_ATF6_ATF6_ALPHA_ACTIVATES_CHAPERONES",	"REACTOME_POST_CHAPERONIN_TUBULIN_FOLDING_PATHWAY",	"REACTOME_RHOBTB1_GTPASE_CYCLE",	"REACTOME_HEDGEHOG_ON_STATE",	"REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_TRANSLATION",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_TGF_BETA_RECEPTOR_SIGNALING_ACTIVATES_SMADS",	"REACTOME_ACTIVATED_NTRK2_SIGNALS_THROUGH_RAS",	"REACTOME_ACTIVATED_NTRK3_SIGNALS_THROUGH_RAS",
                                          "REACTOME_SHC_MEDIATED_CASCADE_FGFR3",	"REACTOME_SHC_MEDIATED_CASCADE_FGFR4",	"REACTOME_SIGNALING_BY_ERBB4",	"REACTOME_EGFR_TRANSACTIVATION_BY_GASTRIN",	"REACTOME_VEGFR2_MEDIATED_CELL_PROLIFERATION",	"REACTOME_SIGNALING_BY_NTRK2_TRKB",	"REACTOME_AMYLOID_FIBER_FORMATION",	"REACTOME_RHO_GTPASES_ACTIVATE_PAKS",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_LATE_ENDOSOMAL_MICROAUTOPHAGY",	"REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",	"REACTOME_COSTIMULATION_BY_THE_CD28_FAMILY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",
                                          "REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_SIGNALING_BY_ALK_IN_CANCER",	"REACTOME_PROCESSING_OF_CAPPED_INTRON_CONTAINING_PRE_MRNA",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_MRNA_SPLICING_MINOR_PATHWAY",	"REACTOME_IRON_UPTAKE_AND_TRANSPORT",	"REACTOME_DDX58_IFIH1_MEDIATED_INDUCTION_OF_INTERFERON_ALPHA_BETA",	"REACTOME_DEFECTIVE_INTRINSIC_PATHWAY_FOR_APOPTOSIS",	"REACTOME_FCERI_MEDIATED_MAPK_ACTIVATION",	"REACTOME_SIGNALING_BY_FGFR2",	"REACTOME_C_TYPE_LECTIN_RECEPTORS_CLRS",	"REACTOME_SIGNALING_BY_FGFR",	"REACTOME_SIGNALING_BY_HEDGEHOG",	"REACTOME_SIGNALING_BY_CSF3_G_CSF",
                                          "REACTOME_SIGNALING_BY_ERYTHROPOIETIN",	"REACTOME_NERVOUS_SYSTEM_DEVELOPMENT",	"REACTOME_OVARIAN_TUMOR_DOMAIN_PROTEASES",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_DNA_REPLICATION_PRE_INITIATION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_B_WICH_COMPLEX_POSITIVELY_REGULATES_RRNA_EXPRESSION",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_PTEN_REGULATION",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_HIV_INFECTION",	"REACTOME_SMAD2_SMAD3_SMAD4_HETEROTRIMER_REGULATES_TRANSCRIPTION",	"REACTOME_CELL_SURFACE_INTERACTIONS_AT_THE_VASCULAR_WALL",	"REACTOME_NOD1_2_SIGNALING_PATHWAY",	
                                          "REACTOME_NEGATIVE_REGULATORS_OF_DDX58_IFIH1_SIGNALING",	"REACTOME_TAK1_DEPENDENT_IKK_AND_NF_KAPPA_B_ACTIVATION",	"REACTOME_NOTCH1_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_DISEASES_OF_PROGRAMMED_CELL_DEATH",	"REACTOME_POSITIVE_EPIGENETIC_REGULATION_OF_RRNA_EXPRESSION",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_UCH_PROTEINASES",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_REGULATION_OF_TNFR1_SIGNALING",	"REACTOME_PLASMA_LIPOPROTEIN_ASSEMBLY_REMODELING_AND_CLEARANCE",	"REACTOME_FORMATION_OF_INCISION_COMPLEX_IN_GG_NER",	"REACTOME_SIGNALING_BY_TGF_BETA_RECEPTOR_COMPLEX",	"REACTOME_SIGNALING_BY_TGFB_FAMILY_MEMBERS",
                                          "REACTOME_SENESCENCE_ASSOCIATED_SECRETORY_PHENOTYPE_SASP",	"REACTOME_MITOCHONDRIAL_PROTEIN_IMPORT",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_DNA_REPLICATION",	"REACTOME_RNA_POLYMERASE_II_TRANSCRIPTION_TERMINATION",	"REACTOME_FORMATION_OF_TC_NER_PRE_INCISION_COMPLEX",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_EPIGENETIC_REGULATION_OF_GENE_EXPRESSION",	"REACTOME_SIGNALING_BY_NOTCH",	"REACTOME_APOPTOSIS",	"REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",	
                                          "REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_S_PHASE",	"REACTOME_ESTROGEN_DEPENDENT_GENE_EXPRESSION",	"REACTOME_EXTRA_NUCLEAR_ESTROGEN_SIGNALING",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_L1CAM_INTERACTIONS",	"REACTOME_UB_SPECIFIC_PROCESSING_PROTEASES",	"REACTOME_AUTOPHAGY",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_SIGNALING_BY_VEGF",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX1",	"REACTOME_PROTEIN_UBIQUITINATION",	"REACTOME_INTRACELLULAR_SIGNALING_BY_SECOND_MESSENGERS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_DISEASES_OF_SIGNAL_TRANSDUCTION_BY_GROWTH_FACTOR_RECEPTORS_AND_SECOND_MESSENGERS",
                                          "REACTOME_SEPARATION_OF_SISTER_CHROMATIDS",	"REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	"REACTOME_REGULATION_OF_PLK1_ACTIVITY_AT_G2_M_TRANSITION",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_CYTOPROTECTION_BY_HMOX1",	"REACTOME_PROTEIN_LOCALIZATION",	"REACTOME_RHOG_GTPASE_CYCLE",	"REACTOME_TRANSCRIPTION_OF_THE_HIV_GENOME",	"REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	"REACTOME_DUAL_INCISION_IN_TC_NER",	"REACTOME_CLATHRIN_MEDIATED_ENDOCYTOSIS",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_METAPHASE_AND_ANAPHASE",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_TP53",	"REACTOME_POTENTIAL_THERAPEUTICS_FOR_SARS",	"REACTOME_RHO_GTPASE_EFFECTORS",
                                          "REACTOME_ESR_MEDIATED_SIGNALING",	"REACTOME_HEMOSTASIS",	"REACTOME_NEUROTRANSMITTER_RECEPTORS_AND_POSTSYNAPTIC_SIGNAL_TRANSMISSION",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_DEVELOPMENTAL_BIOLOGY",	"REACTOME_ANTIGEN_PROCESSING_UBIQUITINATION_PROTEASOME_DEGRADATION",	"REACTOME_NEDDYLATION",	"REACTOME_SARS_COV_2_HOST_INTERACTIONS",	"REACTOME_M_PHASE",	"REACTOME_NUCLEOTIDE_EXCISION_REPAIR",	"REACTOME_TRANSMISSION_ACROSS_CHEMICAL_SYNAPSES",	"REACTOME_HIV_LIFE_CYCLE",	"REACTOME_METABOLISM_OF_AMINO_ACIDS_AND_DERIVATIVES",	"REACTOME_DEUBIQUITINATION",	"REACTOME_SIGNALING_BY_RECEPTOR_TYROSINE_KINASES",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",
                                          "REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_CELL_CYCLE_CHECKPOINTS",	"REACTOME_SARS_COV_INFECTIONS",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES",	"REACTOME_CELL_CYCLE_MITOTIC",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_SIGNALING_BY_RHO_GTPASES_MIRO_GTPASES_AND_RHOBTB3",	"REACTOME_NEUTROPHIL_DEGRANULATION")
#,
#"RespondervsNonreponder_neg_reactome" = c("REACTOME_INTERLEUKIN_18_SIGNALING")
)


#data_1 = M1T4vs.all_pos_reactome 
#data_2 = M1T4vs.all_neg_reactome
#data_3 = M2T3_pos_reactome 
#data_4 = M2T3_neg_reactome


ggvenn(a, c("M1T4vs.all_pos_reactome", "RespondervsNonreponder_pos_reactome"))            # draw two-set venn
ggvenn(a, c("M3T4vs.all_pos_reactome", "RespondervsNonreponder_pos_reactome"))
#ggvenn(a, c("Set 1", "Set 2", "Set 3"))   # draw three-set venn
ggvenn(a)   # without set names, the first 4 elements in list will be chose to draw four-set venn








M1T4pos_responderpos_intersect= as.data.frame(Reduce(intersect, list(set1,set3)))
M1T4pos_responderpos_intersect =t(M1T4pos_responderpos_intersect)
M3T4pos_responderpos_intersect= as.data.frame(Reduce(intersect, list(set2,set3)))
M3T4pos_responderpos_intersect =t(M3T4pos_responderpos_intersect)
M1T4_M3T4_responder_all_intsect = as.data.frame(Reduce(intersect, list(set1,set2,set3)))
M1T4_M3T4_responder_all_intsect =t(M1T4_M3T4_responder_all_intsect)
M3T4_responder_only <- as.data.frame(Reduce(setdiff,list(M3T4pos_responderpos_intersect, M1T4_M3T4_responder_all_intsect)))
M3T4_responder_only =t(M3T4_responder_only)
M1T4_responder_only <- as.data.frame(Reduce(setdiff,list(M1T4pos_responderpos_intersect, M1T4_M3T4_responder_all_intsect)))
M1T4_responder_only =t(M1T4_responder_only)



n<-max(length(M1T4pos_responderpos_intersect),length(M3T4pos_responderpos_intersect),length(M1T4_M3T4_responder_all_intsect),length(M3T4_responder_only),length(M1T4_responder_only))
length(M1T4pos_responderpos_intersect)<-n
length(M3T4pos_responderpos_intersect)<-n
length(M1T4_M3T4_responder_all_intsect)<-n
length(M3T4_responder_only)<-n
length(M1T4_responder_only)<-n


M1T4_M3T4_responder_intersect = as.data.frame(cbind(M1T4pos_responderpos_intersect,M3T4pos_responderpos_intersect,M1T4_M3T4_responder_all_intsect,M3T4_responder_only,M1T4_responder_only))
colnames(M1T4_M3T4_responder_intersect)

M1T4_M3T4_responder_intersect<- M1T4_M3T4_responder_intersect %>% rename("V1" = "M1T4pos_responderpos_intersect")


write.csv(M1T4_M3T4_responder_intersect, "_Tcell_MaT4_pos_M3T4_pos_responder_pos_REACTOME_intersect.csv")


########################################################################################################################################################################################################################################################################################
#######################################################################    fgsea: GOBP results comparison     ############################################################################################################################################
#######################################################################    T cell                                 ############################################################################################################################################
########################################################################################################################################################################################################################################################################################

M1T4vs.all_pos_gobp = c("GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",	"GOBP_PROTEIN_REFOLDING",	"GOBP_DE_NOVO_PROTEIN_FOLDING",	"GOBP_RESPONSE_TO_HEAT",	"GOBP_CHAPERONE_COFACTOR_DEPENDENT_PROTEIN_REFOLDING",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_ACTIVATION",
                        "GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_8_PRODUCTION",	"GOBP_PROTEIN_FOLDING",
                        "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ATP_BIOSYNTHETIC_PROCESS",	"GOBP_CELLULAR_RESPONSE_TO_HEAT",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_COMPLEX_ASSEMBLY",
                        "GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",
                        "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_ATP_METABOLIC_PROCESS",
                        "GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_INCLUSION_BODY_ASSEMBLY",
                        "GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_ICOSANOID_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_VASOCONSTRICTION",	
                        "GOBP_SKELETAL_MUSCLE_CONTRACTION",	"GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",	"GOBP_AEROBIC_RESPIRATION",
                        "GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_RESPONSE_TO_TUMOR_CELL",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",
                        "GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_RESPONSE_TO_TEMPERATURE_STIMULUS",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_STRUCTURE",	"GOBP_MICROTUBULE_BASED_PROTEIN_TRANSPORT",	
                        "GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",
                        "GOBP_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_HOMOTYPIC_CELL_CELL_ADHESION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	
                        "GOBP_IMMUNE_RESPONSE_TO_TUMOR_CELL",	"GOBP_PLATELET_AGGREGATION",	"GOBP_POSITIVE_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_BY_HOST_OF_VIRAL_TRANSCRIPTION",	"GOBP_POSITIVE_REGULATION_OF_TELOMERASE_ACTIVITY",
                        "GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_MIGRATION",	"GOBP_REGULATION_OF_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",
                        "GOBP_PYRIMIDINE_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_MONOCYTE_CHEMOTAXIS",	
                        "GOBP_REGULATION_OF_ENDOPLASMIC_RETICULUM_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_LAMELLIPODIUM_ASSEMBLY",	"GOBP_MACROPHAGE_CHEMOTAXIS",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_ACTIN_CYTOSKELETON",
                        "GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_REGULATION_OF_INCLUSION_BODY_ASSEMBLY",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_SMOOTH_MUSCLE_CONTRACTION",	"GOBP_MUSCLE_CELL_CELLULAR_HOMEOSTASIS",	"GOBP_ADHESION_OF_SYMBIONT_TO_HOST",
                        "GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_CELLULAR_RESPIRATION",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_CELLULAR_RESPONSE_TO_VASCULAR_ENDOTHELIAL_GROWTH_FACTOR_STIMULUS",
                        "GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_PODOSOME_ASSEMBLY",	"GOBP_RESPONSE_TO_SALT_STRESS",	"GOBP_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_ELECTRON_TRANSPORT_CHAIN",	"GOBP_PYROPTOSIS",
                        "GOBP_PYRIMIDINE_CONTAINING_COMPOUND_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_POSITIVE_REGULATION_OF_NUCLEOTIDE_BINDING_OLIGOMERIZATION_DOMAIN_CONTAINING_2_SIGNALING_PATHWAY",	
                        "GOBP_POSITIVE_REGULATION_OF_NUCLEOTIDE_BINDING_OLIGOMERIZATION_DOMAIN_CONTAINING_SIGNALING_PATHWAY",	"GOBP_MACROPHAGE_MIGRATION",	"GOBP_NATURAL_KILLER_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NATURAL_KILLER_CELL_DEGRANULATION",
                        "GOBP_REGULATION_OF_PROTEIN_POLYMERIZATION",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_MACROPHAGE_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_CELL_DIFFERENTIATION",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",
                        "GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_RELAXATION_OF_VASCULAR_ASSOCIATED_SMOOTH_MUSCLE",	"GOBP_ENDOTHELIAL_CELL_CHEMOTAXIS",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",
                        "GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_1_PRODUCTION",	"GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CHEMOKINE_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_ENDOPLASMIC_RETICULUM_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",
                        "GOBP_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_REGULATION_OF_RESPONSE_TO_TUMOR_CELL",	"GOBP_POSITIVE_REGULATION_OF_MICROTUBULE_NUCLEATION",	"GOBP_PLATELET_ACTIVATION",	"GOBP_REGULATION_OF_MITOCHONDRIAL_MEMBRANE_POTENTIAL",
                        "GOBP_MYELOID_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	"GOBP_SIGNAL_TRANSDUCTION_IN_ABSENCE_OF_LIGAND",	"GOBP_REGULATION_OF_LEUKOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",
                        "GOBP_REGULATION_OF_NATURAL_KILLER_CELL_CHEMOTAXIS",	"GOBP_HYPOTHALAMUS_CELL_DIFFERENTIATION",	"GOBP_HYPOTHALAMUS_GONADOTROPHIN_RELEASING_HORMONE_NEURON_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_MICROGLIAL_CELL_ACTIVATION",	"GOBP_MITOCHONDRIAL_TRANSLATION",
                        "GOBP_REGULATION_OF_NITRIC_OXIDE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_TUBE_SIZE",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_FIBROBLAST_PROLIFERATION",	"GOBP_CENTRAL_NERVOUS_SYSTEM_NEURON_DEVELOPMENT",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",
                        "GOBP_INTEGRIN_MEDIATED_SIGNALING_PATHWAY",	"GOBP_PROTEIN_LOCALIZATION_TO_MITOCHONDRION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_SYMBIONT",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_MORPHOGENESIS_OF_A_POLARIZED_EPITHELIUM",	"GOBP_CELL_KILLING",	"GOBP_INTERLEUKIN_8_PRODUCTION",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",
                        "GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_CELLULAR_RESPONSE_TO_UNFOLDED_PROTEIN",	"GOBP_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY_IN_RESPONSE_TO_ENDOPLASMIC_RETICULUM_STRESS",	"GOBP_REGULATION_OF_MUSCLE_CONTRACTION",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",	"GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_POLYMERIZATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",
                        "GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_MITOCHONDRIAL_GENE_EXPRESSION",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_INTERLEUKIN_1_PRODUCTION",	"GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_CELLULAR_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_HEMOSTASIS",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",
                        "GOBP_POSITIVE_REGULATION_OF_ION_TRANSPORT",	"GOBP_MITOCHONDRIAL_TRANSPORT",	"GOBP_RESPONSE_TO_TOXIC_SUBSTANCE",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_PATTERN_RECOGNITION_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_PROTEIN_POLYMERIZATION",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ANATOMICAL_STRUCTURE_SIZE",	"GOBP_NEGATIVE_REGULATION_OF_DNA_BINDING_TRANSCRIPTION_FACTOR_ACTIVITY",	"GOBP_TISSUE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_HYDROLASE_ACTIVITY",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_CELL_CHEMOTAXIS",
                        "GOBP_REGULATION_OF_MEMBRANE_POTENTIAL",	"GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_T_CELL_ACTIVATION",	"GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_MITOCHONDRION_ORGANIZATION",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",
                        "GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_VIRAL_PROCESS",	"GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_ORGANOPHOSPHATE_BIOSYNTHETIC_PROCESS",
                        "GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_DNA_BINDING_TRANSCRIPTION_FACTOR_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_SMALL_MOLECULE_CATABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_I_KAPPAB_KINASE_NF_KAPPAB_SIGNALING",	"GOBP_EXOCYTOSIS",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_TAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_REGULATION_OF_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_HYDROLASE_ACTIVITY",	
                        "GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_TRANSPORT",	"GOBP_REGULATION_OF_VESICLE_MEDIATED_TRANSPORT",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_CARBOHYDRATE_DERIVATIVE_BIOSYNTHETIC_PROCESS",	"GOBP_ESTABLISHMENT_OF_PROTEIN_LOCALIZATION_TO_ORGANELLE",
                        "GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",
                        "GOBP_SMALL_GTPASE_MEDIATED_SIGNAL_TRANSDUCTION",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE"
)
M3T4vs.all_pos_gobp = c("GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_POLARIZED_EPITHELIAL_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_MORPHOGENESIS_OF_A_POLARIZED_EPITHELIUM",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	
                       "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_VIRAL_GENOME_REPLICATION",	"GOBP_ESTABLISHMENT_OF_EPITHELIAL_CELL_APICAL_BASAL_POLARITY",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_ESTABLISHMENT_OR_MAINTENANCE_OF_MONOPOLAR_CELL_POLARITY",	"GOBP_RESPONSE_TO_TYPE_I_INTERFERON",	"GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_DETECTION_OF_BIOTIC_STIMULUS",	"GOBP_CELL_KILLING",	"GOBP_CYTOLYSIS",	"GOBP_DETECTION_OF_OTHER_ORGANISM",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_VIRAL_PROCESS",	"GOBP_NATURAL_KILLER_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_DETECTION_OF_EXTERNAL_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_CELL_KILLING",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	
                       "GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_NEGATIVE_REGULATION_OF_T_CELL_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_TOLL_LIKE_RECEPTOR_4_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_MIGRATION",	"GOBP_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_GLAND_MORPHOGENESIS",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_TOLERANCE_INDUCTION",	"GOBP_INTERFERON_BETA_PRODUCTION",
                       "GOBP_POSITIVE_REGULATION_OF_INTERFERON_BETA_PRODUCTION",	"GOBP_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_INTERFERON_GAMMA_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_SEMAPHORIN_PLEXIN_SIGNALING_PATHWAY",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_NEGATIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_CAMP_MEDIATED_SIGNALING",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNE_RESPONSE_TO_TUMOR_CELL",	"GOBP_EXOCYTOSIS",	"GOBP_ERK1_AND_ERK2_CASCADE",	"GOBP_ADENYLATE_CYCLASE_MODULATING_G_PROTEIN_COUPLED_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",	"GOBP_CELLULAR_RESPONSE_TO_VIRUS",	"GOBP_MAST_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_NEGATIVE_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_REGULATION_OF_CELL_KILLING",	"GOBP_CELLULAR_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_RESPONSE_TO_FIBROBLAST_GROWTH_FACTOR",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_EOSINOPHIL_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_ACTIN_FILAMENT_BUNDLE_ASSEMBLY",	"GOBP_CELL_CELL_RECOGNITION",	"GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_TISSUE_MIGRATION",
                       "GOBP_REGULATION_OF_SYNAPSE_STRUCTURE_OR_ACTIVITY",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_TOLL_LIKE_RECEPTOR_7_SIGNALING_PATHWAY",
                       "GOBP_REGULATION_OF_RESPONSE_TO_CYTOKINE_STIMULUS",	"GOBP_STRESS_FIBER_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_CLEAVAGE_FURROW_FORMATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_CADHERIN",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_INTERLEUKIN_6_PRODUCTION",	"GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_ACTIN_FILAMENT_BUNDLE_ORGANIZATION",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION_VIA_T_CELL_RECEPTOR_CONTACT_WITH_ANTIGEN_BOUND_TO_MHC_MOLECULE_ON_ANTIGEN_PRESENTING_CELL",	"GOBP_REGULATION_OF_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_T_CELL_ACTIVATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_RESPONSE_TO_MOLECULE_OF_BACTERIAL_ORIGIN",	"GOBP_LYMPHOCYTE_MIGRATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_T_CELL_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_CELL_SUBSTRATE_ADHESION",	"GOBP_ACTOMYOSIN_STRUCTURE_ORGANIZATION",
                       "GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_CYTOKINE_MEDIATED_SIGNALING_PATHWAY",	"GOBP_PROTEIN_DEPHOSPHORYLATION",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_REGULATION_OF_CELL_MORPHOGENESIS",	"GOBP_DEPHOSPHORYLATION",	"GOBP_REGULATION_OF_LEUKOCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_REGULATION_OF_ION_TRANSPORT",	"GOBP_TAXIS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_POPULATION_PROLIFERATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE_BASED_ON_SOMATIC_RECOMBINATION_OF_IMMUNE_RECEPTORS_BUILT_FROM_IMMUNOGLOBULIN_SUPERFAMILY_DOMAINS",	"GOBP_G_PROTEIN_COUPLED_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_MORPHOGENESIS_OF_AN_EPITHELIUM",	"GOBP_ACTIVATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_CELLULAR_COMPONENT_BIOGENESIS",	"GOBP_REGULATION_OF_SECRETION",	"GOBP_TRANSMEMBRANE_RECEPTOR_PROTEIN_TYROSINE_KINASE_SIGNALING_PATHWAY",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_TUBE_MORPHOGENESIS"
)

RespondervsNonreponder_pos_gobp =c("GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_CYTOCHROME_C_TO_OXYGEN",	"GOBP_RESPONSE_TO_COPPER_ION",	"GOBP_CELLULAR_RESPONSE_TO_ZINC_ION",	"GOBP_STRESS_RESPONSE_TO_METAL_ION",	
                                   "GOBP_DETOXIFICATION_OF_COPPER_ION",	"GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_RESPONSE_TO_ZINC_ION",	"GOBP_CELLULAR_RESPONSE_TO_CADMIUM_ION",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",
                                   "GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_UBIQUINOL_TO_CYTOCHROME_C",	"GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_CELLULAR_RESPONSE_TO_COPPER_ION",	"GOBP_ATP_BIOSYNTHETIC_PROCESS",
                                   "GOBP_CELL_ADHESION_MOLECULE_PRODUCTION",	"GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_REGULATION_OF_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",
                                   "GOBP_REGULATION_OF_ENDODEOXYRIBONUCLEASE_ACTIVITY",	"GOBP_AEROBIC_RESPIRATION",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",
                                   "GOBP_POSITIVE_REGULATION_OF_CYTOPLASMIC_TRANSLATION",	"GOBP_TRANSEPITHELIAL_TRANSPORT",	"GOBP_RESPONSE_TO_ELECTRICAL_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_LAMELLIPODIUM_ASSEMBLY",	"GOBP_POSITIVE_CHEMOTAXIS",
                                   "GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_FORMATION_OF_CYTOPLASMIC_TRANSLATION_INITIATION_COMPLEX",	"GOBP_REGULATION_OF_CYTOPLASMIC_TRANSLATION",
                                   "GOBP_POSITIVE_REGULATION_OF_LAMELLIPODIUM_ORGANIZATION",	"GOBP_HYDROGEN_PEROXIDE_METABOLIC_PROCESS",	"GOBP_RESPONSE_TO_CADMIUM_ION",	"GOBP_RESPONSE_TO_INTERLEUKIN_3",	"GOBP_ZINC_ION_HOMEOSTASIS",
                                   "GOBP_7_METHYLGUANOSINE_CAP_HYPERMETHYLATION",	"GOBP_INDUCTION_OF_POSITIVE_CHEMOTAXIS",	"GOBP_BICARBONATE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",
                                   "GOBP_TRANSEPITHELIAL_CHLORIDE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_FATTY_ACID_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_UNSATURATED_FATTY_ACID_BIOSYNTHETIC_PROCESS",	"GOBP_CELLULAR_RESPIRATION",
                                   "GOBP_REGULATION_OF_VESICLE_FUSION",	"GOBP_REGULATION_OF_LAMELLIPODIUM_ORGANIZATION",	"GOBP_REGULATION_OF_LAMELLIPODIUM_ASSEMBLY",	"GOBP_REGULATION_OF_NUCLEAR_TRANSCRIBED_MRNA_CATABOLIC_PROCESS_DEADENYLATION_DEPENDENT_DECAY",
                                   "GOBP_POSITIVE_REGULATION_OF_RAC_PROTEIN_SIGNAL_TRANSDUCTION",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_ELECTRON_TRANSPORT_CHAIN",
                                   "GOBP_COMPLEMENT_ACTIVATION",	"GOBP_DETOXIFICATION",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",
                                   "GOBP_U2_TYPE_PRESPLICEOSOME_ASSEMBLY",	"GOBP_REGULATION_OF_OXIDATIVE_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_MRNA_SPLICING_VIA_SPLICEOSOME",	"GOBP_LENS_FIBER_CELL_DEVELOPMENT",
                                   "GOBP_PLASMA_CELL_DIFFERENTIATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_RESPONSE_TO_MURAMYL_DIPEPTIDE",	"GOBP_CYTOPLASMIC_TRANSLATION",
                                   "GOBP_ANTIMICROBIAL_HUMORAL_RESPONSE",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_ATP_METABOLIC_PROCESS",	"GOBP_CYTOPLASMIC_TRANSLATIONAL_INITIATION",	"GOBP_APOPTOTIC_CELL_CLEARANCE",
                                   "GOBP_POSITIVE_REGULATION_OF_TRANSCRIPTION_BY_RNA_POLYMERASE_I",	"GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_REGULATION_OF_AEROBIC_RESPIRATION",
                                   "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_GLUTATHIONE_METABOLIC_PROCESS",	"GOBP_CELLULAR_TRANSITION_METAL_ION_HOMEOSTASIS",	"GOBP_MONOCYTE_DIFFERENTIATION",
                                   "GOBP_REGULATION_OF_MITOCHONDRIAL_MEMBRANE_POTENTIAL",	"GOBP_REGULATION_OF_ALTERNATIVE_MRNA_SPLICING_VIA_SPLICEOSOME",	"GOBP_REGULATION_OF_MRNA_PROCESSING",	"GOBP_RETINA_HOMEOSTASIS",
                                   "GOBP_RNA_SPLICING_VIA_TRANSESTERIFICATION_REACTIONS",	"GOBP_REGULATION_OF_RNA_SPLICING",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_LOCALIZATION_TO_CELL_PERIPHERY",
                                   "GOBP_HUMORAL_IMMUNE_RESPONSE",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_ALTERNATIVE_MRNA_SPLICING_VIA_SPLICEOSOME",	"GOBP_TRANSLATIONAL_INITIATION",
                                   "GOBP_POSITIVE_REGULATION_OF_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_HUMORAL_IMMUNE_RESPONSE_MEDIATED_BY_CIRCULATING_IMMUNOGLOBULIN",	"GOBP_RESPONSE_TO_COLD",	"GOBP_LAMELLIPODIUM_ASSEMBLY",
                                   "GOBP_NEGATIVE_REGULATION_OF_MRNA_METABOLIC_PROCESS",	"GOBP_CENTRAL_NERVOUS_SYSTEM_NEURON_DEVELOPMENT",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_EXTRINSIC_APOPTOTIC_SIGNALING_PATHWAY_VIA_DEATH_DOMAIN_RECEPTORS",	"GOBP_REGULATION_OF_T_HELPER_CELL_DIFFERENTIATION",	"GOBP_CELLULAR_RESPONSE_TO_INORGANIC_SUBSTANCE",	"GOBP_REGULATION_OF_DNA_BINDING",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",
                                   "GOBP_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	"GOBP_REGULATION_OF_OXIDATIVE_STRESS_INDUCED_CELL_DEATH",	"GOBP_RNA_SPLICING",	"GOBP_REGULATION_OF_TRANSLATIONAL_INITIATION",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	
                                   "GOBP_CELLULAR_METAL_ION_HOMEOSTASIS",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",	"GOBP_POSITIVE_REGULATION_OF_CELL_MORPHOGENESIS_INVOLVED_IN_DIFFERENTIATION",	"GOBP_RESPONSE_TO_TOXIC_SUBSTANCE",	"GOBP_FIBROBLAST_PROLIFERATION",
                                   "GOBP_TRANSITION_METAL_ION_HOMEOSTASIS",	"GOBP_RESPONSE_TO_FIBROBLAST_GROWTH_FACTOR",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_LEUKOCYTE_DIFFERENTIATION",	"GOBP_LAMELLIPODIUM_ORGANIZATION",	"GOBP_DIVALENT_INORGANIC_CATION_HOMEOSTASIS",
                                   "GOBP_NEGATIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_TRANSPORTER_ACTIVITY",	"GOBP_MRNA_PROCESSING",	"GOBP_SPLICEOSOMAL_COMPLEX_ASSEMBLY",	"GOBP_INORGANIC_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_CELLULAR_OXIDANT_DETOXIFICATION",
                                   "GOBP_POSITIVE_REGULATION_OF_TRANSLATION",	"GOBP_CELLULAR_RESPONSE_TO_TOXIC_SUBSTANCE",	"GOBP_MITOCHONDRIAL_TRANSMEMBRANE_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",
                                   "GOBP_POSITIVE_REGULATION_OF_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_INTERFERON_GAMMA_PRODUCTION",	"GOBP_TELOMERE_MAINTENANCE_VIA_TELOMERE_LENGTHENING",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",
                                   "GOBP_SEQUESTERING_OF_CALCIUM_ION",	"GOBP_METAL_ION_HOMEOSTASIS",	"GOBP_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_REGULATION_OF_CYSTEINE_TYPE_ENDOPEPTIDASE_ACTIVITY",	"GOBP_VIRAL_GENE_EXPRESSION",
                                   "GOBP_REGULATION_OF_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_RIBONUCLEOPROTEIN_COMPLEX_SUBUNIT_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_ENTRY_INTO_HOST",	"GOBP_PROTEIN_KINASE_B_SIGNALING",
                                   "GOBP_REGULATION_OF_NIK_NF_KAPPAB_SIGNALING",	"GOBP_REGULATION_OF_MRNA_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_PLASMA_MEMBRANE_BOUNDED_CELL_PROJECTION_ASSEMBLY",	"GOBP_CELL_RECOGNITION",	"GOBP_RESPONSE_TO_METAL_ION",
                                   "GOBP_REGULATION_OF_PROTEIN_LOCALIZATION_TO_CELL_PERIPHERY",	"GOBP_CATION_TRANSMEMBRANE_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_NEURON_APOPTOTIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELLULAR_AMIDE_METABOLIC_PROCESS",	"GOBP_CELLULAR_ION_HOMEOSTASIS",
                                   "GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_CALCIUM_ION_TRANSPORT",	"GOBP_PROTEIN_FOLDING",	"GOBP_POSITIVE_REGULATION_OF_CYSTEINE_TYPE_ENDOPEPTIDASE_ACTIVITY",	"GOBP_REGULATION_OF_METAL_ION_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_GROWTH",
                                   "GOBP_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_RESPONSE_TO_TEMPERATURE_STIMULUS",	"GOBP_VIRAL_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_ERK1_AND_ERK2_CASCADE",	"GOBP_CATION_TRANSPORT",
                                   "GOBP_CELLULAR_RESPONSE_TO_UNFOLDED_PROTEIN",	"GOBP_REGULATION_OF_EXTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_POSITIVE_REGULATION_OF_ION_TRANSPORT",	"GOBP_ION_TRANSMEMBRANE_TRANSPORT",
                                   "GOBP_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_NIK_NF_KAPPAB_SIGNALING",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",	"GOBP_RESPONSE_TO_INORGANIC_SUBSTANCE",	"GOBP_REGULATION_OF_ION_TRANSPORT",	"GOBP_REGULATION_OF_TRANSPORTER_ACTIVITY",
                                   "GOBP_NEGATIVE_REGULATION_OF_PROTEOLYSIS",	"GOBP_PROTEIN_STABILIZATION",	"GOBP_CALCIUM_ION_HOMEOSTASIS",	"GOBP_POSITIVE_REGULATION_OF_I_KAPPAB_KINASE_NF_KAPPAB_SIGNALING",	"GOBP_LEUKOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_HEMOPOIESIS",	"GOBP_ION_HOMEOSTASIS",	"GOBP_NEURON_APOPTOTIC_PROCESS",	"GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_CELLULAR_HOMEOSTASIS",	"GOBP_MAINTENANCE_OF_LOCATION_IN_CELL",	"GOBP_PEPTIDE_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_CATION_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_PROTEOLYSIS",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_TRANSPORT",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_REGULATION_OF_CELL_MORPHOGENESIS",	"GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_REGULATION_OF_PROTEIN_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_PROTEOLYSIS",	"GOBP_REGULATION_OF_MEMBRANE_POTENTIAL",	"GOBP_RESPONSE_TO_ENDOPLASMIC_RETICULUM_STRESS",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_MITOCHONDRION_ORGANIZATION",	"GOBP_REGULATION_OF_PROTEIN_STABILITY",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_I_KAPPAB_KINASE_NF_KAPPAB_SIGNALING",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_REGULATION_OF_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_RIBONUCLEOPROTEIN_COMPLEX_BIOGENESIS",	"GOBP_MAINTENANCE_OF_LOCATION",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_CELLULAR_AMIDE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_REGULATION_OF_BINDING",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_PROTEIN_MODIFICATION_BY_SMALL_PROTEIN_CONJUGATION_OR_REMOVAL",	"GOBP_METAL_ION_TRANSPORT",	"GOBP_RESPONSE_TO_WOUNDING",	"GOBP_REGULATION_OF_CELLULAR_MACROMOLECULE_BIOSYNTHETIC_PROCESS",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_CELL_ADHESION",	"GOBP_RESPONSE_TO_BACTERIUM",	"GOBP_POSITIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_PROTEASOMAL_PROTEIN_CATABOLIC_PROCESS",	"GOBP_POST_TRANSCRIPTIONAL_REGULATION_OF_GENE_EXPRESSION",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_EPITHELIAL_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_TRANSPORT",	"GOBP_RESPONSE_TO_GROWTH_FACTOR",	"GOBP_POSITIVE_REGULATION_OF_LOCOMOTION",	"GOBP_T_CELL_ACTIVATION",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_MONONUCLEAR_CELL_DIFFERENTIATION",	"GOBP_NEGATIVE_REGULATION_OF_DEVELOPMENTAL_PROCESS",	"GOBP_LEUKOCYTE_DIFFERENTIATION",	"GOBP_CELL_CELL_ADHESION",	"GOBP_NEGATIVE_REGULATION_OF_MULTICELLULAR_ORGANISMAL_PROCESS"
)
#RespondervsNonreponder_neg_reactome = c()



data_1 = M1T4vs.all_pos_gobp 
data_2 = M3T4vs.all_pos_gobp
data_3 = RespondervsNonreponder_pos_gobp 
#data_4 = RespondervsNonreponder_neg_reactome


set1 = as.list(data_1)
set2 = as.list(data_2)
set3 = as.list(data_3)
#set4 = as.list(data_4)

venn.diagram(x=list(set1, set2, set3), 
             category.names = c("MaT4","McT4","R"), filename = "venndiagram_Tcell_GOBP_MaT4_McT4_Responder.png", output=T, 
             imagetype = "png", height = 2000, width = 2000, resolution = 300, compression = "lzw",
             fill=c("skyblue", "yellowgreen","pink"), 
             cex = 2,  lwd=2, lty='blank', fontface="bold", fontfamily="sans", cat.cex=1, cat.fontface="bold", cat.default.pos="outer", cat.pos=c(0, 0, 200),cat.fontfamily="sans")




a <- list("M1T4vs.all_pos_gobp" = c("GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",	"GOBP_PROTEIN_REFOLDING",	"GOBP_DE_NOVO_PROTEIN_FOLDING",	"GOBP_RESPONSE_TO_HEAT",	"GOBP_CHAPERONE_COFACTOR_DEPENDENT_PROTEIN_REFOLDING",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_ACTIVATION",
                                    "GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_8_PRODUCTION",	"GOBP_PROTEIN_FOLDING",
                                    "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ATP_BIOSYNTHETIC_PROCESS",	"GOBP_CELLULAR_RESPONSE_TO_HEAT",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_COMPLEX_ASSEMBLY",
                                    "GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",
                                    "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_ATP_METABOLIC_PROCESS",
                                    "GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_INCLUSION_BODY_ASSEMBLY",
                                    "GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_ICOSANOID_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_VASOCONSTRICTION",	
                                    "GOBP_SKELETAL_MUSCLE_CONTRACTION",	"GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",	"GOBP_AEROBIC_RESPIRATION",
                                    "GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_RESPONSE_TO_TUMOR_CELL",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",
                                    "GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_RESPONSE_TO_TEMPERATURE_STIMULUS",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_STRUCTURE",	"GOBP_MICROTUBULE_BASED_PROTEIN_TRANSPORT",	
                                    "GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",
                                    "GOBP_REGULATION_OF_ERYTHROCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_HOMOTYPIC_CELL_CELL_ADHESION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	
                                    "GOBP_IMMUNE_RESPONSE_TO_TUMOR_CELL",	"GOBP_PLATELET_AGGREGATION",	"GOBP_POSITIVE_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_BY_HOST_OF_VIRAL_TRANSCRIPTION",	"GOBP_POSITIVE_REGULATION_OF_TELOMERASE_ACTIVITY",
                                    "GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_MIGRATION",	"GOBP_REGULATION_OF_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_REGULATION_OF_MONONUCLEAR_CELL_MIGRATION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",
                                    "GOBP_PYRIMIDINE_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_MONOCYTE_CHEMOTAXIS",	
                                    "GOBP_REGULATION_OF_ENDOPLASMIC_RETICULUM_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_LAMELLIPODIUM_ASSEMBLY",	"GOBP_MACROPHAGE_CHEMOTAXIS",	"GOBP_MODIFICATION_OF_POSTSYNAPTIC_ACTIN_CYTOSKELETON",
                                    "GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_REGULATION_OF_INCLUSION_BODY_ASSEMBLY",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_SMOOTH_MUSCLE_CONTRACTION",	"GOBP_MUSCLE_CELL_CELLULAR_HOMEOSTASIS",	"GOBP_ADHESION_OF_SYMBIONT_TO_HOST",
                                    "GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_CELLULAR_RESPIRATION",	"GOBP_CELLULAR_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_CELLULAR_RESPONSE_TO_VASCULAR_ENDOTHELIAL_GROWTH_FACTOR_STIMULUS",
                                    "GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_PODOSOME_ASSEMBLY",	"GOBP_RESPONSE_TO_SALT_STRESS",	"GOBP_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_ELECTRON_TRANSPORT_CHAIN",	"GOBP_PYROPTOSIS",
                                    "GOBP_PYRIMIDINE_CONTAINING_COMPOUND_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_POSITIVE_REGULATION_OF_NUCLEOTIDE_BINDING_OLIGOMERIZATION_DOMAIN_CONTAINING_2_SIGNALING_PATHWAY",	
                                    "GOBP_POSITIVE_REGULATION_OF_NUCLEOTIDE_BINDING_OLIGOMERIZATION_DOMAIN_CONTAINING_SIGNALING_PATHWAY",	"GOBP_MACROPHAGE_MIGRATION",	"GOBP_NATURAL_KILLER_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NATURAL_KILLER_CELL_DEGRANULATION",
                                    "GOBP_REGULATION_OF_PROTEIN_POLYMERIZATION",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_MACROPHAGE_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_CELL_DIFFERENTIATION",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",
                                    "GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_RELAXATION_OF_VASCULAR_ASSOCIATED_SMOOTH_MUSCLE",	"GOBP_ENDOTHELIAL_CELL_CHEMOTAXIS",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",
                                    "GOBP_POSITIVE_REGULATION_OF_INTERLEUKIN_1_PRODUCTION",	"GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CHEMOKINE_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_ENDOPLASMIC_RETICULUM_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",
                                    "GOBP_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_REGULATION_OF_RESPONSE_TO_TUMOR_CELL",	"GOBP_POSITIVE_REGULATION_OF_MICROTUBULE_NUCLEATION",	"GOBP_PLATELET_ACTIVATION",	"GOBP_REGULATION_OF_MITOCHONDRIAL_MEMBRANE_POTENTIAL",
                                    "GOBP_MYELOID_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	"GOBP_SIGNAL_TRANSDUCTION_IN_ABSENCE_OF_LIGAND",	"GOBP_REGULATION_OF_LEUKOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",
                                    "GOBP_REGULATION_OF_NATURAL_KILLER_CELL_CHEMOTAXIS",	"GOBP_HYPOTHALAMUS_CELL_DIFFERENTIATION",	"GOBP_HYPOTHALAMUS_GONADOTROPHIN_RELEASING_HORMONE_NEURON_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_MICROGLIAL_CELL_ACTIVATION",	"GOBP_MITOCHONDRIAL_TRANSLATION",
                                    "GOBP_REGULATION_OF_NITRIC_OXIDE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_TUBE_SIZE",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_NUCLEOSIDE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_FIBROBLAST_PROLIFERATION",	"GOBP_CENTRAL_NERVOUS_SYSTEM_NEURON_DEVELOPMENT",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",
                                    "GOBP_INTEGRIN_MEDIATED_SIGNALING_PATHWAY",	"GOBP_PROTEIN_LOCALIZATION_TO_MITOCHONDRION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_SYMBIONT",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_MORPHOGENESIS_OF_A_POLARIZED_EPITHELIUM",	"GOBP_CELL_KILLING",	"GOBP_INTERLEUKIN_8_PRODUCTION",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",
                                    "GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_CELLULAR_RESPONSE_TO_UNFOLDED_PROTEIN",	"GOBP_TUMOR_NECROSIS_FACTOR_MEDIATED_SIGNALING_PATHWAY",	"GOBP_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY_IN_RESPONSE_TO_ENDOPLASMIC_RETICULUM_STRESS",	"GOBP_REGULATION_OF_MUSCLE_CONTRACTION",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",	"GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_POLYMERIZATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",
                                    "GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_MITOCHONDRIAL_GENE_EXPRESSION",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_INTERLEUKIN_1_PRODUCTION",	"GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_CELLULAR_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_HEMOSTASIS",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",
                                    "GOBP_POSITIVE_REGULATION_OF_ION_TRANSPORT",	"GOBP_MITOCHONDRIAL_TRANSPORT",	"GOBP_RESPONSE_TO_TOXIC_SUBSTANCE",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_PATTERN_RECOGNITION_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_PROTEIN_POLYMERIZATION",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ANATOMICAL_STRUCTURE_SIZE",	"GOBP_NEGATIVE_REGULATION_OF_DNA_BINDING_TRANSCRIPTION_FACTOR_ACTIVITY",	"GOBP_TISSUE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_HYDROLASE_ACTIVITY",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_CELL_CHEMOTAXIS",
                                    "GOBP_REGULATION_OF_MEMBRANE_POTENTIAL",	"GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_T_CELL_ACTIVATION",	"GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_MITOCHONDRION_ORGANIZATION",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",
                                    "GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_VIRAL_PROCESS",	"GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_ORGANOPHOSPHATE_BIOSYNTHETIC_PROCESS",
                                    "GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_DNA_BINDING_TRANSCRIPTION_FACTOR_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_SMALL_MOLECULE_CATABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_I_KAPPAB_KINASE_NF_KAPPAB_SIGNALING",	"GOBP_EXOCYTOSIS",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_TAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_REGULATION_OF_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_HYDROLASE_ACTIVITY",	
                                    "GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_TRANSPORT",	"GOBP_REGULATION_OF_VESICLE_MEDIATED_TRANSPORT",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_CARBOHYDRATE_DERIVATIVE_BIOSYNTHETIC_PROCESS",	"GOBP_ESTABLISHMENT_OF_PROTEIN_LOCALIZATION_TO_ORGANELLE",
                                    "GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",
                                    "GOBP_SMALL_GTPASE_MEDIATED_SIGNAL_TRANSDUCTION",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE"
),
"M3T4vs.all_pos_gobp" = c("GOBP_NEGATIVE_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_ANTIGEN",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	"GOBP_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_POLARIZED_EPITHELIAL_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_VIRAL_GENOME_REPLICATION",	"GOBP_DEFENSE_RESPONSE_TO_SYMBIONT",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_MORPHOGENESIS_OF_A_POLARIZED_EPITHELIUM",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	
                          "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_RESPONSE_TO_INTERFERON_GAMMA",	"GOBP_VIRAL_GENOME_REPLICATION",	"GOBP_ESTABLISHMENT_OF_EPITHELIAL_CELL_APICAL_BASAL_POLARITY",	"GOBP_RESPONSE_TO_VIRUS",	"GOBP_ESTABLISHMENT_OR_MAINTENANCE_OF_MONOPOLAR_CELL_POLARITY",	"GOBP_RESPONSE_TO_TYPE_I_INTERFERON",	"GOBP_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_LYMPHOCYTE_CHEMOTAXIS",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_CLASS_II_PROTEIN_COMPLEX",	"GOBP_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY",	"GOBP_DETECTION_OF_BIOTIC_STIMULUS",	"GOBP_CELL_KILLING",	"GOBP_CYTOLYSIS",	"GOBP_DETECTION_OF_OTHER_ORGANISM",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_RESPONSE_TO_INTERFERON_BETA",	"GOBP_VIRAL_PROCESS",	"GOBP_NATURAL_KILLER_CELL_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_DETECTION_OF_EXTERNAL_BIOTIC_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_CELL_KILLING",	"GOBP_POSITIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	
                          "GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_NEGATIVE_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_LEUKOCYTE_DEGRANULATION",	"GOBP_NEGATIVE_REGULATION_OF_T_CELL_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_TOLL_LIKE_RECEPTOR_4_SIGNALING_PATHWAY",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_REGULATION_OF_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_INTERFERON_GAMMA_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_CYTOKINE_PRODUCTION",	"GOBP_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_POSITIVE_REGULATION_OF_MACROPHAGE_MIGRATION",	"GOBP_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_NEGATIVE_REGULATION_OF_IMMUNE_SYSTEM_PROCESS",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_ENDOGENOUS_ANTIGEN",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_DENDRITIC_CELL_DIFFERENTIATION",	"GOBP_GLAND_MORPHOGENESIS",	"GOBP_POSITIVE_REGULATION_OF_LYMPHOCYTE_MIGRATION",	"GOBP_TOLERANCE_INDUCTION",	"GOBP_INTERFERON_BETA_PRODUCTION",
                          "GOBP_POSITIVE_REGULATION_OF_INTERFERON_BETA_PRODUCTION",	"GOBP_REGULATION_OF_CELL_ADHESION_MEDIATED_BY_INTEGRIN",	"GOBP_POSITIVE_REGULATION_OF_TYPE_I_INTERFERON_PRODUCTION",	"GOBP_INTERFERON_GAMMA_PRODUCTION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_SEMAPHORIN_PLEXIN_SIGNALING_PATHWAY",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_TUMOR_NECROSIS_FACTOR_SUPERFAMILY_CYTOKINE_PRODUCTION",	"GOBP_REGULATED_EXOCYTOSIS",	"GOBP_NEGATIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_NEGATIVE_REGULATION_OF_INNATE_IMMUNE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_CAMP_MEDIATED_SIGNALING",	"GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNE_RESPONSE_TO_TUMOR_CELL",	"GOBP_EXOCYTOSIS",	"GOBP_ERK1_AND_ERK2_CASCADE",	"GOBP_ADENYLATE_CYCLASE_MODULATING_G_PROTEIN_COUPLED_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_MYELOID_LEUKOCYTE_ACTIVATION",	"GOBP_CELLULAR_RESPONSE_TO_VIRUS",	"GOBP_MAST_CELL_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE_TO_VIRUS_BY_HOST",	"GOBP_NEGATIVE_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_REGULATION_OF_CELL_KILLING",	"GOBP_CELLULAR_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_RESPONSE_TO_FIBROBLAST_GROWTH_FACTOR",	"GOBP_ACTIN_POLYMERIZATION_OR_DEPOLYMERIZATION",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_EOSINOPHIL_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_ACTIN_FILAMENT_BUNDLE_ASSEMBLY",	"GOBP_CELL_CELL_RECOGNITION",	"GOBP_IMMUNE_EFFECTOR_PROCESS",	"GOBP_TISSUE_MIGRATION",
                          "GOBP_REGULATION_OF_SYNAPSE_STRUCTURE_OR_ACTIVITY",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_LENGTH",	"GOBP_TOLL_LIKE_RECEPTOR_7_SIGNALING_PATHWAY",
                          "GOBP_REGULATION_OF_RESPONSE_TO_CYTOKINE_STIMULUS",	"GOBP_STRESS_FIBER_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_CYTOTOXICITY",	"GOBP_CLEAVAGE_FURROW_FORMATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION_MEDIATED_BY_CADHERIN",	"GOBP_REGULATION_OF_IMMUNE_EFFECTOR_PROCESS",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_BIOTIC_STIMULUS",	"GOBP_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_NEGATIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_INTERLEUKIN_6_PRODUCTION",	"GOBP_REGULATION_OF_EPITHELIAL_CELL_MIGRATION",	"GOBP_ACTIN_FILAMENT_BUNDLE_ORGANIZATION",	"GOBP_POSITIVE_REGULATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION_VIA_T_CELL_RECEPTOR_CONTACT_WITH_ANTIGEN_BOUND_TO_MHC_MOLECULE_ON_ANTIGEN_PRESENTING_CELL",	"GOBP_REGULATION_OF_LYMPHOCYTE_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_LEUKOCYTE_MEDIATED_IMMUNITY",	"GOBP_T_CELL_ACTIVATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE",	"GOBP_MONONUCLEAR_CELL_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_LOCOMOTION",	"GOBP_RESPONSE_TO_MOLECULE_OF_BACTERIAL_ORIGIN",	"GOBP_LYMPHOCYTE_MIGRATION",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_PRODUCTION_OF_MOLECULAR_MEDIATOR_OF_IMMUNE_RESPONSE",	"GOBP_T_CELL_MEDIATED_IMMUNITY",	"GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_POSITIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_CELL_SUBSTRATE_ADHESION",	"GOBP_ACTOMYOSIN_STRUCTURE_ORGANIZATION",
                          "GOBP_REGULATION_OF_LEUKOCYTE_PROLIFERATION",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_CYTOKINE_MEDIATED_SIGNALING_PATHWAY",	"GOBP_PROTEIN_DEPHOSPHORYLATION",	"GOBP_NEGATIVE_REGULATION_OF_RESPONSE_TO_EXTERNAL_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_DEFENSE_RESPONSE",	"GOBP_IMMUNE_RESPONSE_REGULATING_SIGNALING_PATHWAY",	"GOBP_LEUKOCYTE_MIGRATION",	"GOBP_REGULATION_OF_CELL_MORPHOGENESIS",	"GOBP_DEPHOSPHORYLATION",	"GOBP_REGULATION_OF_LEUKOCYTE_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_REGULATION_OF_ION_TRANSPORT",	"GOBP_TAXIS",	"GOBP_NEGATIVE_REGULATION_OF_CELL_POPULATION_PROLIFERATION",	"GOBP_ADAPTIVE_IMMUNE_RESPONSE_BASED_ON_SOMATIC_RECOMBINATION_OF_IMMUNE_RECEPTORS_BUILT_FROM_IMMUNOGLOBULIN_SUPERFAMILY_DOMAINS",	"GOBP_G_PROTEIN_COUPLED_RECEPTOR_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_CYTOSKELETON_ORGANIZATION",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_MORPHOGENESIS_OF_AN_EPITHELIUM",	"GOBP_ACTIVATION_OF_IMMUNE_RESPONSE",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_REGULATION_OF_PROTEIN_CONTAINING_COMPLEX_ASSEMBLY",	"GOBP_POSITIVE_REGULATION_OF_CELLULAR_COMPONENT_BIOGENESIS",	"GOBP_REGULATION_OF_SECRETION",	"GOBP_TRANSMEMBRANE_RECEPTOR_PROTEIN_TYROSINE_KINASE_SIGNALING_PATHWAY",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_TUBE_MORPHOGENESIS"
),
"RespondervsNonreponder_pos_gobp" = c("GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_CYTOCHROME_C_TO_OXYGEN",	"GOBP_RESPONSE_TO_COPPER_ION",	"GOBP_CELLULAR_RESPONSE_TO_ZINC_ION",	"GOBP_STRESS_RESPONSE_TO_METAL_ION",	
                                      "GOBP_DETOXIFICATION_OF_COPPER_ION",	"GOBP_OXIDATIVE_PHOSPHORYLATION",	"GOBP_RESPONSE_TO_ZINC_ION",	"GOBP_CELLULAR_RESPONSE_TO_CADMIUM_ION",	"GOBP_ATP_SYNTHESIS_COUPLED_ELECTRON_TRANSPORT",
                                      "GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_UBIQUINOL_TO_CYTOCHROME_C",	"GOBP_PROTON_MOTIVE_FORCE_DRIVEN_ATP_SYNTHESIS",	"GOBP_CELLULAR_RESPONSE_TO_COPPER_ION",	"GOBP_ATP_BIOSYNTHETIC_PROCESS",
                                      "GOBP_CELL_ADHESION_MOLECULE_PRODUCTION",	"GOBP_RESPIRATORY_ELECTRON_TRANSPORT_CHAIN",	"GOBP_REGULATION_OF_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_PROTON_TRANSMEMBRANE_TRANSPORT",
                                      "GOBP_REGULATION_OF_ENDODEOXYRIBONUCLEASE_ACTIVITY",	"GOBP_AEROBIC_RESPIRATION",	"GOBP_REGULATION_OF_VIRAL_ENTRY_INTO_HOST_CELL",	"GOBP_MITOCHONDRIAL_ELECTRON_TRANSPORT_NADH_TO_UBIQUINONE",
                                      "GOBP_POSITIVE_REGULATION_OF_CYTOPLASMIC_TRANSLATION",	"GOBP_TRANSEPITHELIAL_TRANSPORT",	"GOBP_RESPONSE_TO_ELECTRICAL_STIMULUS",	"GOBP_POSITIVE_REGULATION_OF_LAMELLIPODIUM_ASSEMBLY",	"GOBP_POSITIVE_CHEMOTAXIS",
                                      "GOBP_NUCLEOSIDE_TRIPHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_FORMATION_OF_CYTOPLASMIC_TRANSLATION_INITIATION_COMPLEX",	"GOBP_REGULATION_OF_CYTOPLASMIC_TRANSLATION",
                                      "GOBP_POSITIVE_REGULATION_OF_LAMELLIPODIUM_ORGANIZATION",	"GOBP_HYDROGEN_PEROXIDE_METABOLIC_PROCESS",	"GOBP_RESPONSE_TO_CADMIUM_ION",	"GOBP_RESPONSE_TO_INTERLEUKIN_3",	"GOBP_ZINC_ION_HOMEOSTASIS",
                                      "GOBP_7_METHYLGUANOSINE_CAP_HYPERMETHYLATION",	"GOBP_INDUCTION_OF_POSITIVE_CHEMOTAXIS",	"GOBP_BICARBONATE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",
                                      "GOBP_TRANSEPITHELIAL_CHLORIDE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_FATTY_ACID_BIOSYNTHETIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_UNSATURATED_FATTY_ACID_BIOSYNTHETIC_PROCESS",	"GOBP_CELLULAR_RESPIRATION",
                                      "GOBP_REGULATION_OF_VESICLE_FUSION",	"GOBP_REGULATION_OF_LAMELLIPODIUM_ORGANIZATION",	"GOBP_REGULATION_OF_LAMELLIPODIUM_ASSEMBLY",	"GOBP_REGULATION_OF_NUCLEAR_TRANSCRIBED_MRNA_CATABOLIC_PROCESS_DEADENYLATION_DEPENDENT_DECAY",
                                      "GOBP_POSITIVE_REGULATION_OF_RAC_PROTEIN_SIGNAL_TRANSDUCTION",	"GOBP_NADH_DEHYDROGENASE_COMPLEX_ASSEMBLY",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_LIFE_CYCLE",	"GOBP_ELECTRON_TRANSPORT_CHAIN",
                                      "GOBP_COMPLEMENT_ACTIVATION",	"GOBP_DETOXIFICATION",	"GOBP_PEPTIDE_ANTIGEN_ASSEMBLY_WITH_MHC_PROTEIN_COMPLEX",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN_VIA_MHC_CLASS_II",
                                      "GOBP_U2_TYPE_PRESPLICEOSOME_ASSEMBLY",	"GOBP_REGULATION_OF_OXIDATIVE_STRESS_INDUCED_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_REGULATION_OF_MRNA_SPLICING_VIA_SPLICEOSOME",	"GOBP_LENS_FIBER_CELL_DEVELOPMENT",
                                      "GOBP_PLASMA_CELL_DIFFERENTIATION",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_PEPTIDE_OR_POLYSACCHARIDE_ANTIGEN_VIA_MHC_CLASS_II",	"GOBP_RESPONSE_TO_MURAMYL_DIPEPTIDE",	"GOBP_CYTOPLASMIC_TRANSLATION",
                                      "GOBP_ANTIMICROBIAL_HUMORAL_RESPONSE",	"GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_PEPTIDE_ANTIGEN",	"GOBP_ATP_METABOLIC_PROCESS",	"GOBP_CYTOPLASMIC_TRANSLATIONAL_INITIATION",	"GOBP_APOPTOTIC_CELL_CLEARANCE",
                                      "GOBP_POSITIVE_REGULATION_OF_TRANSCRIPTION_BY_RNA_POLYMERASE_I",	"GOBP_ENERGY_DERIVATION_BY_OXIDATION_OF_ORGANIC_COMPOUNDS",	"GOBP_REGULATION_OF_AEROBIC_RESPIRATION",
                                      "GOBP_ANTIGEN_PROCESSING_AND_PRESENTATION_OF_EXOGENOUS_ANTIGEN",	"GOBP_GLUTATHIONE_METABOLIC_PROCESS",	"GOBP_CELLULAR_TRANSITION_METAL_ION_HOMEOSTASIS",	"GOBP_MONOCYTE_DIFFERENTIATION",
                                      "GOBP_REGULATION_OF_MITOCHONDRIAL_MEMBRANE_POTENTIAL",	"GOBP_REGULATION_OF_ALTERNATIVE_MRNA_SPLICING_VIA_SPLICEOSOME",	"GOBP_REGULATION_OF_MRNA_PROCESSING",	"GOBP_RETINA_HOMEOSTASIS",
                                      "GOBP_RNA_SPLICING_VIA_TRANSESTERIFICATION_REACTIONS",	"GOBP_REGULATION_OF_RNA_SPLICING",	"GOBP_NUCLEOSIDE_TRIPHOSPHATE_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_PROTEIN_LOCALIZATION_TO_CELL_PERIPHERY",
                                      "GOBP_HUMORAL_IMMUNE_RESPONSE",	"GOBP_MITOCHONDRIAL_RESPIRATORY_CHAIN_COMPLEX_ASSEMBLY",	"GOBP_ALTERNATIVE_MRNA_SPLICING_VIA_SPLICEOSOME",	"GOBP_TRANSLATIONAL_INITIATION",
                                      "GOBP_POSITIVE_REGULATION_OF_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_HUMORAL_IMMUNE_RESPONSE_MEDIATED_BY_CIRCULATING_IMMUNOGLOBULIN",	"GOBP_RESPONSE_TO_COLD",	"GOBP_LAMELLIPODIUM_ASSEMBLY",
                                      "GOBP_NEGATIVE_REGULATION_OF_MRNA_METABOLIC_PROCESS",	"GOBP_CENTRAL_NERVOUS_SYSTEM_NEURON_DEVELOPMENT",	"GOBP_PURINE_CONTAINING_COMPOUND_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_EXTRINSIC_APOPTOTIC_SIGNALING_PATHWAY_VIA_DEATH_DOMAIN_RECEPTORS",	"GOBP_REGULATION_OF_T_HELPER_CELL_DIFFERENTIATION",	"GOBP_CELLULAR_RESPONSE_TO_INORGANIC_SUBSTANCE",	"GOBP_REGULATION_OF_DNA_BINDING",	"GOBP_RIBOSE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",
                                      "GOBP_REGULATION_OF_SUBSTRATE_ADHESION_DEPENDENT_CELL_SPREADING",	"GOBP_REGULATION_OF_OXIDATIVE_STRESS_INDUCED_CELL_DEATH",	"GOBP_RNA_SPLICING",	"GOBP_REGULATION_OF_TRANSLATIONAL_INITIATION",	"GOBP_REGULATION_OF_VIRAL_PROCESS",	
                                      "GOBP_CELLULAR_METAL_ION_HOMEOSTASIS",	"GOBP_GENERATION_OF_PRECURSOR_METABOLITES_AND_ENERGY",	"GOBP_POSITIVE_REGULATION_OF_CELL_MORPHOGENESIS_INVOLVED_IN_DIFFERENTIATION",	"GOBP_RESPONSE_TO_TOXIC_SUBSTANCE",	"GOBP_FIBROBLAST_PROLIFERATION",
                                      "GOBP_TRANSITION_METAL_ION_HOMEOSTASIS",	"GOBP_RESPONSE_TO_FIBROBLAST_GROWTH_FACTOR",	"GOBP_POSITIVE_REGULATION_OF_MYELOID_LEUKOCYTE_DIFFERENTIATION",	"GOBP_LAMELLIPODIUM_ORGANIZATION",	"GOBP_DIVALENT_INORGANIC_CATION_HOMEOSTASIS",
                                      "GOBP_NEGATIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_POSITIVE_REGULATION_OF_TRANSPORTER_ACTIVITY",	"GOBP_MRNA_PROCESSING",	"GOBP_SPLICEOSOMAL_COMPLEX_ASSEMBLY",	"GOBP_INORGANIC_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_CELLULAR_OXIDANT_DETOXIFICATION",
                                      "GOBP_POSITIVE_REGULATION_OF_TRANSLATION",	"GOBP_CELLULAR_RESPONSE_TO_TOXIC_SUBSTANCE",	"GOBP_MITOCHONDRIAL_TRANSMEMBRANE_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_CHAPERONE_MEDIATED_PROTEIN_FOLDING",
                                      "GOBP_POSITIVE_REGULATION_OF_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_NUCLEOSIDE_PHOSPHATE_BIOSYNTHETIC_PROCESS",	"GOBP_INTERFERON_GAMMA_PRODUCTION",	"GOBP_TELOMERE_MAINTENANCE_VIA_TELOMERE_LENGTHENING",	"GOBP_REGULATION_OF_PEPTIDASE_ACTIVITY",
                                      "GOBP_SEQUESTERING_OF_CALCIUM_ION",	"GOBP_METAL_ION_HOMEOSTASIS",	"GOBP_VIRAL_GENOME_REPLICATION",	"GOBP_NEGATIVE_REGULATION_OF_VIRAL_PROCESS",	"GOBP_REGULATION_OF_CYSTEINE_TYPE_ENDOPEPTIDASE_ACTIVITY",	"GOBP_VIRAL_GENE_EXPRESSION",
                                      "GOBP_REGULATION_OF_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_RIBONUCLEOPROTEIN_COMPLEX_SUBUNIT_ORGANIZATION",	"GOBP_NEGATIVE_REGULATION_OF_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_ENTRY_INTO_HOST",	"GOBP_PROTEIN_KINASE_B_SIGNALING",
                                      "GOBP_REGULATION_OF_NIK_NF_KAPPAB_SIGNALING",	"GOBP_REGULATION_OF_MRNA_METABOLIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_PLASMA_MEMBRANE_BOUNDED_CELL_PROJECTION_ASSEMBLY",	"GOBP_CELL_RECOGNITION",	"GOBP_RESPONSE_TO_METAL_ION",
                                      "GOBP_REGULATION_OF_PROTEIN_LOCALIZATION_TO_CELL_PERIPHERY",	"GOBP_CATION_TRANSMEMBRANE_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_NEURON_APOPTOTIC_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_CELLULAR_AMIDE_METABOLIC_PROCESS",	"GOBP_CELLULAR_ION_HOMEOSTASIS",
                                      "GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_CALCIUM_ION_TRANSPORT",	"GOBP_PROTEIN_FOLDING",	"GOBP_POSITIVE_REGULATION_OF_CYSTEINE_TYPE_ENDOPEPTIDASE_ACTIVITY",	"GOBP_REGULATION_OF_METAL_ION_TRANSPORT",	"GOBP_NEGATIVE_REGULATION_OF_GROWTH",
                                      "GOBP_REGULATION_OF_CALCIUM_ION_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_PEPTIDASE_ACTIVITY",	"GOBP_VIRAL_LIFE_CYCLE",	"GOBP_RESPONSE_TO_TEMPERATURE_STIMULUS",	"GOBP_VIRAL_PROCESS",	"GOBP_POSITIVE_REGULATION_OF_ERK1_AND_ERK2_CASCADE",	"GOBP_CATION_TRANSPORT",
                                      "GOBP_CELLULAR_RESPONSE_TO_UNFOLDED_PROTEIN",	"GOBP_REGULATION_OF_EXTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_SYMBIOTIC_INTERACTION",	"GOBP_POSITIVE_REGULATION_OF_ION_TRANSPORT",	"GOBP_ION_TRANSMEMBRANE_TRANSPORT",
                                      "GOBP_INTRINSIC_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_NIK_NF_KAPPAB_SIGNALING",	"GOBP_BIOLOGICAL_PROCESS_INVOLVED_IN_INTERACTION_WITH_HOST",	"GOBP_RESPONSE_TO_INORGANIC_SUBSTANCE",	"GOBP_REGULATION_OF_ION_TRANSPORT",	"GOBP_REGULATION_OF_TRANSPORTER_ACTIVITY",
                                      "GOBP_NEGATIVE_REGULATION_OF_PROTEOLYSIS",	"GOBP_PROTEIN_STABILIZATION",	"GOBP_CALCIUM_ION_HOMEOSTASIS",	"GOBP_POSITIVE_REGULATION_OF_I_KAPPAB_KINASE_NF_KAPPAB_SIGNALING",	"GOBP_LEUKOCYTE_CHEMOTAXIS",	"GOBP_POSITIVE_REGULATION_OF_HEMOPOIESIS",	"GOBP_ION_HOMEOSTASIS",	"GOBP_NEURON_APOPTOTIC_PROCESS",	"GOBP_PURINE_CONTAINING_COMPOUND_METABOLIC_PROCESS",	"GOBP_CELLULAR_HOMEOSTASIS",	"GOBP_MAINTENANCE_OF_LOCATION_IN_CELL",	"GOBP_PEPTIDE_BIOSYNTHETIC_PROCESS",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_REGULATION_OF_CATION_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_PROTEOLYSIS",	"GOBP_REGULATION_OF_ACTIN_FILAMENT_BASED_PROCESS",	"GOBP_REGULATION_OF_LEUKOCYTE_MIGRATION",	"GOBP_NEGATIVE_REGULATION_OF_TRANSPORT",	"GOBP_REGULATION_OF_T_CELL_ACTIVATION",	"GOBP_REGULATION_OF_CELL_MORPHOGENESIS",	"GOBP_RESPONSE_TO_TOPOLOGICALLY_INCORRECT_PROTEIN",	"GOBP_CELL_CHEMOTAXIS",	"GOBP_REGULATION_OF_PROTEIN_CATABOLIC_PROCESS",	"GOBP_REGULATION_OF_PROTEOLYSIS",	"GOBP_REGULATION_OF_MEMBRANE_POTENTIAL",	"GOBP_RESPONSE_TO_ENDOPLASMIC_RETICULUM_STRESS",	"GOBP_POSITIVE_REGULATION_OF_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_MITOCHONDRION_ORGANIZATION",	"GOBP_REGULATION_OF_PROTEIN_STABILITY",	"GOBP_POSITIVE_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_ION_TRANSMEMBRANE_TRANSPORT",	"GOBP_APOPTOTIC_SIGNALING_PATHWAY",	"GOBP_I_KAPPAB_KINASE_NF_KAPPAB_SIGNALING",	"GOBP_AMEBOIDAL_TYPE_CELL_MIGRATION",	"GOBP_REGULATION_OF_TRANSMEMBRANE_TRANSPORT",	"GOBP_POSITIVE_REGULATION_OF_CELL_ADHESION",	"GOBP_RIBONUCLEOPROTEIN_COMPLEX_BIOGENESIS",	"GOBP_MAINTENANCE_OF_LOCATION",	"GOBP_NUCLEOBASE_CONTAINING_SMALL_MOLECULE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_CELLULAR_AMIDE_METABOLIC_PROCESS",	"GOBP_REGULATION_OF_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_REGULATION_OF_CELLULAR_COMPONENT_SIZE",	"GOBP_REGULATION_OF_BINDING",	"GOBP_REGULATION_OF_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_PROTEIN_MODIFICATION_BY_SMALL_PROTEIN_CONJUGATION_OR_REMOVAL",	"GOBP_METAL_ION_TRANSPORT",	"GOBP_RESPONSE_TO_WOUNDING",	"GOBP_REGULATION_OF_CELLULAR_MACROMOLECULE_BIOSYNTHETIC_PROCESS",	"GOBP_LEUKOCYTE_CELL_CELL_ADHESION",	"GOBP_REGULATION_OF_CELL_ADHESION",	"GOBP_RESPONSE_TO_BACTERIUM",	"GOBP_POSITIVE_REGULATION_OF_CELL_ACTIVATION",	"GOBP_PROTEASOMAL_PROTEIN_CATABOLIC_PROCESS",	"GOBP_POST_TRANSCRIPTIONAL_REGULATION_OF_GENE_EXPRESSION",	"GOBP_ACTIN_FILAMENT_ORGANIZATION",	"GOBP_EPITHELIAL_CELL_DIFFERENTIATION",	"GOBP_REGULATION_OF_HEMOPOIESIS",	"GOBP_REGULATION_OF_LYMPHOCYTE_ACTIVATION",	"GOBP_NEGATIVE_REGULATION_OF_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_CELL_DIFFERENTIATION",	"GOBP_POSITIVE_REGULATION_OF_TRANSPORT",	"GOBP_RESPONSE_TO_GROWTH_FACTOR",	"GOBP_POSITIVE_REGULATION_OF_LOCOMOTION",	"GOBP_T_CELL_ACTIVATION",	"GOBP_SUPRAMOLECULAR_FIBER_ORGANIZATION",	"GOBP_MONONUCLEAR_CELL_DIFFERENTIATION",	"GOBP_NEGATIVE_REGULATION_OF_DEVELOPMENTAL_PROCESS",	"GOBP_LEUKOCYTE_DIFFERENTIATION",	"GOBP_CELL_CELL_ADHESION",	"GOBP_NEGATIVE_REGULATION_OF_MULTICELLULAR_ORGANISMAL_PROCESS"
)
#,
#"RespondervsNonreponder_neg_reactome" = c("REACTOME_INTERLEUKIN_18_SIGNALING")
)


#data_1 = M1T4vs.all_pos_reactome 
#data_2 = M1T4vs.all_neg_reactome
#data_3 = M2T3_pos_reactome 
#data_4 = M2T3_neg_reactome


ggvenn(a, c("M1T4vs.all_pos_gobp", "RespondervsNonreponder_pos_gobp"))            # draw two-set venn
ggvenn(a, c("M3T4vs.all_pos_gobp", "RespondervsNonreponder_pos_gobp"))
#ggvenn(a, c("Set 1", "Set 2", "Set 3"))   # draw three-set venn
ggvenn(a)   # without set names, the first 4 elements in list will be chose to draw four-set venn








M1T4pos_responderpos_intersect= as.data.frame(Reduce(intersect, list(set1,set3)))
M1T4pos_responderpos_intersect =t(M1T4pos_responderpos_intersect)
M3T4pos_responderpos_intersect= as.data.frame(Reduce(intersect, list(set2,set3)))
M3T4pos_responderpos_intersect =t(M3T4pos_responderpos_intersect)
M1T4_M3T4_responder_all_intsect = as.data.frame(Reduce(intersect, list(set1,set2,set3)))
M1T4_M3T4_responder_all_intsect =t(M1T4_M3T4_responder_all_intsect)


M3T4_responder_only <- as.data.frame(Reduce(setdiff,list(M3T4pos_responderpos_intersect, M1T4_M3T4_responder_all_intsect)))
M3T4_responder_only =t(M3T4_responder_only)
M1T4_responder_only <- as.data.frame(Reduce(setdiff,list(M1T4pos_responderpos_intersect, M1T4_M3T4_responder_all_intsect)))
M1T4_responder_only =t(M1T4_responder_only)


n<-max(length(M1T4pos_responderpos_intersect),length(M3T4pos_responderpos_intersect),length(M1T4_M3T4_responder_all_intsect),length(M3T4_responder_only),length(M1T4_responder_only))
length(M1T4pos_responderpos_intersect)<-n
length(M3T4pos_responderpos_intersect)<-n
length(M1T4_M3T4_responder_all_intsect)<-n
length(M3T4_responder_only)<-n
length(M1T4_responder_only)<-n


M1T4_M3T4_responder_intersect = as.data.frame(cbind(M1T4pos_responderpos_intersect,M3T4pos_responderpos_intersect,M1T4_M3T4_responder_all_intsect,M3T4_responder_only,M1T4_responder_only))
colnames(M1T4_M3T4_responder_intersect)

M1T4_M3T4_responder_intersect<- M1T4_M3T4_responder_intersect %>% rename("V1" = "M1T4pos_responderpos_intersect")


write.csv(M1T4_M3T4_responder_intersect, "_Tcell_MaT4_pos_M3T4_pos_responder_pos_GOBP_intersect.csv")
