==> 2_venndiagram_from_T_cell_fgsea_quad.venn_B1T4_and_responder_nonresponder_04-25-23.R <==
rm(list=ls())
#install.packages("RDocumentation")
#install.packages("VennDiagram")
#install.packages('ggvenn')

#library(VennDiagram)
library(RColorBrewer)
library(ggplot2)
library(RTCGA)
library(dplyr)
library(RTCGAToolbox)
library(TCGAbiolinks)
library(data.table)
library(RDocumentation)
library(ggvenn)

setwd("F:/Ko/PD-1_responders/fgsea")


#fgsea: REACTOME results comparison



B1T4vs.all_pos_reactome = c("REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_PD_1_SIGNALING",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_FOLDING_OF_ACTIN_BY_CCT_TRIC",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",
                            "REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_FORMATION_OF_TUBULIN_FOLDING_INTERMEDIATES_BY_CCT_TRIC",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_COOPERATION_OF_PREFOLDIN_AND_TRIC_CCT_IN_ACTIN_AND_TUBULIN_FOLDING",	"REACTOME_PYROPTOSIS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",
                            "REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_ADRENALINE_NORADRENALINE_INHIBITS_INSULIN_SECRETION",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_SODIUM_CALCIUM_EXCHANGERS",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	
                            "REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_TCR_SIGNALING",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_OAS_ANTIVIRAL_RESPONSE",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_NUCLEOTIDE_CATABOLISM",	
                            "REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	"REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_THE_NLRP3_INFLAMMASOME",	"REACTOME_NOTCH2_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_ROCKS",
                            "REACTOME_GLUCONEOGENESIS",	"REACTOME_GENE_AND_PROTEIN_EXPRESSION_BY_JAK_STAT_SIGNALING_AFTER_INTERLEUKIN_12_STIMULATION",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	"REACTOME_ACYL_CHAIN_REMODELLING_OF_PE",	"REACTOME_INWARDLY_RECTIFYING_K_CHANNELS",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",
                            "REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",	"REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",	"REACTOME_SENSORY_PROCESSING_OF_SOUND_BY_OUTER_HAIR_CELLS_OF_THE_COCHLEA",	"REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_KEAP1_NFE2L2_PATHWAY",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_PURINERGIC_SIGNALING_IN_LEISHMANIASIS_INFECTION",	"REACTOME_HEDGEHOG_ON_STATE",
                            "REACTOME_UCH_PROTEINASES",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	"REACTOME_RHOH_GTPASE_CYCLE",	"REACTOME_RHO_GTPASES_ACTIVATE_PKNS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_INTERLEUKIN_12_SIGNALING",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	
                            "REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",	"REACTOME_SENSORY_PROCESSING_OF_SOUND",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_SIGNALING_BY_MODERATE_KINASE_ACTIVITY_BRAF_MUTANTS",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",
                            "REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",	"REACTOME_APOPTOSIS",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_REGULATED_NECROSIS",	"REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_SIGNALING_BY_HEDGEHOG",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",
                            "REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",	"REACTOME_PTEN_REGULATION",	"REACTOME_INTRINSIC_PATHWAY_FOR_APOPTOSIS",	"REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX1",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_C_TYPE_LECTIN_RECEPTORS_CLRS",	"REACTOME_G_ALPHA_12_13_SIGNALLING_EVENTS",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_DNA_REPLICATION_PRE_INITIATION",	"REACTOME_TP53_REGULATES_METABOLIC_GENES",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",
                            "REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_L1CAM_INTERACTIONS",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_UB_SPECIFIC_PROCESSING_PROTEASES",	"REACTOME_DNA_REPLICATION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_LEISHMANIA_INFECTION",	"REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",
                            "REACTOME_DEUBIQUITINATION",	"REACTOME_MRNA_SPLICING",	"REACTOME_NEDDYLATION",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_SIGNALING_BY_NOTCH",	"REACTOME_HIV_INFECTION",	"REACTOME_S_PHASE",	"REACTOME_INTRACELLULAR_SIGNALING_BY_SECOND_MESSENGERS",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_HEMOSTASIS",	"REACTOME_SEPARATION_OF_SISTER_CHROMATIDS",	"REACTOME_PROCESSING_OF_CAPPED_INTRON_CONTAINING_PRE_MRNA",	"REACTOME_RHO_GTPASE_EFFECTORS",	"REACTOME_MITOTIC_METAPHASE_AND_ANAPHASE",	"REACTOME_SIGNALING_BY_RECEPTOR_TYROSINE_KINASES",
                            "REACTOME_ANTIGEN_PROCESSING_UBIQUITINATION_PROTEASOME_DEGRADATION",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_DISEASES_OF_SIGNAL_TRANSDUCTION_BY_GROWTH_FACTOR_RECEPTORS_AND_SECOND_MESSENGERS"
)
B1T4vs.all_neg_reactome = c("REACTOME_DEFENSINS",	"REACTOME_ACTIVATION_OF_MATRIX_METALLOPROTEINASES",	"REACTOME_INITIAL_TRIGGERING_OF_COMPLEMENT",	"REACTOME_SCAVENGING_OF_HEME_FROM_PLASMA",	"REACTOME_COMPLEMENT_CASCADE",	"REACTOME_CREATION_OF_C4_AND_C2_ACTIVATORS",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY"
)

RespondervsNonreponder_pos_reactome =c("REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_CHAPERONE_MEDIATED_AUTOPHAGY",	"REACTOME_CONSTITUTIVE_SIGNALING_BY_LIGAND_RESPONSIVE_EGFR_CANCER_VARIANTS",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_INLB_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELL",	"REACTOME_RESPONSE_TO_METAL_IONS",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY",	"REACTOME_SIGNALING_BY_EGFR_IN_CANCER",	"REACTOME_DOWNREGULATION_OF_ERBB2_ERBB3_SIGNALING",	"REACTOME_LISTERIA_MONOCYTOGENES_ENTRY_INTO_HOST_CELLS",
                                       "REACTOME_EGFR_DOWNREGULATION",	"REACTOME_NEGATIVE_REGULATION_OF_MET_ACTIVITY",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION",	"REACTOME_TGF_BETA_RECEPTOR_SIGNALING_IN_EMT_EPITHELIAL_TO_MESENCHYMAL_TRANSITION",	"REACTOME_IRAK2_MEDIATED_ACTIVATION_OF_TAK1_COMPLEX",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_METALLOTHIONEINS_BIND_METALS",	"REACTOME_RHO_GTPASES_ACTIVATE_KTN1",	"REACTOME_PTK6_REGULATES_RTKS_AND_THEIR_EFFECTORS_AKT1_AND_DOK1",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_TICAM1_TRAF6_DEPENDENT_INDUCTION_OF_TAK1_COMPLEX",	
                                       "REACTOME_FLT3_SIGNALING_BY_CBL_MUTANTS",	"REACTOME_BINDING_AND_UPTAKE_OF_LIGANDS_BY_SCAVENGER_RECEPTORS",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_REGULATION_OF_BACH1_ACTIVITY",	"REACTOME_IRAK1_RECRUITS_IKK_COMPLEX",	"REACTOME_SIGNALING_BY_NOTCH1_HD_DOMAIN_MUTANTS_IN_CANCER",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_FOLDING_OF_ACTIN_BY_CCT_TRIC",	"REACTOME_INLA_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELLS",	"REACTOME_TICAM1_RIP1_MEDIATED_IKK_COMPLEX_RECRUITMENT",	"REACTOME_GENE_AND_PROTEIN_EXPRESSION_BY_JAK_STAT_SIGNALING_AFTER_INTERLEUKIN_12_STIMULATION",	
                                       "REACTOME_MITOPHAGY",	"REACTOME_PEXOPHAGY",	"REACTOME_SIGNALING_BY_NTRK3_TRKC",	"REACTOME_GLYCOGEN_STORAGE_DISEASES",	"REACTOME_COOPERATION_OF_PREFOLDIN_AND_TRIC_CCT_IN_ACTIN_AND_TUBULIN_FOLDING",	"REACTOME_RESPONSE_OF_MTB_TO_PHAGOCYTOSIS",	"REACTOME_PINK1_PRKN_MEDIATED_MITOPHAGY",	"REACTOME_KILLING_MECHANISMS",	"REACTOME_UPTAKE_AND_ACTIONS_OF_BACTERIAL_TOXINS",	"REACTOME_ACTIVATION_OF_THE_MRNA_UPON_BINDING_OF_THE_CAP_BINDING_COMPLEX_AND_EIFS_AND_SUBSEQUENT_BINDING_TO_43S",	"REACTOME_GLYCOGEN_SYNTHESIS",	"REACTOME_FORMATION_OF_TUBULIN_FOLDING_INTERMEDIATES_BY_CCT_TRIC",	"REACTOME_TRAF6_MEDIATED_INDUCTION_OF_TAK1_COMPLEX_WITHIN_TLR4_COMPLEX",	
                                       "REACTOME_IKK_COMPLEX_RECRUITMENT_MEDIATED_BY_RIP1",	"REACTOME_CONSTITUTIVE_SIGNALING_BY_EGFRVIII",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_PD_1_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_IQGAPS",	"REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_SIGNALING_BY_FGFR4_IN_DISEASE",	"REACTOME_PTK6_REGULATES_RHO_GTPASES_RAS_GTPASE_AND_MAP_KINASES",	"REACTOME_ASSEMBLY_OF_THE_HIV_VIRION",	"REACTOME_SEMA3A_PAK_DEPENDENT_AXON_REPULSION",	"REACTOME_AGGREPHAGY",	"REACTOME_NF_KB_IS_ACTIVATED_AND_SIGNALS_SURVIVAL",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	
                                       "REACTOME_SIGNALING_BY_ERBB2",	"REACTOME_SIGNALING_BY_FGFR1",	"REACTOME_SIGNALING_BY_FGFR3",	"REACTOME_SIGNALING_BY_FGFR4",	"REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",	"REACTOME_DETOXIFICATION_OF_REACTIVE_OXYGEN_SPECIES",	"REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_P75NTR_SIGNALS_VIA_NF_KB",	"REACTOME_PREVENTION_OF_PHAGOSOMAL_LYSOSOMAL_FUSION",	"REACTOME_DOWNREGULATION_OF_ERBB4_SIGNALING",	"REACTOME_SIGNALING_BY_ERBB2_ECD_MUTANTS",	"REACTOME_SIGNALING_BY_PTK6",	"REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_INTEGRATION_OF_PROVIRUS",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR2_SIGNALING",
                                       "REACTOME_NEGATIVE_REGULATION_OF_FGFR3_SIGNALING",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR4_SIGNALING",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR1_SIGNALING",	"REACTOME_SIGNALING_BY_ROBO_RECEPTORS",	"REACTOME_JNK_C_JUN_KINASES_PHOSPHORYLATION_AND_ACTIVATION_MEDIATED_BY_ACTIVATED_HUMAN_TAK1",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_REGULATION_OF_INNATE_IMMUNE_RESPONSES_TO_CYTOSOLIC_DNA",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_INFECTION_WITH_MYCOBACTERIUM_TUBERCULOSIS",	"REACTOME_NEGATIVE_REGULATION_OF_FLT3",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",	"REACTOME_FLT3_SIGNALING_IN_DISEASE",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_PCP_CE_PATHWAY",	
                                       "REACTOME_SPRY_REGULATION_OF_FGF_SIGNALING",	"REACTOME_NUCLEAR_EVENTS_STIMULATED_BY_ALK_SIGNALING_IN_CANCER",	"REACTOME_REGULATION_OF_EXPRESSION_OF_SLITS_AND_ROBOS",	"REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_SIGNALING_BY_MET",	"REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_GABA_SYNTHESIS_RELEASE_REUPTAKE_AND_DEGRADATION",	"REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",	"REACTOME_P75NTR_RECRUITS_SIGNALLING_COMPLEXES",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_MYOCLONIC_EPILEPSY_OF_LAFORA",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_REGULATION_OF_SIGNALING_BY_CBL",	
                                       "REACTOME_SIGNALING_BY_EGFR",	"REACTOME_STIMULI_SENSING_CHANNELS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_ALPHA_PROTEIN_KINASE_1_SIGNALING_PATHWAY",	"REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	"REACTOME_INTERLEUKIN_12_SIGNALING",	"REACTOME_UPTAKE_AND_FUNCTION_OF_DIPHTHERIA_TOXIN",	"REACTOME_RAS_PROCESSING",	"REACTOME_MRNA_SPLICING",	"REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_MAP3K8_TPL2_DEPENDENT_MAPK1_3_ACTIVATION",	"REACTOME_TRANSPORT_OF_NUCLEOSIDES_AND_FREE_PURINE_AND_PYRIMIDINE_BASES_ACROSS_THE_PLASMA_MEMBRANE",	
                                       "REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	"REACTOME_ACTIVATION_OF_RAC1_DOWNSTREAM_OF_NMDARS",	"REACTOME_REGULATION_OF_PTEN_LOCALIZATION",	"REACTOME_SUPPRESSION_OF_PHAGOSOMAL_MATURATION",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_SCAVENGING_BY_CLASS_A_RECEPTORS",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",	"REACTOME_GP1B_IX_V_ACTIVATION_SIGNALLING",	"REACTOME_REGULATION_OF_NF_KAPPA_B_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	
                                       "REACTOME_CALCINEURIN_ACTIVATES_NFAT",	"REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_2_LTR_CIRCLE_FORMATION",	"REACTOME_TCR_SIGNALING",	"REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_SIGNALING_BY_ERBB2_IN_CANCER",	"REACTOME_GPVI_MEDIATED_ACTIVATION_CASCADE",	"REACTOME_EARLY_PHASE_OF_HIV_LIFE_CYCLE",	"REACTOME_NEGATIVE_REGULATION_OF_MAPK_PATHWAY",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",
                                       "REACTOME_VLDLR_INTERNALISATION_AND_DEGRADATION",	"REACTOME_NOTCH3_ACTIVATION_AND_TRANSMISSION_OF_SIGNAL_TO_THE_NUCLEUS",	"REACTOME_KEAP1_NFE2L2_PATHWAY",	"REACTOME_SHC_MEDIATED_CASCADE_FGFR1",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_TP53_REGULATES_METABOLIC_GENES",	"REACTOME_FLT3_SIGNALING",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_GLYCOGEN_METABOLISM",
                                       "REACTOME_SYNTHESIS_OF_5_EICOSATETRAENOIC_ACIDS",	"REACTOME_MITOCHONDRIAL_BIOGENESIS",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_ATF6_ATF6_ALPHA_ACTIVATES_CHAPERONES",	"REACTOME_POST_CHAPERONIN_TUBULIN_FOLDING_PATHWAY",	"REACTOME_RHOBTB1_GTPASE_CYCLE",	"REACTOME_HEDGEHOG_ON_STATE",	"REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_TRANSLATION",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_TGF_BETA_RECEPTOR_SIGNALING_ACTIVATES_SMADS",	"REACTOME_ACTIVATED_NTRK2_SIGNALS_THROUGH_RAS",	"REACTOME_ACTIVATED_NTRK3_SIGNALS_THROUGH_RAS",
                                       "REACTOME_SHC_MEDIATED_CASCADE_FGFR3",	"REACTOME_SHC_MEDIATED_CASCADE_FGFR4",	"REACTOME_SIGNALING_BY_ERBB4",	"REACTOME_EGFR_TRANSACTIVATION_BY_GASTRIN",	"REACTOME_VEGFR2_MEDIATED_CELL_PROLIFERATION",	"REACTOME_SIGNALING_BY_NTRK2_TRKB",	"REACTOME_AMYLOID_FIBER_FORMATION",	"REACTOME_RHO_GTPASES_ACTIVATE_PAKS",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_LATE_ENDOSOMAL_MICROAUTOPHAGY",	"REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",	"REACTOME_COSTIMULATION_BY_THE_CD28_FAMILY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",
                                       "REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_SIGNALING_BY_ALK_IN_CANCER",	"REACTOME_PROCESSING_OF_CAPPED_INTRON_CONTAINING_PRE_MRNA",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_MRNA_SPLICING_MINOR_PATHWAY",	"REACTOME_IRON_UPTAKE_AND_TRANSPORT",	"REACTOME_DDX58_IFIH1_MEDIATED_INDUCTION_OF_INTERFERON_ALPHA_BETA",	"REACTOME_DEFECTIVE_INTRINSIC_PATHWAY_FOR_APOPTOSIS",	"REACTOME_FCERI_MEDIATED_MAPK_ACTIVATION",	"REACTOME_SIGNALING_BY_FGFR2",	"REACTOME_C_TYPE_LECTIN_RECEPTORS_CLRS",	"REACTOME_SIGNALING_BY_FGFR",	"REACTOME_SIGNALING_BY_HEDGEHOG",	"REACTOME_SIGNALING_BY_CSF3_G_CSF",
                                       "REACTOME_SIGNALING_BY_ERYTHROPOIETIN",	"REACTOME_NERVOUS_SYSTEM_DEVELOPMENT",	"REACTOME_OVARIAN_TUMOR_DOMAIN_PROTEASES",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_DNA_REPLICATION_PRE_INITIATION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_B_WICH_COMPLEX_POSITIVELY_REGULATES_RRNA_EXPRESSION",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_PTEN_REGULATION",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_HIV_INFECTION",	"REACTOME_SMAD2_SMAD3_SMAD4_HETEROTRIMER_REGULATES_TRANSCRIPTION",	"REACTOME_CELL_SURFACE_INTERACTIONS_AT_THE_VASCULAR_WALL",	"REACTOME_NOD1_2_SIGNALING_PATHWAY",	
                                       "REACTOME_NEGATIVE_REGULATORS_OF_DDX58_IFIH1_SIGNALING",	"REACTOME_TAK1_DEPENDENT_IKK_AND_NF_KAPPA_B_ACTIVATION",	"REACTOME_NOTCH1_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_DISEASES_OF_PROGRAMMED_CELL_DEATH",	"REACTOME_POSITIVE_EPIGENETIC_REGULATION_OF_RRNA_EXPRESSION",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_UCH_PROTEINASES",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_REGULATION_OF_TNFR1_SIGNALING",	"REACTOME_PLASMA_LIPOPROTEIN_ASSEMBLY_REMODELING_AND_CLEARANCE",	"REACTOME_FORMATION_OF_INCISION_COMPLEX_IN_GG_NER",	"REACTOME_SIGNALING_BY_TGF_BETA_RECEPTOR_COMPLEX",	"REACTOME_SIGNALING_BY_TGFB_FAMILY_MEMBERS",
                                       "REACTOME_SENESCENCE_ASSOCIATED_SECRETORY_PHENOTYPE_SASP",	"REACTOME_MITOCHONDRIAL_PROTEIN_IMPORT",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_DNA_REPLICATION",	"REACTOME_RNA_POLYMERASE_II_TRANSCRIPTION_TERMINATION",	"REACTOME_FORMATION_OF_TC_NER_PRE_INCISION_COMPLEX",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_EPIGENETIC_REGULATION_OF_GENE_EXPRESSION",	"REACTOME_SIGNALING_BY_NOTCH",	"REACTOME_APOPTOSIS",	"REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",	
                                       "REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_S_PHASE",	"REACTOME_ESTROGEN_DEPENDENT_GENE_EXPRESSION",	"REACTOME_EXTRA_NUCLEAR_ESTROGEN_SIGNALING",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_L1CAM_INTERACTIONS",	"REACTOME_UB_SPECIFIC_PROCESSING_PROTEASES",	"REACTOME_AUTOPHAGY",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_SIGNALING_BY_VEGF",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX1",	"REACTOME_PROTEIN_UBIQUITINATION",	"REACTOME_INTRACELLULAR_SIGNALING_BY_SECOND_MESSENGERS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_DISEASES_OF_SIGNAL_TRANSDUCTION_BY_GROWTH_FACTOR_RECEPTORS_AND_SECOND_MESSENGERS",
                                       "REACTOME_SEPARATION_OF_SISTER_CHROMATIDS",	"REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	"REACTOME_REGULATION_OF_PLK1_ACTIVITY_AT_G2_M_TRANSITION",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_CYTOPROTECTION_BY_HMOX1",	"REACTOME_PROTEIN_LOCALIZATION",	"REACTOME_RHOG_GTPASE_CYCLE",	"REACTOME_TRANSCRIPTION_OF_THE_HIV_GENOME",	"REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	"REACTOME_DUAL_INCISION_IN_TC_NER",	"REACTOME_CLATHRIN_MEDIATED_ENDOCYTOSIS",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_METAPHASE_AND_ANAPHASE",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_TP53",	"REACTOME_POTENTIAL_THERAPEUTICS_FOR_SARS",	"REACTOME_RHO_GTPASE_EFFECTORS",
                                       "REACTOME_ESR_MEDIATED_SIGNALING",	"REACTOME_HEMOSTASIS",	"REACTOME_NEUROTRANSMITTER_RECEPTORS_AND_POSTSYNAPTIC_SIGNAL_TRANSMISSION",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_DEVELOPMENTAL_BIOLOGY",	"REACTOME_ANTIGEN_PROCESSING_UBIQUITINATION_PROTEASOME_DEGRADATION",	"REACTOME_NEDDYLATION",	"REACTOME_SARS_COV_2_HOST_INTERACTIONS",	"REACTOME_M_PHASE",	"REACTOME_NUCLEOTIDE_EXCISION_REPAIR",	"REACTOME_TRANSMISSION_ACROSS_CHEMICAL_SYNAPSES",	"REACTOME_HIV_LIFE_CYCLE",	"REACTOME_METABOLISM_OF_AMINO_ACIDS_AND_DERIVATIVES",	"REACTOME_DEUBIQUITINATION",	"REACTOME_SIGNALING_BY_RECEPTOR_TYROSINE_KINASES",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",
                                       "REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_CELL_CYCLE_CHECKPOINTS",	"REACTOME_SARS_COV_INFECTIONS",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES",	"REACTOME_CELL_CYCLE_MITOTIC",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_SIGNALING_BY_RHO_GTPASES_MIRO_GTPASES_AND_RHOBTB3",	"REACTOME_NEUTROPHIL_DEGRANULATION")
RespondervsNonreponder_neg_reactome = c("REACTOME_INTERLEUKIN_18_SIGNALING")



data_1 = B1T4vs.all_pos_reactome 
data_2 = B1T4vs.all_neg_reactome
data_3 = RespondervsNonreponder_pos_reactome 
data_4 = RespondervsNonreponder_neg_reactome


set1 = as.list(data_1)
set2 = as.list(data_2)
set3 = as.list(data_3)
set4 = as.list(data_4)

a <- list("B1T4vs.all_pos_reactome" = c("REACTOME_ATTENUATION_PHASE",	"REACTOME_HSF1_DEPENDENT_TRANSACTIVATION",	"REACTOME_PD_1_SIGNALING",	"REACTOME_HSF1_ACTIVATION",	"REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_INTERFERON_ALPHA_BETA_SIGNALING",	"REACTOME_HSP90_CHAPERONE_CYCLE_FOR_STEROID_HORMONE_RECEPTORS_SHR_IN_THE_PRESENCE_OF_LIGAND",	"REACTOME_FOLDING_OF_ACTIN_BY_CCT_TRIC",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",
                                        "REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_FORMATION_OF_TUBULIN_FOLDING_INTERMEDIATES_BY_CCT_TRIC",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_REGULATION_OF_HSF1_MEDIATED_HEAT_SHOCK_RESPONSE",	"REACTOME_COOPERATION_OF_PREFOLDIN_AND_TRIC_CCT_IN_ACTIN_AND_TUBULIN_FOLDING",	"REACTOME_PYROPTOSIS",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_SCAVENGING_BY_CLASS_F_RECEPTORS",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",
                                        "REACTOME_RHO_GTPASES_ACTIVATE_RHOTEKIN_AND_RHOPHILINS",	"REACTOME_CELLULAR_RESPONSE_TO_HEAT_STRESS",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",	"REACTOME_SEMA4D_IN_SEMAPHORIN_SIGNALING",	"REACTOME_PCP_CE_PATHWAY",	"REACTOME_ADRENALINE_NORADRENALINE_INHIBITS_INSULIN_SECRETION",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_SODIUM_CALCIUM_EXCHANGERS",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	
                                        "REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_TCR_SIGNALING",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_SEMA4D_INDUCED_CELL_MIGRATION_AND_GROWTH_CONE_COLLAPSE",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_OAS_ANTIVIRAL_RESPONSE",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_RHOBTB_GTPASE_CYCLE",	"REACTOME_NUCLEOTIDE_CATABOLISM",	
                                        "REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	"REACTOME_ANTIGEN_PRESENTATION_FOLDING_ASSEMBLY_AND_PEPTIDE_LOADING_OF_CLASS_I_MHC",	"REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_THE_NLRP3_INFLAMMASOME",	"REACTOME_NOTCH2_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_ROCKS",
                                        "REACTOME_GLUCONEOGENESIS",	"REACTOME_GENE_AND_PROTEIN_EXPRESSION_BY_JAK_STAT_SIGNALING_AFTER_INTERLEUKIN_12_STIMULATION",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	"REACTOME_ACYL_CHAIN_REMODELLING_OF_PE",	"REACTOME_INWARDLY_RECTIFYING_K_CHANNELS",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",
                                        "REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",	"REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",	"REACTOME_SENSORY_PROCESSING_OF_SOUND_BY_OUTER_HAIR_CELLS_OF_THE_COCHLEA",	"REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_KEAP1_NFE2L2_PATHWAY",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_PURINERGIC_SIGNALING_IN_LEISHMANIASIS_INFECTION",	"REACTOME_HEDGEHOG_ON_STATE",
                                        "REACTOME_UCH_PROTEINASES",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",	"REACTOME_RHOH_GTPASE_CYCLE",	"REACTOME_RHO_GTPASES_ACTIVATE_PKNS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_INTERLEUKIN_12_SIGNALING",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	
                                        "REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",	"REACTOME_SENSORY_PROCESSING_OF_SOUND",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_SIGNALING_BY_MODERATE_KINASE_ACTIVITY_BRAF_MUTANTS",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",
                                        "REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_RESPONSE_TO_ELEVATED_PLATELET_CYTOSOLIC_CA2",	"REACTOME_APOPTOSIS",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_REGULATED_NECROSIS",	"REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_SIGNALING_BY_HEDGEHOG",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",
                                        "REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",	"REACTOME_PTEN_REGULATION",	"REACTOME_INTRINSIC_PATHWAY_FOR_APOPTOSIS",	"REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX1",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_C_TYPE_LECTIN_RECEPTORS_CLRS",	"REACTOME_G_ALPHA_12_13_SIGNALLING_EVENTS",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_DNA_REPLICATION_PRE_INITIATION",	"REACTOME_TP53_REGULATES_METABOLIC_GENES",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",
                                        "REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_L1CAM_INTERACTIONS",	"REACTOME_NEUTROPHIL_DEGRANULATION",	"REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_UB_SPECIFIC_PROCESSING_PROTEASES",	"REACTOME_DNA_REPLICATION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_LEISHMANIA_INFECTION",	"REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",
                                        "REACTOME_DEUBIQUITINATION",	"REACTOME_MRNA_SPLICING",	"REACTOME_NEDDYLATION",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_SIGNALING_BY_NOTCH",	"REACTOME_HIV_INFECTION",	"REACTOME_S_PHASE",	"REACTOME_INTRACELLULAR_SIGNALING_BY_SECOND_MESSENGERS",	"REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_HEMOSTASIS",	"REACTOME_SEPARATION_OF_SISTER_CHROMATIDS",	"REACTOME_PROCESSING_OF_CAPPED_INTRON_CONTAINING_PRE_MRNA",	"REACTOME_RHO_GTPASE_EFFECTORS",	"REACTOME_MITOTIC_METAPHASE_AND_ANAPHASE",	"REACTOME_SIGNALING_BY_RECEPTOR_TYROSINE_KINASES",
                                        "REACTOME_ANTIGEN_PROCESSING_UBIQUITINATION_PROTEASOME_DEGRADATION",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",	"REACTOME_DISEASES_OF_SIGNAL_TRANSDUCTION_BY_GROWTH_FACTOR_RECEPTORS_AND_SECOND_MESSENGERS"),
          "B1T4vs.all_neg_reactome" = c("REACTOME_DEFENSINS",	"REACTOME_ACTIVATION_OF_MATRIX_METALLOPROTEINASES",	"REACTOME_INITIAL_TRIGGERING_OF_COMPLEMENT",	"REACTOME_SCAVENGING_OF_HEME_FROM_PLASMA",	"REACTOME_COMPLEMENT_CASCADE",	"REACTOME_CREATION_OF_C4_AND_C2_ACTIVATORS",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY"),
          "RespondervsNonreponder_pos_reactome" = c("REACTOME_FORMATION_OF_ATP_BY_CHEMIOSMOTIC_COUPLING",	"REACTOME_CHAPERONE_MEDIATED_AUTOPHAGY",	"REACTOME_CONSTITUTIVE_SIGNALING_BY_LIGAND_RESPONSIVE_EGFR_CANCER_VARIANTS",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT_ATP_SYNTHESIS_BY_CHEMIOSMOTIC_COUPLING_AND_HEAT_PRODUCTION_BY_UNCOUPLING_PROTEINS",	"REACTOME_INLB_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELL",	"REACTOME_RESPONSE_TO_METAL_IONS",	"REACTOME_CRISTAE_FORMATION",	"REACTOME_SARS_COV_2_MODULATES_HOST_TRANSLATION_MACHINERY",	"REACTOME_SIGNALING_BY_EGFR_IN_CANCER",	"REACTOME_DOWNREGULATION_OF_ERBB2_ERBB3_SIGNALING",	"REACTOME_LISTERIA_MONOCYTOGENES_ENTRY_INTO_HOST_CELLS",
                                                    "REACTOME_EGFR_DOWNREGULATION",	"REACTOME_NEGATIVE_REGULATION_OF_MET_ACTIVITY",	"REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION",	"REACTOME_TGF_BETA_RECEPTOR_SIGNALING_IN_EMT_EPITHELIAL_TO_MESENCHYMAL_TRANSITION",	"REACTOME_IRAK2_MEDIATED_ACTIVATION_OF_TAK1_COMPLEX",	"REACTOME_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_METALLOTHIONEINS_BIND_METALS",	"REACTOME_RHO_GTPASES_ACTIVATE_KTN1",	"REACTOME_PTK6_REGULATES_RTKS_AND_THEIR_EFFECTORS_AKT1_AND_DOK1",	"REACTOME_GENERATION_OF_SECOND_MESSENGER_MOLECULES",	"REACTOME_SRP_DEPENDENT_COTRANSLATIONAL_PROTEIN_TARGETING_TO_MEMBRANE",	"REACTOME_TICAM1_TRAF6_DEPENDENT_INDUCTION_OF_TAK1_COMPLEX",	
                                                    "REACTOME_FLT3_SIGNALING_BY_CBL_MUTANTS",	"REACTOME_BINDING_AND_UPTAKE_OF_LIGANDS_BY_SCAVENGER_RECEPTORS",	"REACTOME_EUKARYOTIC_TRANSLATION_INITIATION",	"REACTOME_THE_CITRIC_ACID_TCA_CYCLE_AND_RESPIRATORY_ELECTRON_TRANSPORT",	"REACTOME_REGULATION_OF_BACH1_ACTIVITY",	"REACTOME_IRAK1_RECRUITS_IKK_COMPLEX",	"REACTOME_SIGNALING_BY_NOTCH1_HD_DOMAIN_MUTANTS_IN_CANCER",	"REACTOME_COMPLEX_I_BIOGENESIS",	"REACTOME_FOLDING_OF_ACTIN_BY_CCT_TRIC",	"REACTOME_INLA_MEDIATED_ENTRY_OF_LISTERIA_MONOCYTOGENES_INTO_HOST_CELLS",	"REACTOME_TICAM1_RIP1_MEDIATED_IKK_COMPLEX_RECRUITMENT",	"REACTOME_GENE_AND_PROTEIN_EXPRESSION_BY_JAK_STAT_SIGNALING_AFTER_INTERLEUKIN_12_STIMULATION",	
                                                    "REACTOME_MITOPHAGY",	"REACTOME_PEXOPHAGY",	"REACTOME_SIGNALING_BY_NTRK3_TRKC",	"REACTOME_GLYCOGEN_STORAGE_DISEASES",	"REACTOME_COOPERATION_OF_PREFOLDIN_AND_TRIC_CCT_IN_ACTIN_AND_TUBULIN_FOLDING",	"REACTOME_RESPONSE_OF_MTB_TO_PHAGOCYTOSIS",	"REACTOME_PINK1_PRKN_MEDIATED_MITOPHAGY",	"REACTOME_KILLING_MECHANISMS",	"REACTOME_UPTAKE_AND_ACTIONS_OF_BACTERIAL_TOXINS",	"REACTOME_ACTIVATION_OF_THE_MRNA_UPON_BINDING_OF_THE_CAP_BINDING_COMPLEX_AND_EIFS_AND_SUBSEQUENT_BINDING_TO_43S",	"REACTOME_GLYCOGEN_SYNTHESIS",	"REACTOME_FORMATION_OF_TUBULIN_FOLDING_INTERMEDIATES_BY_CCT_TRIC",	"REACTOME_TRAF6_MEDIATED_INDUCTION_OF_TAK1_COMPLEX_WITHIN_TLR4_COMPLEX",	
                                                    "REACTOME_IKK_COMPLEX_RECRUITMENT_MEDIATED_BY_RIP1",	"REACTOME_CONSTITUTIVE_SIGNALING_BY_EGFRVIII",	"REACTOME_THE_ROLE_OF_GTSE1_IN_G2_M_PROGRESSION_AFTER_G2_CHECKPOINT",	"REACTOME_PD_1_SIGNALING",	"REACTOME_RHO_GTPASES_ACTIVATE_IQGAPS",	"REACTOME_NEGATIVE_REGULATION_OF_NOTCH4_SIGNALING",	"REACTOME_SIGNALING_BY_FGFR4_IN_DISEASE",	"REACTOME_PTK6_REGULATES_RHO_GTPASES_RAS_GTPASE_AND_MAP_KINASES",	"REACTOME_ASSEMBLY_OF_THE_HIV_VIRION",	"REACTOME_SEMA3A_PAK_DEPENDENT_AXON_REPULSION",	"REACTOME_AGGREPHAGY",	"REACTOME_NF_KB_IS_ACTIVATED_AND_SIGNALS_SURVIVAL",	"REACTOME_METABOLISM_OF_POLYAMINES",	"REACTOME_AUF1_HNRNP_D0_BINDS_AND_DESTABILIZES_MRNA",	
                                                    "REACTOME_SIGNALING_BY_ERBB2",	"REACTOME_SIGNALING_BY_FGFR1",	"REACTOME_SIGNALING_BY_FGFR3",	"REACTOME_SIGNALING_BY_FGFR4",	"REACTOME_DOWNSTREAM_SIGNALING_EVENTS_OF_B_CELL_RECEPTOR_BCR",	"REACTOME_DETOXIFICATION_OF_REACTIVE_OXYGEN_SPECIES",	"REACTOME_SIGNALING_BY_NOTCH4",	"REACTOME_P75NTR_SIGNALS_VIA_NF_KB",	"REACTOME_PREVENTION_OF_PHAGOSOMAL_LYSOSOMAL_FUSION",	"REACTOME_DOWNREGULATION_OF_ERBB4_SIGNALING",	"REACTOME_SIGNALING_BY_ERBB2_ECD_MUTANTS",	"REACTOME_SIGNALING_BY_PTK6",	"REACTOME_REGULATION_OF_RUNX3_EXPRESSION_AND_ACTIVITY",	"REACTOME_INTEGRATION_OF_PROVIRUS",	"REACTOME_STRIATED_MUSCLE_CONTRACTION",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR2_SIGNALING",
                                                    "REACTOME_NEGATIVE_REGULATION_OF_FGFR3_SIGNALING",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR4_SIGNALING",	"REACTOME_NEGATIVE_REGULATION_OF_FGFR1_SIGNALING",	"REACTOME_SIGNALING_BY_ROBO_RECEPTORS",	"REACTOME_JNK_C_JUN_KINASES_PHOSPHORYLATION_AND_ACTIVATION_MEDIATED_BY_ACTIVATED_HUMAN_TAK1",	"REACTOME_SMOOTH_MUSCLE_CONTRACTION",	"REACTOME_REGULATION_OF_INNATE_IMMUNE_RESPONSES_TO_CYTOSOLIC_DNA",	"REACTOME_ASYMMETRIC_LOCALIZATION_OF_PCP_PROTEINS",	"REACTOME_INFECTION_WITH_MYCOBACTERIUM_TUBERCULOSIS",	"REACTOME_NEGATIVE_REGULATION_OF_FLT3",	"REACTOME_SCF_BETA_TRCP_MEDIATED_DEGRADATION_OF_EMI1",	"REACTOME_FLT3_SIGNALING_IN_DISEASE",	"REACTOME_STABILIZATION_OF_P53",	"REACTOME_PCP_CE_PATHWAY",	
                                                    "REACTOME_SPRY_REGULATION_OF_FGF_SIGNALING",	"REACTOME_NUCLEAR_EVENTS_STIMULATED_BY_ALK_SIGNALING_IN_CANCER",	"REACTOME_REGULATION_OF_EXPRESSION_OF_SLITS_AND_ROBOS",	"REACTOME_SCF_SKP2_MEDIATED_DEGRADATION_OF_P27_P21",	"REACTOME_SIGNALING_BY_MET",	"REACTOME_SELECTIVE_AUTOPHAGY",	"REACTOME_GABA_SYNTHESIS_RELEASE_REUPTAKE_AND_DEGRADATION",	"REACTOME_REGULATION_OF_RUNX2_EXPRESSION_AND_ACTIVITY",	"REACTOME_P75NTR_RECRUITS_SIGNALLING_COMPLEXES",	"REACTOME_CELLULAR_RESPONSE_TO_HYPOXIA",	"REACTOME_DEGRADATION_OF_AXIN",	"REACTOME_MYOCLONIC_EPILEPSY_OF_LAFORA",	"REACTOME_DEGRADATION_OF_DVL",	"REACTOME_MAPK6_MAPK4_SIGNALING",	"REACTOME_REGULATION_OF_SIGNALING_BY_CBL",	
                                                    "REACTOME_SIGNALING_BY_EGFR",	"REACTOME_STIMULI_SENSING_CHANNELS",	"REACTOME_APC_C_CDH1_MEDIATED_DEGRADATION_OF_CDC20_AND_OTHER_APC_C_CDH1_TARGETED_PROTEINS_IN_LATE_MITOSIS_EARLY_G1",	"REACTOME_ALPHA_PROTEIN_KINASE_1_SIGNALING_PATHWAY",	"REACTOME_FCERI_MEDIATED_NF_KB_ACTIVATION",	"REACTOME_INTERLEUKIN_12_SIGNALING",	"REACTOME_UPTAKE_AND_FUNCTION_OF_DIPHTHERIA_TOXIN",	"REACTOME_RAS_PROCESSING",	"REACTOME_MRNA_SPLICING",	"REACTOME_CYCLIN_A_CDK2_ASSOCIATED_EVENTS_AT_S_PHASE_ENTRY",	"REACTOME_MAP3K8_TPL2_DEPENDENT_MAPK1_3_ACTIVATION",	"REACTOME_TRANSPORT_OF_NUCLEOSIDES_AND_FREE_PURINE_AND_PYRIMIDINE_BASES_ACROSS_THE_PLASMA_MEMBRANE",	
                                                    "REACTOME_G1_S_DNA_DAMAGE_CHECKPOINTS",	"REACTOME_ACTIVATION_OF_RAC1_DOWNSTREAM_OF_NMDARS",	"REACTOME_REGULATION_OF_PTEN_LOCALIZATION",	"REACTOME_SUPPRESSION_OF_PHAGOSOMAL_MATURATION",	"REACTOME_RHOBTB2_GTPASE_CYCLE",	"REACTOME_RHO_GTPASES_ACTIVATE_WASPS_AND_WAVES",	"REACTOME_DEGRADATION_OF_GLI1_BY_THE_PROTEASOME",	"REACTOME_CROSS_PRESENTATION_OF_SOLUBLE_EXOGENOUS_ANTIGENS_ENDOSOMES",	"REACTOME_NUCLEAR_EVENTS_MEDIATED_BY_NFE2L2",	"REACTOME_SCAVENGING_BY_CLASS_A_RECEPTORS",	"REACTOME_HEDGEHOG_LIGAND_BIOGENESIS",	"REACTOME_GP1B_IX_V_ACTIVATION_SIGNALLING",	"REACTOME_REGULATION_OF_NF_KAPPA_B_SIGNALING",	"REACTOME_RHOBTB_GTPASE_CYCLE",	
                                                    "REACTOME_CALCINEURIN_ACTIVATES_NFAT",	"REACTOME_DECTIN_1_MEDIATED_NONCANONICAL_NF_KB_SIGNALING",	"REACTOME_2_LTR_CIRCLE_FORMATION",	"REACTOME_TCR_SIGNALING",	"REACTOME_SIGNALING_BY_THE_B_CELL_RECEPTOR_BCR",	"REACTOME_REGULATION_OF_RAS_BY_GAPS",	"REACTOME_HEDGEHOG_OFF_STATE",	"REACTOME_SIGNALING_BY_ERBB2_IN_CANCER",	"REACTOME_GPVI_MEDIATED_ACTIVATION_CASCADE",	"REACTOME_EARLY_PHASE_OF_HIV_LIFE_CYCLE",	"REACTOME_NEGATIVE_REGULATION_OF_MAPK_PATHWAY",	"REACTOME_REGULATION_OF_PTEN_STABILITY_AND_ACTIVITY",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX3",	"REACTOME_DEFECTIVE_CFTR_CAUSES_CYSTIC_FIBROSIS",	"REACTOME_EPHB_MEDIATED_FORWARD_SIGNALING",
                                                    "REACTOME_VLDLR_INTERNALISATION_AND_DEGRADATION",	"REACTOME_NOTCH3_ACTIVATION_AND_TRANSMISSION_OF_SIGNAL_TO_THE_NUCLEUS",	"REACTOME_KEAP1_NFE2L2_PATHWAY",	"REACTOME_SHC_MEDIATED_CASCADE_FGFR1",	"REACTOME_APC_C_MEDIATED_DEGRADATION_OF_CELL_CYCLE_PROTEINS",	"REACTOME_FC_EPSILON_RECEPTOR_FCERI_SIGNALING",	"REACTOME_CELLULAR_RESPONSE_TO_CHEMICAL_STRESS",	"REACTOME_TP53_REGULATES_METABOLIC_GENES",	"REACTOME_FLT3_SIGNALING",	"REACTOME_DEGRADATION_OF_BETA_CATENIN_BY_THE_DESTRUCTION_COMPLEX",	"REACTOME_REGULATION_OF_MRNA_STABILITY_BY_PROTEINS_THAT_BIND_AU_RICH_ELEMENTS",	"REACTOME_ABC_TRANSPORTER_DISORDERS",	"REACTOME_GLYCOGEN_METABOLISM",
                                                    "REACTOME_SYNTHESIS_OF_5_EICOSATETRAENOIC_ACIDS",	"REACTOME_MITOCHONDRIAL_BIOGENESIS",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX2",	"REACTOME_ATF6_ATF6_ALPHA_ACTIVATES_CHAPERONES",	"REACTOME_POST_CHAPERONIN_TUBULIN_FOLDING_PATHWAY",	"REACTOME_RHOBTB1_GTPASE_CYCLE",	"REACTOME_HEDGEHOG_ON_STATE",	"REACTOME_HOST_INTERACTIONS_OF_HIV_FACTORS",	"REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_DIFFERENTIATION_OF_HSCS",	"REACTOME_TRANSLATION",	"REACTOME_GAP_JUNCTION_TRAFFICKING_AND_REGULATION",	"REACTOME_TGF_BETA_RECEPTOR_SIGNALING_ACTIVATES_SMADS",	"REACTOME_ACTIVATED_NTRK2_SIGNALS_THROUGH_RAS",	"REACTOME_ACTIVATED_NTRK3_SIGNALS_THROUGH_RAS",
                                                    "REACTOME_SHC_MEDIATED_CASCADE_FGFR3",	"REACTOME_SHC_MEDIATED_CASCADE_FGFR4",	"REACTOME_SIGNALING_BY_ERBB4",	"REACTOME_EGFR_TRANSACTIVATION_BY_GASTRIN",	"REACTOME_VEGFR2_MEDIATED_CELL_PROLIFERATION",	"REACTOME_SIGNALING_BY_NTRK2_TRKB",	"REACTOME_AMYLOID_FIBER_FORMATION",	"REACTOME_RHO_GTPASES_ACTIVATE_PAKS",	"REACTOME_COOPERATION_OF_PDCL_PHLP1_AND_TRIC_CCT_IN_G_PROTEIN_BETA_FOLDING",	"REACTOME_LATE_ENDOSOMAL_MICROAUTOPHAGY",	"REACTOME_INTERLEUKIN_12_FAMILY_SIGNALING",	"REACTOME_COSTIMULATION_BY_THE_CD28_FAMILY",	"REACTOME_BETA_CATENIN_INDEPENDENT_WNT_SIGNALING",	"REACTOME_CLEC7A_DECTIN_1_SIGNALING",	"REACTOME_ABC_FAMILY_PROTEINS_MEDIATED_TRANSPORT",
                                                    "REACTOME_TNFR2_NON_CANONICAL_NF_KB_PATHWAY",	"REACTOME_SIGNALING_BY_ALK_IN_CANCER",	"REACTOME_PROCESSING_OF_CAPPED_INTRON_CONTAINING_PRE_MRNA",	"REACTOME_ORC1_REMOVAL_FROM_CHROMATIN",	"REACTOME_SWITCHING_OF_ORIGINS_TO_A_POST_REPLICATIVE_STATE",	"REACTOME_MRNA_SPLICING_MINOR_PATHWAY",	"REACTOME_IRON_UPTAKE_AND_TRANSPORT",	"REACTOME_DDX58_IFIH1_MEDIATED_INDUCTION_OF_INTERFERON_ALPHA_BETA",	"REACTOME_DEFECTIVE_INTRINSIC_PATHWAY_FOR_APOPTOSIS",	"REACTOME_FCERI_MEDIATED_MAPK_ACTIVATION",	"REACTOME_SIGNALING_BY_FGFR2",	"REACTOME_C_TYPE_LECTIN_RECEPTORS_CLRS",	"REACTOME_SIGNALING_BY_FGFR",	"REACTOME_SIGNALING_BY_HEDGEHOG",	"REACTOME_SIGNALING_BY_CSF3_G_CSF",
                                                    "REACTOME_SIGNALING_BY_ERYTHROPOIETIN",	"REACTOME_NERVOUS_SYSTEM_DEVELOPMENT",	"REACTOME_OVARIAN_TUMOR_DOMAIN_PROTEASES",	"REACTOME_GLUCONEOGENESIS",	"REACTOME_DNA_REPLICATION_PRE_INITIATION",	"REACTOME_MITOTIC_G2_G2_M_PHASES",	"REACTOME_B_WICH_COMPLEX_POSITIVELY_REGULATES_RRNA_EXPRESSION",	"REACTOME_ANTIGEN_PROCESSING_CROSS_PRESENTATION",	"REACTOME_SEMAPHORIN_INTERACTIONS",	"REACTOME_PTEN_REGULATION",	"REACTOME_INTERLEUKIN_1_SIGNALING",	"REACTOME_HIV_INFECTION",	"REACTOME_SMAD2_SMAD3_SMAD4_HETEROTRIMER_REGULATES_TRANSCRIPTION",	"REACTOME_CELL_SURFACE_INTERACTIONS_AT_THE_VASCULAR_WALL",	"REACTOME_NOD1_2_SIGNALING_PATHWAY",	
                                                    "REACTOME_NEGATIVE_REGULATORS_OF_DDX58_IFIH1_SIGNALING",	"REACTOME_TAK1_DEPENDENT_IKK_AND_NF_KAPPA_B_ACTIVATION",	"REACTOME_NOTCH1_INTRACELLULAR_DOMAIN_REGULATES_TRANSCRIPTION",	"REACTOME_DISEASES_OF_PROGRAMMED_CELL_DEATH",	"REACTOME_POSITIVE_EPIGENETIC_REGULATION_OF_RRNA_EXPRESSION",	"REACTOME_PROTEIN_FOLDING",	"REACTOME_UCH_PROTEINASES",	"REACTOME_MHC_CLASS_II_ANTIGEN_PRESENTATION",	"REACTOME_REGULATION_OF_TNFR1_SIGNALING",	"REACTOME_PLASMA_LIPOPROTEIN_ASSEMBLY_REMODELING_AND_CLEARANCE",	"REACTOME_FORMATION_OF_INCISION_COMPLEX_IN_GG_NER",	"REACTOME_SIGNALING_BY_TGF_BETA_RECEPTOR_COMPLEX",	"REACTOME_SIGNALING_BY_TGFB_FAMILY_MEMBERS",
                                                    "REACTOME_SENESCENCE_ASSOCIATED_SECRETORY_PHENOTYPE_SASP",	"REACTOME_MITOCHONDRIAL_PROTEIN_IMPORT",	"REACTOME_SYNTHESIS_OF_DNA",	"REACTOME_INTERFERON_SIGNALING",	"REACTOME_DNA_REPLICATION",	"REACTOME_RNA_POLYMERASE_II_TRANSCRIPTION_TERMINATION",	"REACTOME_FORMATION_OF_TC_NER_PRE_INCISION_COMPLEX",	"REACTOME_INTERLEUKIN_1_FAMILY_SIGNALING",	"REACTOME_PROGRAMMED_CELL_DEATH",	"REACTOME_EPH_EPHRIN_SIGNALING",	"REACTOME_MITOTIC_G1_PHASE_AND_G1_S_TRANSITION",	"REACTOME_SIGNALING_BY_WNT",	"REACTOME_EPIGENETIC_REGULATION_OF_GENE_EXPRESSION",	"REACTOME_SIGNALING_BY_NOTCH",	"REACTOME_APOPTOSIS",	"REACTOME_TCF_DEPENDENT_SIGNALING_IN_RESPONSE_TO_WNT",	
                                                    "REACTOME_INTERFERON_GAMMA_SIGNALING",	"REACTOME_S_PHASE",	"REACTOME_ESTROGEN_DEPENDENT_GENE_EXPRESSION",	"REACTOME_EXTRA_NUCLEAR_ESTROGEN_SIGNALING",	"REACTOME_MAPK_FAMILY_SIGNALING_CASCADES",	"REACTOME_L1CAM_INTERACTIONS",	"REACTOME_UB_SPECIFIC_PROCESSING_PROTEASES",	"REACTOME_AUTOPHAGY",	"REACTOME_FCGAMMA_RECEPTOR_FCGR_DEPENDENT_PHAGOCYTOSIS",	"REACTOME_INFLUENZA_INFECTION",	"REACTOME_SIGNALING_BY_VEGF",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_RUNX1",	"REACTOME_PROTEIN_UBIQUITINATION",	"REACTOME_INTRACELLULAR_SIGNALING_BY_SECOND_MESSENGERS",	"REACTOME_MITOCHONDRIAL_TRANSLATION",	"REACTOME_DISEASES_OF_SIGNAL_TRANSDUCTION_BY_GROWTH_FACTOR_RECEPTORS_AND_SECOND_MESSENGERS",
                                                    "REACTOME_SEPARATION_OF_SISTER_CHROMATIDS",	"REACTOME_TRANSLOCATION_OF_SLC2A4_GLUT4_TO_THE_PLASMA_MEMBRANE",	"REACTOME_REGULATION_OF_PLK1_ACTIVITY_AT_G2_M_TRANSITION",	"REACTOME_G2_M_CHECKPOINTS",	"REACTOME_CYTOPROTECTION_BY_HMOX1",	"REACTOME_PROTEIN_LOCALIZATION",	"REACTOME_RHOG_GTPASE_CYCLE",	"REACTOME_TRANSCRIPTION_OF_THE_HIV_GENOME",	"REACTOME_DISORDERS_OF_TRANSMEMBRANE_TRANSPORTERS",	"REACTOME_DUAL_INCISION_IN_TC_NER",	"REACTOME_CLATHRIN_MEDIATED_ENDOCYTOSIS",	"REACTOME_SIGNALING_BY_INTERLEUKINS",	"REACTOME_MITOTIC_METAPHASE_AND_ANAPHASE",	"REACTOME_TRANSCRIPTIONAL_REGULATION_BY_TP53",	"REACTOME_POTENTIAL_THERAPEUTICS_FOR_SARS",	"REACTOME_RHO_GTPASE_EFFECTORS",
                                                    "REACTOME_ESR_MEDIATED_SIGNALING",	"REACTOME_HEMOSTASIS",	"REACTOME_NEUROTRANSMITTER_RECEPTORS_AND_POSTSYNAPTIC_SIGNAL_TRANSMISSION",	"REACTOME_PLATELET_ACTIVATION_SIGNALING_AND_AGGREGATION",	"REACTOME_DEVELOPMENTAL_BIOLOGY",	"REACTOME_ANTIGEN_PROCESSING_UBIQUITINATION_PROTEASOME_DEGRADATION",	"REACTOME_NEDDYLATION",	"REACTOME_SARS_COV_2_HOST_INTERACTIONS",	"REACTOME_M_PHASE",	"REACTOME_NUCLEOTIDE_EXCISION_REPAIR",	"REACTOME_TRANSMISSION_ACROSS_CHEMICAL_SYNAPSES",	"REACTOME_HIV_LIFE_CYCLE",	"REACTOME_METABOLISM_OF_AMINO_ACIDS_AND_DERIVATIVES",	"REACTOME_DEUBIQUITINATION",	"REACTOME_SIGNALING_BY_RECEPTOR_TYROSINE_KINASES",	"REACTOME_ORGANELLE_BIOGENESIS_AND_MAINTENANCE",
                                                    "REACTOME_CLASS_I_MHC_MEDIATED_ANTIGEN_PROCESSING_PRESENTATION",	"REACTOME_CELL_CYCLE_CHECKPOINTS",	"REACTOME_SARS_COV_INFECTIONS",	"REACTOME_TRANSPORT_OF_SMALL_MOLECULES",	"REACTOME_CELL_CYCLE_MITOTIC",	"REACTOME_RHO_GTPASE_CYCLE",	"REACTOME_SIGNALING_BY_RHO_GTPASES_MIRO_GTPASES_AND_RHOBTB3",	"REACTOME_NEUTROPHIL_DEGRANULATION"),


          "RespondervsNonreponder_neg_reactome" = c("REACTOME_INTERLEUKIN_18_SIGNALING"))


#data_1 = B1T4vs.all_pos_reactome 
#data_2 = B1T4vs.all_neg_reactome
#data_3 = M2T3_pos_reactome 
#data_4 = M2T3_neg_reactome


ggvenn(a, c("B1T4vs.all_pos_reactome", "RespondervsNonreponder_pos_reactome"))            # draw two-set venn
ggvenn(a, c("B1T4vs.all_neg_reactome", "RespondervsNonreponder_neg_reactome"))
#ggvenn(a, c("Set 1", "Set 2", "Set 3"))   # draw three-set venn
ggvenn(a)   # without set names, the first 4 elements in list will be chose to draw four-set venn








B1T4pos_responderpos_intersect= as.data.frame(Reduce(intersect, list(set1,set3)))
B1T4pos_responderpos_intersect =t(B1T4pos_responderpos_intersect)
B1T4neg_responderneg_intersect= as.data.frame(Reduce(intersect, list(set2,set4)))
B1T4neg_responderneg_intersect =t(B1T4neg_responderneg_intersect)


n<-max(length(B1T4pos_responderpos_intersect),length(B1T4neg_responderneg_intersect))
length(B1T4pos_responderpos_intersect)<-n
length(B1T4neg_responderneg_intersect)<-n

B1T4_responder_intersect = as.data.frame(cbind(B1T4pos_responderpos_intersect,B1T4neg_responderneg_intersect))
B1T4_responder_intersect$B1T4pos_responderpos_intersect <- B1T4_responder_intersect$V1

write.csv(B1T4_responder_intersect, "_Tcell_B1T4_responder_REACTOME_intersect_pos_neg_neg_pos.csv")
